Filter lines in the matrix of intensities w.r.t. some criteria

Returns the SummarizedExperiment object with a extra column in rowData(). The extra column indicates feature(s) to delete in 1 and 0 otherwise. The user chooses the threshold, either the percentage of NAS per row (default) or the number of samples containing NAs per row allowed. Then, the filter tags the lines that do not respect this condition. The condition may be on the whole line or condition by condition.

The different methods are : - "WholeMatrix": given a threshold th, only the lines that contain at least th values are kept. - "AllCond": given a threshold th, only the lines which contain at least th values for each of the conditions are kept. - "AtLeastOneCond": given a threshold th, only the lines that contain at least th values, and for at least one condition, are kept.

MVrowsTagToOne(object, type, th = 0, percent = TRUE)

Arguments

object
An object of class QFeatures

type
Method used to choose the lines to delete. Values are : "None", "EmptyLines", "WholeMatrix", "AllCond", "AtLeastOneCond"

th
Either a numeric between 0 and 1 where only the lines which contain at least th Or a integer between 0 and maximum number of samples for 'WholeMatrix' and between 0 and maximum number of replicate for "AllCond" and "AtLeastOneCond", where only the lines which contain at least th values are kept.

percent
TRUE by default. When FALSE, use the number of samples

Value

The object of class SummarizedExperiment with extra column in rowData indicating 1 for the lines to remove, else 0.

Author

Enora Fremy, Samuel Wieczorek

Examples

library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept
res <- MVrowsTagToOne(object, type = "AtLeastOneCond", th=0.7, percent=TRUE)