Applies aov() on a vector of protein abundances using the design derived from the sample names (simple aov wrapper)

Applies aov() on a vector of protein abundances using the design derived from the sample names (simple aov wrapper)

Usage

OWAnova(current_protein, conditions)

Arguments

current_protein

a real vector

conditions

the list of groups the protein belongs to

Value

see aov()

Author

Thomas Burger

Examples

protein_abundance <- rep(rnorm(3, mean= 18, sd=2), each=3) + rnorm(9)
groups <- c(rep("group1",3),rep("group2",3),rep("group3",3))
OWAnova(protein_abundance,groups)
#> Call:
#>    aov(formula = intensities ~ conditions, data = NULL)
#> 
#> Terms:
#>                 conditions Residuals
#> Sum of Squares    4.800036  2.772799
#> Deg. of Freedom          2         6
#> 
#> Residual standard error: 0.6798037
#> Estimated effects may be unbalanced