Method to aggregate peptides into proteins with the iterative approach.

Method to aggregate peptides into proteins with the iterative approach

aggIter(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)

Arguments

qPepData
A matrix of intensities of peptides.

X
An adjacency matrix

conditions
xxx

init.method
The method used to initialize the iterative algotirhm. Default is 'Sum'.

method
The method used for xxx. Default value is 'Mean'.

n
xxxx

Value

A matrix of protein intensities.

Author

Samuel Wieczorek

Examples

library(QFeatures)

#> Le chargement a nécessité le package : MultiAssayExperiment

#> Le chargement a nécessité le package : SummarizedExperiment

#> Le chargement a nécessité le package : MatrixGenerics

#> Warning: le package 'MatrixGenerics' a été compilé avec la version R 4.1.1

#> Le chargement a nécessité le package : matrixStats

#> Warning: le package 'matrixStats' a été compilé avec la version R 4.1.1

#> 
#> Attachement du package : 'MatrixGenerics'

#> Les objets suivants sont masqués depuis 'package:matrixStats':
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars

#> Le chargement a nécessité le package : GenomicRanges

#> Le chargement a nécessité le package : stats4

#> Le chargement a nécessité le package : BiocGenerics

#> Le chargement a nécessité le package : parallel

#> 
#> Attachement du package : 'BiocGenerics'

#> Les objets suivants sont masqués depuis 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> Les objets suivants sont masqués depuis 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs

#> Les objets suivants sont masqués depuis 'package:base':
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Le chargement a nécessité le package : S4Vectors

#> Warning: le package 'S4Vectors' a été compilé avec la version R 4.1.1

#> 
#> Attachement du package : 'S4Vectors'

#> Les objets suivants sont masqués depuis 'package:base':
#> 
#>     expand.grid, I, unname

#> Le chargement a nécessité le package : IRanges

#> 
#> Attachement du package : 'IRanges'

#> L'objet suivant est masqué depuis 'package:grDevices':
#> 
#>     windows

#> Le chargement a nécessité le package : GenomeInfoDb

#> Warning: le package 'GenomeInfoDb' a été compilé avec la version R 4.1.1

#> Le chargement a nécessité le package : Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> 
#> Attachement du package : 'Biobase'

#> L'objet suivant est masqué depuis 'package:MatrixGenerics':
#> 
#>     rowMedians

#> Les objets suivants sont masqués depuis 'package:matrixStats':
#> 
#>     anyMissing, rowMedians

#> 
#> Attachement du package : 'QFeatures'

#> L'objet suivant est masqué depuis 'package:MultiAssayExperiment':
#> 
#>     longFormat

#> L'objet suivant est masqué depuis 'package:base':
#> 
#>     sweep

utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[seq_len(1000),]
obj <- addListAdjacencyMatrices(obj, 2)

#> Error in addListAdjacencyMatrices(obj, 2): impossible de trouver la fonction "addListAdjacencyMatrices"

X <- as.matrix(GetAdjMat(obj[[2]])$all)
conditions <- colData(obj)$Condition
aggIter(assay(obj,2), X, conditions)

#> Error in t(X) %*% qPepData: arguments inadéquats