Aggregate an assay's quantitative features which take into account the peptides shared between proteins

This function aggregates the quantitative features of an assay, applying an aggregation function (`fun`) to sets of features as defined by the `fcol` feature variable. The new assay's features will be named based on the unique `fcol` values. This function is largely inspired by xxxx . The difference is that it can take into account the peptides shared between proteins.

aggregateFeatures_sam(
  object,
  i,
  aggType = "all",
  name,
  meta.names = NULL,
  fun,
  ...
)

Arguments

object
An instance of class [QFeatures].

i
The index or name of the assay which features will be aggregated the create the new assay.

aggType
The type of peptides used for the aggregation. Possibla values are: 'all', 'onlyShared' and 'onlySPec'. This argument automatically selects the corresponding adjacency matrix.

name
A `character(1)` naming the new assay. Default is `newAssay`. Note that the function will fail if there's already an assay with `name`.

meta.names
A vector of character strings that are the metadata of the peptides which needs to be aggregated and kept in the protein dataset

fun
A function used for quantitative feature aggregation. See Details for examples.

...
Additional parameters passed the `fun`.

Value

A `QFeatures` object with an additional assay.

Details

Aggregation is performed by a function that takes a matrix as input and returns a xxxxx. Examples thereof are

- [DaparToolshed:aggSum()] to use the sum of each column (default);

- [DaparToolshed:aggMean()] to use the sum of each column;

- [DaparToolshed:aggIter()] to use the mean of each column;

- [DaparToolshed:aggIterParallel()] same as previous function but use parallelism.

- [DaparToolshed::aggTopn] to use the sum of each column;

See also

Examples

library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept
aggregateFeatures_sam(obj,2, aggType= 'all', name='aggregated', 
meta.names = 'Sequence', fun='aggTopn', n=3)

#> Error in addListAdjacencyMatrices(object, i): impossible de trouver la fonction "addListAdjacencyMatrices"