Builds a plot from a dataframe. Same as compareNormalizationD but uses the library highcharter

Plot to compare the quantitative proteomics data before and after normalization using the package highcharter

Usage

compareNormalizationD_HC(
  qDataBefore,
  qDataAfter,
  keyId = NULL,
  conds = NULL,
  pal = NULL,
  subset.view = NULL,
  n = 100,
  type = "scatter"
)

Arguments

qDataBefore

A dataframe that contains quantitative data before normalization.

qDataAfter

A dataframe that contains quantitative data after normalization.

keyId

xxx

conds

A vector of the conditions (one condition per sample).

pal

xxx

subset.view

xxx

n

An integer that is equal to the maximum number of displayed points. This number must be less or equal to the size of the dataset. If it is less than it, it is a random selection

type

scatter or line

Value

A plot

Author

Samuel Wieczorek

Examples

library(SummarizedExperiment)
data(subR25prot)
qDataBefore <- assay(subR25prot[[2]])
conds <- design.qf(subR25prot)$Condition
id <- rowData(subR25prot[[2]])[, idcol(subR25prot[[2]])]
# pal <- ExtendPalette(2)
qDataAfter <- LOESS(qDataBefore, conds, type = "overall")

n <- 1
compareNormalizationD_HC(
qDataBefore = qDataBefore,
qDataAfter = qDataAfter, 
keyId = id, 
pal = NULL, 
n = n,
subset.view = seq_len(n),
conds = conds)
#> Warning: Color palette set to default.