Bar plot of missing values per lines using highcharter.

This method plots a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins).

listPlotModules()

mod_all_plots_ui(id)

mod_all_plots_server(id, object)

mod_ds_corrmatrix_ui(id)

mod_ds_corrmatrix_server(
  id,
  object,
  rate = reactive({     0.5 }),
  showValues = reactive({     FALSE })
)

mod_ds_density_ui(id)

mod_ds_density_server(id, object, conds, pal.name = reactive({
    NULL
}))

mod_ds_explorer_ui(id)

mod_ds_explorer_server(id, object, which.assay, digits = reactive({
    3
}))

mod_ds_heatmap_ui(id)

mod_ds_heatmap_server(id, object, conds, width = 900)

mod_ds_intensity_ui(id)

mod_ds_intensity_server(id, object, exp.design, ...)

mod_ds_mv_ui(id)

mod_ds_mv_server(id, object, conds, pal.name = reactive({
    NULL
}))

mod_plot_mvImputation_ui(id)

mod_plot_mvImputation_server(
  id,
  object = reactive({     NULL }),
  conds = reactive({     NULL }),
  title = reactive({     NULL }),
  pal.name = reactive({     NULL })
)

mod_ds_pca_ui(id)

mod_ds_pca_server(id, object, conds)

mod_ds_variance_ui(id)

mod_ds_variance_server(id, object, conds, pal.name = NULL)

boxPlotD(object, exp.design, subset = NULL, pal.name = NULL)

corrMatrixPlot(object, rate = 0.5, showValues = FALSE)

CVDist(object, conds = NULL, pal.name = NULL)

densityPlot(qData, conds, pal.name = NULL)

heatmapD(qData, conds, distfun = "euclidean", hclustfun = "complete")

mv.heatmap(
  x,
  col = grDevices::heat.colors(100),
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)

wrapper_pca(object, conds, var.scaling = TRUE, ncp = NULL)

violinPlotD(object, exp.design, subset = NULL, pal.name)

mvPerLinesHisto(qData)

mvPerLinesHistoPerCondition(qData, conds, pal.name)

mvHisto(qData, conds, showValues = FALSE, pal.name = NULL)

Arguments

id

shiny id

object

An instance of the class SummarizedExperiment

rate

The rate parameter to control the exponential law for the gradient of colors

showValues

xxx

conds

xxx

pal.name

xxx

which.assay

A

digits

xxx

width

xxx

exp.design

xxx

...

xxx

title

A character(1) which is

subset

A vector of index indicating rows to highlight

qData

A matrix or array quantitative values

distfun

A character(1) defining the distance used by the clustering algorithm to compute. Default value is 'euclidean'. See help(heatmap.2)

hclustfun

the clustering algorithm used to build the dendrogram. Default value is 'complete'. See help(heatmap.2)

x

A matrix or array containing the quantitative data.

col

Colors used for the image. Defaults to heat colors (heat.colors).

srtCol

Angle of column conds, in degrees from horizontal

labCol

Character vectors with column conds to use.

labRow

Character vectors with row conds to use.

key

Logical indicating whether a color-key should be shown.

key.title

Main title of the color key. If set to NA no title will be plotted.

main

Main title; default to none.

ylab

y-axis title; default to none.

var.scaling

The dimensions to plot

ncp

A integer(1) which represents the umber of dimensions kept in the results.

obj

xxx

base_palette

xxx

legend

A vector of the conditions (one condition per sample).

Value

A plot NA NA NA A boxplot A colored correlation matrix A violinplot

Details

distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins) and per conditions,

Histogram of missing values.

Variance : Builds a densityplot of the CV of entities in numeric matrix. The CV is calculated for each condition present in the dataset (see the slot 'Condition' in the colData() DataFrame)

Heatmap:

The function heatmapD()

The function [] is inspired from the function 'heatmap.2' that displays a numeric matrix. For more information, please refer to the help of the heatmap.2 function.

Missing values

#' - distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins) and per conditions,

Histogram of missing values.

Author

Samuel Wieczorek, Enora Fremy

Samuel Wieczorek, Anais Courtier, Enora Fremy

Samuel Wieczorek, Enora Fremy

Examples