This function is inspired from the function 'heatmap.2' that displays a numeric matrix. For more information, please refer to the help of the heatmap.2 function.

This function is inspired from the function 'heatmap.2' that displays a numeric matrix. For more information, please refer to the help of the heatmap.2 function.

heatmapForMissingValues(
  x,
  col = grDevices::heat.colors(100),
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)

mod_all_plots_server(id, dataIn)

Arguments

x

A matrix that contains quantitative data.

col

Colors used for the image. Defaults to heat colors (heat.colors).

srtCol

Angle of column conds, in degrees from horizontal

labCol

Character vectors with column conds to use.

labRow

Character vectors with row conds to use.

key

Logical indicating whether a color-key should be shown.

key.title

Main title of the color key. If set to NA no title will be plotted.

main

Main title; default to none.

ylab

y-axis title; default to none.

id

xxxx

dataIn

xxxx

Value

A heatmap

Author

Samuel Wieczorek, Enora Fremy

Examples

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
#> Warning: cannot open compressed file '', probable reason 'Invalid argument'
#> Error in gzfile(file, "rb"): cannot open the connection
qData <- assay(Exp1_R25_pept[[2]])[1:1000,]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'assay': object 'Exp1_R25_pept' not found
heatmapForMissingValues(qData)
#> Error in heatmapForMissingValues(qData): object 'qData' not found