Plots a histogram of p-values

This function is a wrappper to the function adjust.p from the cp4p package. It returns the FDR corresponding to the p-values of the differential analysis. The FDR is computed with the function p.adjust{stats}..

This method returns a list of the statistical tests performed with xxx and formatted as a list of two DataFrame: one for the logFC data and one for the p-values data.

This method returns a class QFeatures object with the results of differential analysis

Plots an interactive volcanoplot after the differential analysis. Typically, the log of Fold Change is represented on the X-axis and the log10 of the p-value is drawn on the Y-axis. When the threshold_pVal and the threshold_logFC are set, two lines are drawn respectively on the y-axis and the X-axis to visually distinguish between differential and non differential data. With the use of the package Highcharter, a customizable tooltip appears when the user put the mouse's pointer over a point of the scatter plot.

# S4 method for SummarizedExperiment
diff_analysis_sam(object, ...)

# S4 method for QFeatures
diff_analysis_sam(object, i, name = "diffAnaAssay", ...)

histPValue_HC(pval_ll, bins = 80, pi0 = 1)

diffAnaComputeFDR(
  logFC,
  pval,
  threshold_PVal = 0,
  threshold_LogFC = 0,
  pi0Method = 1
)

Get_AllComparisons(obj)

diffAnalysis(obj, comp, th_pval = 0, th_logFC = 0)

diffAnaVolcanoplot_rCharts(
  df,
  threshold_pVal = 1e-60,
  threshold_logFC = 0,
  conditions = NULL,
  clickFunction = NULL,
  palette = NULL,
  swap = FALSE
)

Arguments

object

An object of class QFeatures or SummarizedExperiment.

...

Additional parameters passed to inner functions.

i

A numeric vector or a character vector giving the index or the name, respectively, of the assay(s) to be processed.

name

A character(1) naming the new assay name. Default is ttestAssay.

pval_ll

A data frame which contains the p-values for the set of comparisons to show.

bins

xxx

pi0

xxx

logFC

The result (logFC values) of the differential analysis processed by 'limmaCompleteTest'

pval

The result (p-values) of the differential analysis processed by 'limmaCompleteTest'

threshold_PVal

The threshold on p-pvalue to distinguish between differential and non-differential data

threshold_LogFC

The threshold on log(Fold Change) to distinguish between differential and non-differential data

pi0Method

The parameter pi0.method of the method adjust.p in the package cp4p

obj

An object of class SUmmarizeExperiment.

comp

The name of the comparison (given by the t-tests) on whihc to apply thresholds

th_pval

xxx

th_logFC

xxx

df

A dataframe which contains the following slots : x : a vector of the log(fold change) values of the differential analysis, y : a vector of the p-value values returned by the differential analysis. index : a vector of the rownames of the data. This dataframe must has been built with the option stringsAsFactors set to FALSE. There may be additional slots which will be used to show informations in the tooltip. The name of these slots must begin with the prefix "tooltip_". It will be automatically removed in the plot.

threshold_pVal

A floating number which represents the p-value that separates differential and non-differential data.

threshold_logFC

A floating number which represents the log of the Fold Change that separates differential and non-differential data.

conditions

A list of the names of condition 1 and 2 used for the differential analysis.

clickFunction

A string that contains a JavaScript function used to show info from slots in df. The variable this.index refers to the slot named index and allows to retrieve the right row to show in the tooltip.

palette

xxx

swap

A boolean that indicates if the conditions have been swaped

Value

A plot The computed FDR value (floating number) A list of two slots: logFC and P_Value. Each slot contains a DataFrame An object of class SummarizedExperimentAn interactive volcanoplot

Author

Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek, Enora Fremy

Examples

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
#> Warning: cannot open compressed file '', probable reason 'Invalid argument'
#> Error in gzfile(file, "rb"): cannot open the connection
object <- Exp1_R25_pept[seq_len(1000),]
#> Error in eval(expr, envir, enclos): object 'Exp1_R25_pept' not found
object <- addAssay(object, QFeatures::filterNA(object[[2]],  pNA = 0), name='filtered')
#> Error in stopifnot(inherits(x, "QFeatures")): object 'object' not found
object <- addListAdjacencyMatrices(object, 3)
#> Error in addListAdjacencyMatrices(object, 3): could not find function "addListAdjacencyMatrices"
sTab <- colData(object)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'colData': object 'object' not found
se <- t_test_sam(object[[3]], sTab, FUN = compute.t.test)
#> Error in t_test_sam(object[[3]], sTab, FUN = compute.t.test): could not find function "t_test_sam"
pval <- metadata(se)$t_test[['25fmol_vs_10fmol_pval']]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'metadata': object 'se' not found
histPValue_HC(pval)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'sort': object 'pval' not found

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
#> Warning: cannot open compressed file '', probable reason 'Invalid argument'
#> Error in gzfile(file, "rb"): cannot open the connection
obj <- Exp1_R25_pept[seq_len(1000)]
#> Error in eval(expr, envir, enclos): object 'Exp1_R25_pept' not found
obj <- addAssay(obj, QFeatures::filterNA(obj[[2]],  pNA = 0), name='filtered')
#> Error in stopifnot(inherits(x, "QFeatures")): object 'obj' not found
se <- t_test_sam(obj[[3]], colData(obj), FUN = compute.t.test)
#> Error in t_test_sam(obj[[3]], colData(obj), FUN = compute.t.test): could not find function "t_test_sam"
ind_logFC <- grep('_logFC', colnames(metadata(se)$t_test))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'grep': error in evaluating the argument 'x' in selecting a method for function 'colnames': error in evaluating the argument 'x' in selecting a method for function 'metadata': object 'se' not found
logFC <- setNames(as.data.frame(metadata(se)$t_test[,ind_logFC]), 
colnames(metadata(se)$t_test)[ind_logFC])
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'colnames': error in evaluating the argument 'x' in selecting a method for function 'metadata': object 'se' not found
ind_pval <- grep('_pval', colnames(metadata(se)$t_test))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'grep': error in evaluating the argument 'x' in selecting a method for function 'colnames': error in evaluating the argument 'x' in selecting a method for function 'metadata': object 'se' not found
pval <- setNames(as.data.frame(metadata(se)$t_test[,ind_pval]), 
colnames(metadata(se)$t_test)[ind_pval])
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'colnames': error in evaluating the argument 'x' in selecting a method for function 'metadata': object 'se' not found
diffAnaComputeFDR(logFC[,1], pval[,1])
#> Error in which(abs(logFC) >= threshold_LogFC): object 'logFC' not found

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
#> Warning: cannot open compressed file '', probable reason 'Invalid argument'
#> Error in gzfile(file, "rb"): cannot open the connection
object <- Exp1_R25_pept[seq_len(1000),]
#> Error in eval(expr, envir, enclos): object 'Exp1_R25_pept' not found
object <- addAssay(object, QFeatures::filterNA(object[[2]],  pNA = 0), name='filtered')
#> Error in stopifnot(inherits(x, "QFeatures")): object 'object' not found
sTab <- colData(object)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'colData': object 'object' not found
gttest.se <- t_test_sam(object[[3]], sTab, FUN = compute.t.test)
#> Error in t_test_sam(object[[3]], sTab, FUN = compute.t.test): could not find function "t_test_sam"
object <- addAssay(object, gttest.se, name='t-test')
#> Error in stopifnot(inherits(x, "QFeatures")): object 'object' not found
comp <- '25fmol_vs_10fmol'
object <- diff_analysis_sam(object, 't-test', name='diffAna', comp)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'diff_analysis_sam': object 'object' not found

allComp <- Get_AllComparisons(object[['diffAna']])
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'metadata': object 'object' not found

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
#> Warning: cannot open compressed file '', probable reason 'Invalid argument'
#> Error in gzfile(file, "rb"): cannot open the connection
obj <- Exp1_R25_pept[seq_len(1000)]
#> Error in eval(expr, envir, enclos): object 'Exp1_R25_pept' not found
obj <- addAssay(obj, QFeatures::filterNA(obj[[2]],  pNA = 0), name='filtered')
#> Error in stopifnot(inherits(x, "QFeatures")): object 'obj' not found
obj <- addAssay(obj, t_test_sam(obj[[3]], colData(obj), FUN = 'compute.t.test'), name='t-test')
#> Error in stopifnot(inherits(x, "QFeatures")): object 'obj' not found
comp <- '25fmol_vs_10fmol'
da.se <- diffAnalysis(obj[['t-test']], comp, th_pval=0, th_logFC=0)
#> Error in diffAnalysis(obj[["t-test"]], comp, th_pval = 0, th_logFC = 0): object 'obj' not found

# \donttest{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DaparToolshedData')
#> Error in find.package(package, lib.loc, verbose = verbose): there is no package called 'DaparToolshedData'
obj <- QFeatures::filterNA(Exp1_R25_pept,i='original') # remove all rows with NAs
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'filterNA': object 'Exp1_R25_pept' not found
qData <- assay(obj[['original']])
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'assay': object 'obj' not found
pData <- as.data.frame(colData(obj))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'as.data.frame': error in evaluating the argument 'x' in selecting a method for function 'colData': object 'obj' not found
data <- limma.complete.test(qData,pData)
#> Error in limma.complete.test(qData, pData): could not find function "limma.complete.test"
df <- data.frame(x=data[,1], y = -log10(data[,2]),index = as.character(rownames(qData)))
#> Error in data[, 1]: object of type 'closure' is not subsettable
colnames(df) <- c("x", "y", "index")
#> Error in `colnames<-`(`*tmp*`, value = c("x", "y", "index")): attempt to set 'colnames' on an object with less than two dimensions
tooltipSlot <- c("Sequence", "Score")
df <- cbind(df,rowData(obj[['original']])[tooltipSlot])
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'rowData': object 'obj' not found
colnames(df) <- gsub(".", "_", colnames(df), fixed=TRUE)
#> Error in `colnames<-`(`*tmp*`, value = character(0)): attempt to set 'colnames' on an object with less than two dimensions
if (ncol(df) > 3){
    colnames(df)[4:ncol(df)] <- 
    paste("tooltip_", colnames(df)[4:ncol(df)], sep="")}
#> Error in if (ncol(df) > 3) {    colnames(df)[4:ncol(df)] <- paste("tooltip_", colnames(df)[4:ncol(df)],         sep = "")}: argument is of length zero
hc_clickFunction <- JS("function(event) {
Shiny.onInputChange('eventPointClicked', [this.index]+'_'+ [this.series.name]);}")
cond <- unique(colData(obj)[['Condition']])
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'unique': error in evaluating the argument 'x' in selecting a method for function 'colData': object 'obj' not found
diffAnaVolcanoplot_rCharts(df, 2.5, 1, cond, hc_clickFunction) 
#> Error in paste("log2 ( mean(", conditions[2], ") / mean(", conditions[1],     ") )", sep = ""): object 'cond' not found
# }