Missing values imputation from a MSnSet object

This method is a wrapper to objects of class MSnSet and imputes missing values with a fixed value.

impute_fixed_value(x, value)

Arguments

x
An assay of a SummarizedExperiment

value
A float .

Value

The object obj which has been imputed

Author

Samuel Wieczorek

Examples

utils::data(Exp1_R25_pept, package='DAPARdata')
impute_fixed_value(assay(Exp1_R25_pept[seq_len(1000)], 2), 0.001)

#> Le chargement a nécessité le package : MSnbase

#> Le chargement a nécessité le package : mzR

#> Warning: mzR has been built against a different Rcpp version (1.0.6)
#> than is installed on your system (1.0.7). This might lead to errors
#> when loading mzR. If you encounter such issues, please send a report,
#> including the output of sessionInfo() to the Bioc support forum at 
#> https://support.bioconductor.org/. For details see also
#> https://github.com/sneumann/mzR/wiki/mzR-Rcpp-compiler-linker-issue.

#> Error: le chargement du package ou de l'espace de noms a échoué pour 'mzR' in .doLoadActions(where, attach) :
#> error in load action .__A__.1 for package mzR: loadModule(module = "Ramp", what = TRUE, env = ns, loadNow = TRUE): Unable to load module "Ramp": tentative d'appliquer un objet qui n'est pas une fonction

#> Error in h(simpleError(msg, call)): erreur d'ï¿½valuation de l'argument 'x' lors de la sï¿½lection d'une mï¿½thode pour la fonction 'assay' : unable to find required package 'MSnbase'