Intensity family plots

These functions are plotting different views of quantitative data. Two plots are available:

boxplotD() xxxx,

violinPlotD() xxxx.

boxPlotD(object, exp.design, subset = NULL, pal.name = NULL)

violinPlotD(object, exp.design, subset = NULL, pal.name)

mod_intensity_plots_ui(id)

mod_intensity_plots_server(id, object, exp.design, ...)

Arguments

object

xxx

exp.design

xxx

subset

A vector of index indicating rows to highlight

pal.name

xxx

id

xxx

...

xxx

Value

A boxplot A violinplot NA NA

Author

Samuel Wieczorek, Anais Courtier, Enora Fremy

Samuel Wieczorek, Anais Courtier, Enora Fremy

Examples

library(QFeatures)
data(ft)

#----------------------------------------------
# View violin plot with and without subset view
#----------------------------------------------

violinPlotD(ft[[1]], design(ft))
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("QFeatures", package = "QFeatures")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("design"), generic = structure("design", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "design"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function 'design' for signature '"QFeatures"'
violinPlotD(ft[[1]], design(ft), subset = c(3,4,6), pal.name = 'Dark2')
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("QFeatures", package = "QFeatures")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("design"), generic = structure("design", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "design"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function 'design' for signature '"QFeatures"'

#----------------------------------------
# xxxx
#----------------------------------------

boxPlotD(ft[[1]], design(ft))
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("QFeatures", package = "QFeatures")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("design"), generic = structure("design", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "design"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function 'design' for signature '"QFeatures"'
boxPlotD(ft[[1]], design(ft), subset = c(2, 3), pal.name = 'Dark2')
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("QFeatures", package = "QFeatures")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("design"), generic = structure("design", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "design"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function 'design' for signature '"QFeatures"'

# ----------------------------------------------------
# A shiny module which can track entities on the plots
# ----------------------------------------------------

if (interactive()){
library(QFeatures)
library(shiny)
library(shinyjs)
library(highcharter)
library(DaparToolshed)

data(ft)

ui <- tagList(
  mod_intensity_plots_ui('iplots'),
  uiOutput('show')
)

server <- function(input, output, session) {
  rv <- reactiveValues(
    tmp = NULL
  )
  
  mod_intensity_plots_server(id = 'iplots',
                             object = reactive({ft[[1]]}),
                             exp.design = reactive({design(ft)}),
                             pal.name = 'Dark2'
                             )
  
}
shinyApp(ui = ui, server = server)
}