Similar to the function is.na() but focused on the equality with the paramter 'type'.

Similar to the function is.na() but focused on the equality with the paramter 'type'.

Usage

is.OfType(data, type)

Arguments

data

A data.frame

type

The value to search in the dataframe

Value

A boolean dataframe

Author

Samuel Wieczorek

Examples

library(QFeatures)
data(subR25prot)
obj <- subR25prot[[1]]
data <- qMetacell(obj)
is.OfType(as.data.frame(data), "MEC")
#>                                                    metacell_Intensity_C_R1
#> CON__A2I7N1;CON__A2I7N0                                              FALSE
#> CON__P00761                                                          FALSE
#> P02768upsedyp|ALBU_HUMAN_upsedyp;CON__P02768-1                       FALSE
#> CON__P04264                                                          FALSE
#> CON__P07477                                                          FALSE
#> CON__P13645                                                          FALSE
#> CON__P35527                                                          FALSE
#> CON__P35908                                                          FALSE
#> CON__P62894                                                          FALSE
#> CON__Q2KIS7                                                          FALSE
#> O00762upsedyp|UBE2C_HUMAN_upsedyp                                    FALSE
#> O76070upsedyp|SYUG_HUMAN_upsedyp                                     FALSE
#> P00167upsedyp|CYB5_HUMAN_upsedyp                                     FALSE
#> P00441upsedyp|SODC_HUMAN_upsedyp                                     FALSE
#> P00709upsedyp|LALBA_HUMAN_upsedyp                                    FALSE
#> P00915upsedyp|CAH1_HUMAN_upsedyp                                     FALSE
#> P00918upsedyp|CAH2_HUMAN_upsedyp                                     FALSE
#> P01008upsedyp|ANT3_HUMAN_upsedyp                                     FALSE
#> P01031upsedyp|CO5_HUMAN_upsedyp                                      FALSE
#> P01112upsedyp|RASH_HUMAN_upsedyp                                     FALSE
#> P01127upsedyp|PDGFB_HUMAN_upsedyp                                    FALSE
#> P01133upsedyp|EGF_HUMAN_upsedyp                                      FALSE
#> P01344upsedyp|IGF2_HUMAN_upsedyp                                     FALSE
#> P01375upsedyp|TNFA_HUMAN_upsedyp                                     FALSE
#> P01579upsedyp|IFNG_HUMAN_upsedyp                                     FALSE
#> P02144upsedyp|MYG_HUMAN_upsedyp                                      FALSE
#> P02741upsedyp|CRP_HUMAN_upsedyp                                      FALSE
#> P02753upsedyp|RET4_HUMAN_upsedyp                                     FALSE
#> P02787upsedyp|TRFE_HUMAN_upsedyp                                     FALSE
#> P02788upsedyp|TRFL_HUMAN_upsedyp                                     FALSE
#> P04040upsedyp|CATA_HUMAN_upsedyp                                     FALSE
#> P05413upsedyp|FABPH_HUMAN_upsedyp                                    FALSE
#> P06396upsedyp|GELS_HUMAN_upsedyp                                     FALSE
#> P06732upsedyp|KCRM_HUMAN_upsedyp                                     FALSE
#> P08263upsedyp|GSTA1_HUMAN_upsedyp                                    FALSE
#> P08758upsedyp|ANXA5_HUMAN_upsedyp                                    FALSE
#> P09211upsedyp|GSTP1_HUMAN_upsedyp                                    FALSE
#> P10145upsedyp|IL8_HUMAN_upsedyp                                      FALSE
#> P10599upsedyp|THIO_HUMAN_upsedyp                                     FALSE
#> P10636-8upsedyp|TAU_HUMAN_upsedyp                                    FALSE
#> P12081upsedyp|SYHC_HUMAN_upsedyp                                     FALSE
#> P15559upsedyp|NQO1_HUMAN_upsedyp                                     FALSE
#> P16083upsedyp|NQO2_HUMAN_upsedyp                                     FALSE
#> P41159upsedyp|LEP_HUMAN_upsedyp                                      FALSE
#> P51965upsedyp|UB2E1_HUMAN_upsedyp                                    FALSE
#> P55957upsedyp|BID_HUMAN_upsedyp                                      FALSE
#> P61626upsedyp|LYSC_HUMAN_upsedyp                                     FALSE
#> P61769upsedyp|B2MG_HUMAN_upsedyp                                     FALSE
#> P62937upsedyp|PPIA_HUMAN_upsedyp                                     FALSE
#> P62979upsedyp|RS27A_HUMAN_upsedyp                                    FALSE
#> P63165upsedyp|SUMO1_HUMAN_upsedyp                                    FALSE
#> P63279upsedyp|UBC9_HUMAN_upsedyp                                     FALSE
#> P68871upsedyp|HBB_HUMAN_upsedyp                                      FALSE
#> P69905upsedyp|HBA_HUMAN_upsedyp                                      FALSE
#> P99999upsedyp|CYC_HUMAN_upsedyp                                      FALSE
#> Q06830upsedyp|PRDX1_HUMAN_upsedyp                                    FALSE
#> Q15843upsedyp|NEDD8_HUMAN_upsedyp                                    FALSE
#> REV__CON__Q3KNV1;REV__CON__P08729;REV__CON__P07744                   FALSE
#> REV__CON__Q0VCM5                                                     FALSE
#> REV__CON__Q14CN4-1                                                   FALSE
#> REV__CON__Q2YDI2                                                     FALSE
#> REV__CON__Q7Z794                                                     FALSE
#> REV__CON__Q9DCV7                                                     FALSE
#> REV__sp|P0C155|CBPS2_YEAST                                           FALSE
#> REV__sp|P0CF34|YL016_YEAST                                           FALSE
#> REV__sp|P32380|SP110_YEAST                                           FALSE
#> REV__sp|P33314|BUD2_YEAST                                            FALSE
#> REV__sp|P36150|SUMT_YEAST                                            FALSE
#> REV__sp|P36775|LONM_YEAST                                            FALSE
#> REV__sp|P38279|YPQ3_YEAST                                            FALSE
#> REV__sp|P39526|LAA1_YEAST                                            FALSE
#> REV__sp|P39531|RCY1_YEAST                                            FALSE
#> REV__sp|P40012|PPOX_YEAST                                            FALSE
#> REV__sp|P40358|JEM1_YEAST                                            FALSE
#> REV__sp|P46678|TFC5_YEAST                                            FALSE
#> REV__sp|P53119|HUL5_YEAST                                            FALSE
#> REV__sp|P53746|FRE4_YEAST                                            FALSE
#> REV__sp|Q02932|KA120_YEAST                                           FALSE
#> REV__sp|Q03660|TR130_YEAST                                           FALSE
#> REV__sp|Q06639|RSC3_YEAST                                            FALSE
#> REV__sp|Q07660|BRE4_YEAST                                            FALSE
#> sp|A5Z2X5|YP010_YEAST                                                FALSE
#> sp|D6VTK4|STE2_YEAST                                                 FALSE
#> sp|O13297|CET1_YEAST                                                 FALSE
#> sp|O13516|RS9A_YEAST                                                 FALSE
#> sp|O13563|RPN13_YEAST                                                FALSE
#> sp|O13585|YP089_YEAST                                                FALSE
#> sp|P05745|RL36A_YEAST;sp|O14455|RL36B_YEAST                          FALSE
#> sp|O14467|MBF1_YEAST                                                 FALSE
#> sp|O14468|YO304_YEAST                                                FALSE
#> sp|O43137|YB085_YEAST                                                FALSE
#> sp|O60200|MDM35_YEAST                                                FALSE
#> sp|O74700|TIM9_YEAST                                                 FALSE
#> sp|O94742|SEM1_YEAST                                                 FALSE
#> sp|P00128|QCR7_YEAST                                                 FALSE
#> sp|P00330|ADH1_YEAST                                                 FALSE
#> sp|P00358|G3P2_YEAST                                                 FALSE
#> sp|P00359|G3P3_YEAST                                                 FALSE
#> sp|P00360|G3P1_YEAST                                                 FALSE
#> sp|P00427|COX6_YEAST                                                 FALSE
#>                                                    metacell_Intensity_C_R2
#> CON__A2I7N1;CON__A2I7N0                                              FALSE
#> CON__P00761                                                          FALSE
#> P02768upsedyp|ALBU_HUMAN_upsedyp;CON__P02768-1                       FALSE
#> CON__P04264                                                          FALSE
#> CON__P07477                                                          FALSE
#> CON__P13645                                                          FALSE
#> CON__P35527                                                          FALSE
#> CON__P35908                                                          FALSE
#> CON__P62894                                                          FALSE
#> CON__Q2KIS7                                                          FALSE
#> O00762upsedyp|UBE2C_HUMAN_upsedyp                                    FALSE
#> O76070upsedyp|SYUG_HUMAN_upsedyp                                     FALSE
#> P00167upsedyp|CYB5_HUMAN_upsedyp                                     FALSE
#> P00441upsedyp|SODC_HUMAN_upsedyp                                     FALSE
#> P00709upsedyp|LALBA_HUMAN_upsedyp                                    FALSE
#> P00915upsedyp|CAH1_HUMAN_upsedyp                                     FALSE
#> P00918upsedyp|CAH2_HUMAN_upsedyp                                     FALSE
#> P01008upsedyp|ANT3_HUMAN_upsedyp                                     FALSE
#> P01031upsedyp|CO5_HUMAN_upsedyp                                      FALSE
#> P01112upsedyp|RASH_HUMAN_upsedyp                                     FALSE
#> P01127upsedyp|PDGFB_HUMAN_upsedyp                                    FALSE
#> P01133upsedyp|EGF_HUMAN_upsedyp                                      FALSE
#> P01344upsedyp|IGF2_HUMAN_upsedyp                                     FALSE
#> P01375upsedyp|TNFA_HUMAN_upsedyp                                     FALSE
#> P01579upsedyp|IFNG_HUMAN_upsedyp                                     FALSE
#> P02144upsedyp|MYG_HUMAN_upsedyp                                      FALSE
#> P02741upsedyp|CRP_HUMAN_upsedyp                                      FALSE
#> P02753upsedyp|RET4_HUMAN_upsedyp                                     FALSE
#> P02787upsedyp|TRFE_HUMAN_upsedyp                                     FALSE
#> P02788upsedyp|TRFL_HUMAN_upsedyp                                     FALSE
#> P04040upsedyp|CATA_HUMAN_upsedyp                                     FALSE
#> P05413upsedyp|FABPH_HUMAN_upsedyp                                    FALSE
#> P06396upsedyp|GELS_HUMAN_upsedyp                                     FALSE
#> P06732upsedyp|KCRM_HUMAN_upsedyp                                     FALSE
#> P08263upsedyp|GSTA1_HUMAN_upsedyp                                    FALSE
#> P08758upsedyp|ANXA5_HUMAN_upsedyp                                    FALSE
#> P09211upsedyp|GSTP1_HUMAN_upsedyp                                    FALSE
#> P10145upsedyp|IL8_HUMAN_upsedyp                                      FALSE
#> P10599upsedyp|THIO_HUMAN_upsedyp                                     FALSE
#> P10636-8upsedyp|TAU_HUMAN_upsedyp                                    FALSE
#> P12081upsedyp|SYHC_HUMAN_upsedyp                                     FALSE
#> P15559upsedyp|NQO1_HUMAN_upsedyp                                     FALSE
#> P16083upsedyp|NQO2_HUMAN_upsedyp                                     FALSE
#> P41159upsedyp|LEP_HUMAN_upsedyp                                      FALSE
#> P51965upsedyp|UB2E1_HUMAN_upsedyp                                    FALSE
#> P55957upsedyp|BID_HUMAN_upsedyp                                      FALSE
#> P61626upsedyp|LYSC_HUMAN_upsedyp                                     FALSE
#> P61769upsedyp|B2MG_HUMAN_upsedyp                                     FALSE
#> P62937upsedyp|PPIA_HUMAN_upsedyp                                     FALSE
#> P62979upsedyp|RS27A_HUMAN_upsedyp                                    FALSE
#> P63165upsedyp|SUMO1_HUMAN_upsedyp                                    FALSE
#> P63279upsedyp|UBC9_HUMAN_upsedyp                                     FALSE
#> P68871upsedyp|HBB_HUMAN_upsedyp                                      FALSE
#> P69905upsedyp|HBA_HUMAN_upsedyp                                      FALSE
#> P99999upsedyp|CYC_HUMAN_upsedyp                                      FALSE
#> Q06830upsedyp|PRDX1_HUMAN_upsedyp                                    FALSE
#> Q15843upsedyp|NEDD8_HUMAN_upsedyp                                    FALSE
#> REV__CON__Q3KNV1;REV__CON__P08729;REV__CON__P07744                   FALSE
#> REV__CON__Q0VCM5                                                     FALSE
#> REV__CON__Q14CN4-1                                                   FALSE
#> REV__CON__Q2YDI2                                                     FALSE
#> REV__CON__Q7Z794                                                     FALSE
#> REV__CON__Q9DCV7                                                     FALSE
#> REV__sp|P0C155|CBPS2_YEAST                                           FALSE
#> REV__sp|P0CF34|YL016_YEAST                                           FALSE
#> REV__sp|P32380|SP110_YEAST                                           FALSE
#> REV__sp|P33314|BUD2_YEAST                                            FALSE
#> REV__sp|P36150|SUMT_YEAST                                            FALSE
#> REV__sp|P36775|LONM_YEAST                                            FALSE
#> REV__sp|P38279|YPQ3_YEAST                                            FALSE
#> REV__sp|P39526|LAA1_YEAST                                            FALSE
#> REV__sp|P39531|RCY1_YEAST                                            FALSE
#> REV__sp|P40012|PPOX_YEAST                                            FALSE
#> REV__sp|P40358|JEM1_YEAST                                            FALSE
#> REV__sp|P46678|TFC5_YEAST                                            FALSE
#> REV__sp|P53119|HUL5_YEAST                                            FALSE
#> REV__sp|P53746|FRE4_YEAST                                            FALSE
#> REV__sp|Q02932|KA120_YEAST                                           FALSE
#> REV__sp|Q03660|TR130_YEAST                                           FALSE
#> REV__sp|Q06639|RSC3_YEAST                                            FALSE
#> REV__sp|Q07660|BRE4_YEAST                                            FALSE
#> sp|A5Z2X5|YP010_YEAST                                                FALSE
#> sp|D6VTK4|STE2_YEAST                                                 FALSE
#> sp|O13297|CET1_YEAST                                                 FALSE
#> sp|O13516|RS9A_YEAST                                                 FALSE
#> sp|O13563|RPN13_YEAST                                                FALSE
#> sp|O13585|YP089_YEAST                                                FALSE
#> sp|P05745|RL36A_YEAST;sp|O14455|RL36B_YEAST                          FALSE
#> sp|O14467|MBF1_YEAST                                                 FALSE
#> sp|O14468|YO304_YEAST                                                FALSE
#> sp|O43137|YB085_YEAST                                                FALSE
#> sp|O60200|MDM35_YEAST                                                FALSE
#> sp|O74700|TIM9_YEAST                                                 FALSE
#> sp|O94742|SEM1_YEAST                                                 FALSE
#> sp|P00128|QCR7_YEAST                                                 FALSE
#> sp|P00330|ADH1_YEAST                                                 FALSE
#> sp|P00358|G3P2_YEAST                                                 FALSE
#> sp|P00359|G3P3_YEAST                                                 FALSE
#> sp|P00360|G3P1_YEAST                                                 FALSE
#> sp|P00427|COX6_YEAST                                                 FALSE
#>                                                    metacell_Intensity_C_R3
#> CON__A2I7N1;CON__A2I7N0                                              FALSE
#> CON__P00761                                                          FALSE
#> P02768upsedyp|ALBU_HUMAN_upsedyp;CON__P02768-1                       FALSE
#> CON__P04264                                                          FALSE
#> CON__P07477                                                          FALSE
#> CON__P13645                                                          FALSE
#> CON__P35527                                                          FALSE
#> CON__P35908                                                          FALSE
#> CON__P62894                                                          FALSE
#> CON__Q2KIS7                                                          FALSE
#> O00762upsedyp|UBE2C_HUMAN_upsedyp                                    FALSE
#> O76070upsedyp|SYUG_HUMAN_upsedyp                                     FALSE
#> P00167upsedyp|CYB5_HUMAN_upsedyp                                     FALSE
#> P00441upsedyp|SODC_HUMAN_upsedyp                                     FALSE
#> P00709upsedyp|LALBA_HUMAN_upsedyp                                    FALSE
#> P00915upsedyp|CAH1_HUMAN_upsedyp                                     FALSE
#> P00918upsedyp|CAH2_HUMAN_upsedyp                                     FALSE
#> P01008upsedyp|ANT3_HUMAN_upsedyp                                     FALSE
#> P01031upsedyp|CO5_HUMAN_upsedyp                                      FALSE
#> P01112upsedyp|RASH_HUMAN_upsedyp                                     FALSE
#> P01127upsedyp|PDGFB_HUMAN_upsedyp                                    FALSE
#> P01133upsedyp|EGF_HUMAN_upsedyp                                      FALSE
#> P01344upsedyp|IGF2_HUMAN_upsedyp                                     FALSE
#> P01375upsedyp|TNFA_HUMAN_upsedyp                                     FALSE
#> P01579upsedyp|IFNG_HUMAN_upsedyp                                     FALSE
#> P02144upsedyp|MYG_HUMAN_upsedyp                                      FALSE
#> P02741upsedyp|CRP_HUMAN_upsedyp                                      FALSE
#> P02753upsedyp|RET4_HUMAN_upsedyp                                     FALSE
#> P02787upsedyp|TRFE_HUMAN_upsedyp                                     FALSE
#> P02788upsedyp|TRFL_HUMAN_upsedyp                                     FALSE
#> P04040upsedyp|CATA_HUMAN_upsedyp                                     FALSE
#> P05413upsedyp|FABPH_HUMAN_upsedyp                                    FALSE
#> P06396upsedyp|GELS_HUMAN_upsedyp                                     FALSE
#> P06732upsedyp|KCRM_HUMAN_upsedyp                                     FALSE
#> P08263upsedyp|GSTA1_HUMAN_upsedyp                                    FALSE
#> P08758upsedyp|ANXA5_HUMAN_upsedyp                                    FALSE
#> P09211upsedyp|GSTP1_HUMAN_upsedyp                                    FALSE
#> P10145upsedyp|IL8_HUMAN_upsedyp                                      FALSE
#> P10599upsedyp|THIO_HUMAN_upsedyp                                     FALSE
#> P10636-8upsedyp|TAU_HUMAN_upsedyp                                    FALSE
#> P12081upsedyp|SYHC_HUMAN_upsedyp                                     FALSE
#> P15559upsedyp|NQO1_HUMAN_upsedyp                                     FALSE
#> P16083upsedyp|NQO2_HUMAN_upsedyp                                     FALSE
#> P41159upsedyp|LEP_HUMAN_upsedyp                                      FALSE
#> P51965upsedyp|UB2E1_HUMAN_upsedyp                                    FALSE
#> P55957upsedyp|BID_HUMAN_upsedyp                                      FALSE
#> P61626upsedyp|LYSC_HUMAN_upsedyp                                     FALSE
#> P61769upsedyp|B2MG_HUMAN_upsedyp                                     FALSE
#> P62937upsedyp|PPIA_HUMAN_upsedyp                                     FALSE
#> P62979upsedyp|RS27A_HUMAN_upsedyp                                    FALSE
#> P63165upsedyp|SUMO1_HUMAN_upsedyp                                    FALSE
#> P63279upsedyp|UBC9_HUMAN_upsedyp                                     FALSE
#> P68871upsedyp|HBB_HUMAN_upsedyp                                      FALSE
#> P69905upsedyp|HBA_HUMAN_upsedyp                                      FALSE
#> P99999upsedyp|CYC_HUMAN_upsedyp                                      FALSE
#> Q06830upsedyp|PRDX1_HUMAN_upsedyp                                    FALSE
#> Q15843upsedyp|NEDD8_HUMAN_upsedyp                                    FALSE
#> REV__CON__Q3KNV1;REV__CON__P08729;REV__CON__P07744                   FALSE
#> REV__CON__Q0VCM5                                                     FALSE
#> REV__CON__Q14CN4-1                                                   FALSE
#> REV__CON__Q2YDI2                                                     FALSE
#> REV__CON__Q7Z794                                                     FALSE
#> REV__CON__Q9DCV7                                                     FALSE
#> REV__sp|P0C155|CBPS2_YEAST                                           FALSE
#> REV__sp|P0CF34|YL016_YEAST                                           FALSE
#> REV__sp|P32380|SP110_YEAST                                           FALSE
#> REV__sp|P33314|BUD2_YEAST                                            FALSE
#> REV__sp|P36150|SUMT_YEAST                                            FALSE
#> REV__sp|P36775|LONM_YEAST                                            FALSE
#> REV__sp|P38279|YPQ3_YEAST                                            FALSE
#> REV__sp|P39526|LAA1_YEAST                                            FALSE
#> REV__sp|P39531|RCY1_YEAST                                            FALSE
#> REV__sp|P40012|PPOX_YEAST                                            FALSE
#> REV__sp|P40358|JEM1_YEAST                                            FALSE
#> REV__sp|P46678|TFC5_YEAST                                            FALSE
#> REV__sp|P53119|HUL5_YEAST                                            FALSE
#> REV__sp|P53746|FRE4_YEAST                                            FALSE
#> REV__sp|Q02932|KA120_YEAST                                           FALSE
#> REV__sp|Q03660|TR130_YEAST                                           FALSE
#> REV__sp|Q06639|RSC3_YEAST                                            FALSE
#> REV__sp|Q07660|BRE4_YEAST                                            FALSE
#> sp|A5Z2X5|YP010_YEAST                                                FALSE
#> sp|D6VTK4|STE2_YEAST                                                 FALSE
#> sp|O13297|CET1_YEAST                                                 FALSE
#> sp|O13516|RS9A_YEAST                                                 FALSE
#> sp|O13563|RPN13_YEAST                                                FALSE
#> sp|O13585|YP089_YEAST                                                FALSE
#> sp|P05745|RL36A_YEAST;sp|O14455|RL36B_YEAST                          FALSE
#> sp|O14467|MBF1_YEAST                                                 FALSE
#> sp|O14468|YO304_YEAST                                                FALSE
#> sp|O43137|YB085_YEAST                                                FALSE
#> sp|O60200|MDM35_YEAST                                                FALSE
#> sp|O74700|TIM9_YEAST                                                 FALSE
#> sp|O94742|SEM1_YEAST                                                 FALSE
#> sp|P00128|QCR7_YEAST                                                 FALSE
#> sp|P00330|ADH1_YEAST                                                 FALSE
#> sp|P00358|G3P2_YEAST                                                 FALSE
#> sp|P00359|G3P3_YEAST                                                 FALSE
#> sp|P00360|G3P1_YEAST                                                 FALSE
#> sp|P00427|COX6_YEAST                                                 FALSE
#>                                                    metacell_Intensity_D_R1
#> CON__A2I7N1;CON__A2I7N0                                              FALSE
#> CON__P00761                                                          FALSE
#> P02768upsedyp|ALBU_HUMAN_upsedyp;CON__P02768-1                       FALSE
#> CON__P04264                                                          FALSE
#> CON__P07477                                                          FALSE
#> CON__P13645                                                          FALSE
#> CON__P35527                                                          FALSE
#> CON__P35908                                                          FALSE
#> CON__P62894                                                          FALSE
#> CON__Q2KIS7                                                          FALSE
#> O00762upsedyp|UBE2C_HUMAN_upsedyp                                    FALSE
#> O76070upsedyp|SYUG_HUMAN_upsedyp                                     FALSE
#> P00167upsedyp|CYB5_HUMAN_upsedyp                                     FALSE
#> P00441upsedyp|SODC_HUMAN_upsedyp                                     FALSE
#> P00709upsedyp|LALBA_HUMAN_upsedyp                                    FALSE
#> P00915upsedyp|CAH1_HUMAN_upsedyp                                     FALSE
#> P00918upsedyp|CAH2_HUMAN_upsedyp                                     FALSE
#> P01008upsedyp|ANT3_HUMAN_upsedyp                                     FALSE
#> P01031upsedyp|CO5_HUMAN_upsedyp                                      FALSE
#> P01112upsedyp|RASH_HUMAN_upsedyp                                     FALSE
#> P01127upsedyp|PDGFB_HUMAN_upsedyp                                    FALSE
#> P01133upsedyp|EGF_HUMAN_upsedyp                                      FALSE
#> P01344upsedyp|IGF2_HUMAN_upsedyp                                     FALSE
#> P01375upsedyp|TNFA_HUMAN_upsedyp                                     FALSE
#> P01579upsedyp|IFNG_HUMAN_upsedyp                                     FALSE
#> P02144upsedyp|MYG_HUMAN_upsedyp                                      FALSE
#> P02741upsedyp|CRP_HUMAN_upsedyp                                      FALSE
#> P02753upsedyp|RET4_HUMAN_upsedyp                                     FALSE
#> P02787upsedyp|TRFE_HUMAN_upsedyp                                     FALSE
#> P02788upsedyp|TRFL_HUMAN_upsedyp                                     FALSE
#> P04040upsedyp|CATA_HUMAN_upsedyp                                     FALSE
#> P05413upsedyp|FABPH_HUMAN_upsedyp                                    FALSE
#> P06396upsedyp|GELS_HUMAN_upsedyp                                     FALSE
#> P06732upsedyp|KCRM_HUMAN_upsedyp                                     FALSE
#> P08263upsedyp|GSTA1_HUMAN_upsedyp                                    FALSE
#> P08758upsedyp|ANXA5_HUMAN_upsedyp                                    FALSE
#> P09211upsedyp|GSTP1_HUMAN_upsedyp                                    FALSE
#> P10145upsedyp|IL8_HUMAN_upsedyp                                      FALSE
#> P10599upsedyp|THIO_HUMAN_upsedyp                                     FALSE
#> P10636-8upsedyp|TAU_HUMAN_upsedyp                                    FALSE
#> P12081upsedyp|SYHC_HUMAN_upsedyp                                     FALSE
#> P15559upsedyp|NQO1_HUMAN_upsedyp                                     FALSE
#> P16083upsedyp|NQO2_HUMAN_upsedyp                                     FALSE
#> P41159upsedyp|LEP_HUMAN_upsedyp                                      FALSE
#> P51965upsedyp|UB2E1_HUMAN_upsedyp                                    FALSE
#> P55957upsedyp|BID_HUMAN_upsedyp                                      FALSE
#> P61626upsedyp|LYSC_HUMAN_upsedyp                                     FALSE
#> P61769upsedyp|B2MG_HUMAN_upsedyp                                     FALSE
#> P62937upsedyp|PPIA_HUMAN_upsedyp                                     FALSE
#> P62979upsedyp|RS27A_HUMAN_upsedyp                                    FALSE
#> P63165upsedyp|SUMO1_HUMAN_upsedyp                                    FALSE
#> P63279upsedyp|UBC9_HUMAN_upsedyp                                     FALSE
#> P68871upsedyp|HBB_HUMAN_upsedyp                                      FALSE
#> P69905upsedyp|HBA_HUMAN_upsedyp                                      FALSE
#> P99999upsedyp|CYC_HUMAN_upsedyp                                      FALSE
#> Q06830upsedyp|PRDX1_HUMAN_upsedyp                                    FALSE
#> Q15843upsedyp|NEDD8_HUMAN_upsedyp                                    FALSE
#> REV__CON__Q3KNV1;REV__CON__P08729;REV__CON__P07744                   FALSE
#> REV__CON__Q0VCM5                                                     FALSE
#> REV__CON__Q14CN4-1                                                   FALSE
#> REV__CON__Q2YDI2                                                     FALSE
#> REV__CON__Q7Z794                                                     FALSE
#> REV__CON__Q9DCV7                                                     FALSE
#> REV__sp|P0C155|CBPS2_YEAST                                           FALSE
#> REV__sp|P0CF34|YL016_YEAST                                           FALSE
#> REV__sp|P32380|SP110_YEAST                                           FALSE
#> REV__sp|P33314|BUD2_YEAST                                            FALSE
#> REV__sp|P36150|SUMT_YEAST                                            FALSE
#> REV__sp|P36775|LONM_YEAST                                            FALSE
#> REV__sp|P38279|YPQ3_YEAST                                            FALSE
#> REV__sp|P39526|LAA1_YEAST                                            FALSE
#> REV__sp|P39531|RCY1_YEAST                                            FALSE
#> REV__sp|P40012|PPOX_YEAST                                            FALSE
#> REV__sp|P40358|JEM1_YEAST                                            FALSE
#> REV__sp|P46678|TFC5_YEAST                                            FALSE
#> REV__sp|P53119|HUL5_YEAST                                            FALSE
#> REV__sp|P53746|FRE4_YEAST                                            FALSE
#> REV__sp|Q02932|KA120_YEAST                                           FALSE
#> REV__sp|Q03660|TR130_YEAST                                           FALSE
#> REV__sp|Q06639|RSC3_YEAST                                            FALSE
#> REV__sp|Q07660|BRE4_YEAST                                            FALSE
#> sp|A5Z2X5|YP010_YEAST                                                FALSE
#> sp|D6VTK4|STE2_YEAST                                                 FALSE
#> sp|O13297|CET1_YEAST                                                 FALSE
#> sp|O13516|RS9A_YEAST                                                 FALSE
#> sp|O13563|RPN13_YEAST                                                FALSE
#> sp|O13585|YP089_YEAST                                                FALSE
#> sp|P05745|RL36A_YEAST;sp|O14455|RL36B_YEAST                          FALSE
#> sp|O14467|MBF1_YEAST                                                 FALSE
#> sp|O14468|YO304_YEAST                                                FALSE
#> sp|O43137|YB085_YEAST                                                FALSE
#> sp|O60200|MDM35_YEAST                                                FALSE
#> sp|O74700|TIM9_YEAST                                                 FALSE
#> sp|O94742|SEM1_YEAST                                                 FALSE
#> sp|P00128|QCR7_YEAST                                                 FALSE
#> sp|P00330|ADH1_YEAST                                                 FALSE
#> sp|P00358|G3P2_YEAST                                                 FALSE
#> sp|P00359|G3P3_YEAST                                                 FALSE
#> sp|P00360|G3P1_YEAST                                                 FALSE
#> sp|P00427|COX6_YEAST                                                 FALSE
#>                                                    metacell_Intensity_D_R2
#> CON__A2I7N1;CON__A2I7N0                                              FALSE
#> CON__P00761                                                          FALSE
#> P02768upsedyp|ALBU_HUMAN_upsedyp;CON__P02768-1                       FALSE
#> CON__P04264                                                          FALSE
#> CON__P07477                                                          FALSE
#> CON__P13645                                                          FALSE
#> CON__P35527                                                          FALSE
#> CON__P35908                                                          FALSE
#> CON__P62894                                                          FALSE
#> CON__Q2KIS7                                                          FALSE
#> O00762upsedyp|UBE2C_HUMAN_upsedyp                                    FALSE
#> O76070upsedyp|SYUG_HUMAN_upsedyp                                     FALSE
#> P00167upsedyp|CYB5_HUMAN_upsedyp                                     FALSE
#> P00441upsedyp|SODC_HUMAN_upsedyp                                     FALSE
#> P00709upsedyp|LALBA_HUMAN_upsedyp                                    FALSE
#> P00915upsedyp|CAH1_HUMAN_upsedyp                                     FALSE
#> P00918upsedyp|CAH2_HUMAN_upsedyp                                     FALSE
#> P01008upsedyp|ANT3_HUMAN_upsedyp                                     FALSE
#> P01031upsedyp|CO5_HUMAN_upsedyp                                      FALSE
#> P01112upsedyp|RASH_HUMAN_upsedyp                                     FALSE
#> P01127upsedyp|PDGFB_HUMAN_upsedyp                                    FALSE
#> P01133upsedyp|EGF_HUMAN_upsedyp                                      FALSE
#> P01344upsedyp|IGF2_HUMAN_upsedyp                                     FALSE
#> P01375upsedyp|TNFA_HUMAN_upsedyp                                     FALSE
#> P01579upsedyp|IFNG_HUMAN_upsedyp                                     FALSE
#> P02144upsedyp|MYG_HUMAN_upsedyp                                      FALSE
#> P02741upsedyp|CRP_HUMAN_upsedyp                                      FALSE
#> P02753upsedyp|RET4_HUMAN_upsedyp                                     FALSE
#> P02787upsedyp|TRFE_HUMAN_upsedyp                                     FALSE
#> P02788upsedyp|TRFL_HUMAN_upsedyp                                     FALSE
#> P04040upsedyp|CATA_HUMAN_upsedyp                                     FALSE
#> P05413upsedyp|FABPH_HUMAN_upsedyp                                    FALSE
#> P06396upsedyp|GELS_HUMAN_upsedyp                                     FALSE
#> P06732upsedyp|KCRM_HUMAN_upsedyp                                     FALSE
#> P08263upsedyp|GSTA1_HUMAN_upsedyp                                    FALSE
#> P08758upsedyp|ANXA5_HUMAN_upsedyp                                    FALSE
#> P09211upsedyp|GSTP1_HUMAN_upsedyp                                    FALSE
#> P10145upsedyp|IL8_HUMAN_upsedyp                                      FALSE
#> P10599upsedyp|THIO_HUMAN_upsedyp                                     FALSE
#> P10636-8upsedyp|TAU_HUMAN_upsedyp                                    FALSE
#> P12081upsedyp|SYHC_HUMAN_upsedyp                                     FALSE
#> P15559upsedyp|NQO1_HUMAN_upsedyp                                     FALSE
#> P16083upsedyp|NQO2_HUMAN_upsedyp                                     FALSE
#> P41159upsedyp|LEP_HUMAN_upsedyp                                      FALSE
#> P51965upsedyp|UB2E1_HUMAN_upsedyp                                    FALSE
#> P55957upsedyp|BID_HUMAN_upsedyp                                      FALSE
#> P61626upsedyp|LYSC_HUMAN_upsedyp                                     FALSE
#> P61769upsedyp|B2MG_HUMAN_upsedyp                                     FALSE
#> P62937upsedyp|PPIA_HUMAN_upsedyp                                     FALSE
#> P62979upsedyp|RS27A_HUMAN_upsedyp                                    FALSE
#> P63165upsedyp|SUMO1_HUMAN_upsedyp                                    FALSE
#> P63279upsedyp|UBC9_HUMAN_upsedyp                                     FALSE
#> P68871upsedyp|HBB_HUMAN_upsedyp                                      FALSE
#> P69905upsedyp|HBA_HUMAN_upsedyp                                      FALSE
#> P99999upsedyp|CYC_HUMAN_upsedyp                                      FALSE
#> Q06830upsedyp|PRDX1_HUMAN_upsedyp                                    FALSE
#> Q15843upsedyp|NEDD8_HUMAN_upsedyp                                    FALSE
#> REV__CON__Q3KNV1;REV__CON__P08729;REV__CON__P07744                   FALSE
#> REV__CON__Q0VCM5                                                     FALSE
#> REV__CON__Q14CN4-1                                                   FALSE
#> REV__CON__Q2YDI2                                                     FALSE
#> REV__CON__Q7Z794                                                     FALSE
#> REV__CON__Q9DCV7                                                     FALSE
#> REV__sp|P0C155|CBPS2_YEAST                                           FALSE
#> REV__sp|P0CF34|YL016_YEAST                                           FALSE
#> REV__sp|P32380|SP110_YEAST                                           FALSE
#> REV__sp|P33314|BUD2_YEAST                                            FALSE
#> REV__sp|P36150|SUMT_YEAST                                            FALSE
#> REV__sp|P36775|LONM_YEAST                                            FALSE
#> REV__sp|P38279|YPQ3_YEAST                                            FALSE
#> REV__sp|P39526|LAA1_YEAST                                            FALSE
#> REV__sp|P39531|RCY1_YEAST                                            FALSE
#> REV__sp|P40012|PPOX_YEAST                                            FALSE
#> REV__sp|P40358|JEM1_YEAST                                            FALSE
#> REV__sp|P46678|TFC5_YEAST                                            FALSE
#> REV__sp|P53119|HUL5_YEAST                                            FALSE
#> REV__sp|P53746|FRE4_YEAST                                            FALSE
#> REV__sp|Q02932|KA120_YEAST                                           FALSE
#> REV__sp|Q03660|TR130_YEAST                                           FALSE
#> REV__sp|Q06639|RSC3_YEAST                                            FALSE
#> REV__sp|Q07660|BRE4_YEAST                                            FALSE
#> sp|A5Z2X5|YP010_YEAST                                                FALSE
#> sp|D6VTK4|STE2_YEAST                                                 FALSE
#> sp|O13297|CET1_YEAST                                                 FALSE
#> sp|O13516|RS9A_YEAST                                                 FALSE
#> sp|O13563|RPN13_YEAST                                                FALSE
#> sp|O13585|YP089_YEAST                                                FALSE
#> sp|P05745|RL36A_YEAST;sp|O14455|RL36B_YEAST                          FALSE
#> sp|O14467|MBF1_YEAST                                                 FALSE
#> sp|O14468|YO304_YEAST                                                FALSE
#> sp|O43137|YB085_YEAST                                                FALSE
#> sp|O60200|MDM35_YEAST                                                FALSE
#> sp|O74700|TIM9_YEAST                                                 FALSE
#> sp|O94742|SEM1_YEAST                                                 FALSE
#> sp|P00128|QCR7_YEAST                                                 FALSE
#> sp|P00330|ADH1_YEAST                                                 FALSE
#> sp|P00358|G3P2_YEAST                                                 FALSE
#> sp|P00359|G3P3_YEAST                                                 FALSE
#> sp|P00360|G3P1_YEAST                                                 FALSE
#> sp|P00427|COX6_YEAST                                                 FALSE
#>                                                    metacell_Intensity_D_R3
#> CON__A2I7N1;CON__A2I7N0                                              FALSE
#> CON__P00761                                                          FALSE
#> P02768upsedyp|ALBU_HUMAN_upsedyp;CON__P02768-1                       FALSE
#> CON__P04264                                                          FALSE
#> CON__P07477                                                          FALSE
#> CON__P13645                                                          FALSE
#> CON__P35527                                                          FALSE
#> CON__P35908                                                          FALSE
#> CON__P62894                                                          FALSE
#> CON__Q2KIS7                                                          FALSE
#> O00762upsedyp|UBE2C_HUMAN_upsedyp                                    FALSE
#> O76070upsedyp|SYUG_HUMAN_upsedyp                                     FALSE
#> P00167upsedyp|CYB5_HUMAN_upsedyp                                     FALSE
#> P00441upsedyp|SODC_HUMAN_upsedyp                                     FALSE
#> P00709upsedyp|LALBA_HUMAN_upsedyp                                    FALSE
#> P00915upsedyp|CAH1_HUMAN_upsedyp                                     FALSE
#> P00918upsedyp|CAH2_HUMAN_upsedyp                                     FALSE
#> P01008upsedyp|ANT3_HUMAN_upsedyp                                     FALSE
#> P01031upsedyp|CO5_HUMAN_upsedyp                                      FALSE
#> P01112upsedyp|RASH_HUMAN_upsedyp                                     FALSE
#> P01127upsedyp|PDGFB_HUMAN_upsedyp                                    FALSE
#> P01133upsedyp|EGF_HUMAN_upsedyp                                      FALSE
#> P01344upsedyp|IGF2_HUMAN_upsedyp                                     FALSE
#> P01375upsedyp|TNFA_HUMAN_upsedyp                                     FALSE
#> P01579upsedyp|IFNG_HUMAN_upsedyp                                     FALSE
#> P02144upsedyp|MYG_HUMAN_upsedyp                                      FALSE
#> P02741upsedyp|CRP_HUMAN_upsedyp                                      FALSE
#> P02753upsedyp|RET4_HUMAN_upsedyp                                     FALSE
#> P02787upsedyp|TRFE_HUMAN_upsedyp                                     FALSE
#> P02788upsedyp|TRFL_HUMAN_upsedyp                                     FALSE
#> P04040upsedyp|CATA_HUMAN_upsedyp                                     FALSE
#> P05413upsedyp|FABPH_HUMAN_upsedyp                                    FALSE
#> P06396upsedyp|GELS_HUMAN_upsedyp                                     FALSE
#> P06732upsedyp|KCRM_HUMAN_upsedyp                                     FALSE
#> P08263upsedyp|GSTA1_HUMAN_upsedyp                                    FALSE
#> P08758upsedyp|ANXA5_HUMAN_upsedyp                                    FALSE
#> P09211upsedyp|GSTP1_HUMAN_upsedyp                                    FALSE
#> P10145upsedyp|IL8_HUMAN_upsedyp                                      FALSE
#> P10599upsedyp|THIO_HUMAN_upsedyp                                     FALSE
#> P10636-8upsedyp|TAU_HUMAN_upsedyp                                    FALSE
#> P12081upsedyp|SYHC_HUMAN_upsedyp                                     FALSE
#> P15559upsedyp|NQO1_HUMAN_upsedyp                                     FALSE
#> P16083upsedyp|NQO2_HUMAN_upsedyp                                     FALSE
#> P41159upsedyp|LEP_HUMAN_upsedyp                                      FALSE
#> P51965upsedyp|UB2E1_HUMAN_upsedyp                                    FALSE
#> P55957upsedyp|BID_HUMAN_upsedyp                                      FALSE
#> P61626upsedyp|LYSC_HUMAN_upsedyp                                     FALSE
#> P61769upsedyp|B2MG_HUMAN_upsedyp                                     FALSE
#> P62937upsedyp|PPIA_HUMAN_upsedyp                                     FALSE
#> P62979upsedyp|RS27A_HUMAN_upsedyp                                    FALSE
#> P63165upsedyp|SUMO1_HUMAN_upsedyp                                    FALSE
#> P63279upsedyp|UBC9_HUMAN_upsedyp                                     FALSE
#> P68871upsedyp|HBB_HUMAN_upsedyp                                      FALSE
#> P69905upsedyp|HBA_HUMAN_upsedyp                                      FALSE
#> P99999upsedyp|CYC_HUMAN_upsedyp                                      FALSE
#> Q06830upsedyp|PRDX1_HUMAN_upsedyp                                    FALSE
#> Q15843upsedyp|NEDD8_HUMAN_upsedyp                                    FALSE
#> REV__CON__Q3KNV1;REV__CON__P08729;REV__CON__P07744                   FALSE
#> REV__CON__Q0VCM5                                                     FALSE
#> REV__CON__Q14CN4-1                                                   FALSE
#> REV__CON__Q2YDI2                                                     FALSE
#> REV__CON__Q7Z794                                                     FALSE
#> REV__CON__Q9DCV7                                                     FALSE
#> REV__sp|P0C155|CBPS2_YEAST                                           FALSE
#> REV__sp|P0CF34|YL016_YEAST                                           FALSE
#> REV__sp|P32380|SP110_YEAST                                           FALSE
#> REV__sp|P33314|BUD2_YEAST                                            FALSE
#> REV__sp|P36150|SUMT_YEAST                                            FALSE
#> REV__sp|P36775|LONM_YEAST                                            FALSE
#> REV__sp|P38279|YPQ3_YEAST                                            FALSE
#> REV__sp|P39526|LAA1_YEAST                                            FALSE
#> REV__sp|P39531|RCY1_YEAST                                            FALSE
#> REV__sp|P40012|PPOX_YEAST                                            FALSE
#> REV__sp|P40358|JEM1_YEAST                                            FALSE
#> REV__sp|P46678|TFC5_YEAST                                            FALSE
#> REV__sp|P53119|HUL5_YEAST                                            FALSE
#> REV__sp|P53746|FRE4_YEAST                                            FALSE
#> REV__sp|Q02932|KA120_YEAST                                           FALSE
#> REV__sp|Q03660|TR130_YEAST                                           FALSE
#> REV__sp|Q06639|RSC3_YEAST                                            FALSE
#> REV__sp|Q07660|BRE4_YEAST                                            FALSE
#> sp|A5Z2X5|YP010_YEAST                                                FALSE
#> sp|D6VTK4|STE2_YEAST                                                 FALSE
#> sp|O13297|CET1_YEAST                                                 FALSE
#> sp|O13516|RS9A_YEAST                                                 FALSE
#> sp|O13563|RPN13_YEAST                                                FALSE
#> sp|O13585|YP089_YEAST                                                FALSE
#> sp|P05745|RL36A_YEAST;sp|O14455|RL36B_YEAST                          FALSE
#> sp|O14467|MBF1_YEAST                                                 FALSE
#> sp|O14468|YO304_YEAST                                                FALSE
#> sp|O43137|YB085_YEAST                                                FALSE
#> sp|O60200|MDM35_YEAST                                                FALSE
#> sp|O74700|TIM9_YEAST                                                 FALSE
#> sp|O94742|SEM1_YEAST                                                 FALSE
#> sp|P00128|QCR7_YEAST                                                 FALSE
#> sp|P00330|ADH1_YEAST                                                 FALSE
#> sp|P00358|G3P2_YEAST                                                 FALSE
#> sp|P00359|G3P3_YEAST                                                 FALSE
#> sp|P00360|G3P1_YEAST                                                 FALSE
#> sp|P00427|COX6_YEAST                                                 FALSE