Computes the number of proteins that are only defined by specific peptides, shared peptides or a mixture of two.

This function computes few values about the adjacency matrix such as the number of proteins that are only defined by specific peptides, shared peptides or a mixture of two.

matAdjStats(X)

Arguments

X
The adjacency matrix with both specific and shared peptides.

Value

A list of values: * nbPeptides: the number of peptides in the matrix, nbSpecificPeptides: the number of specific peptides in the matrix, nbSharedPeptides: the number of shared peptides in the matrix, nbProt: the number of proteins in the matrix, protOnlyUniquePep: the list of proteins only defined by specific peptides, protOnlySharedPep: the list of proteins only defined by shared peptides, protMixPep: the list of proteins defined by both shared and specific peptides.

Author

Samuel Wieczorek

Examples

utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[seq_len(1000),]
obj <- addListAdjacencyMatrices(obj, 2)

#> Error in addListAdjacencyMatrices(obj, 2): impossible de trouver la fonction "addListAdjacencyMatrices"

X <- GetAdjMat(obj[[2]])$all
matAdjStats(X)

#> $nbPeptides
#> [1] 13919
#> 
#> $nbSpecificPeptides
#> [1] 13524
#> 
#> $nbSharedPeptides
#> [1] 395
#> 
#> $nbProt
#> [1] 172
#> 
#> $protOnlyUniquePep
#> character(0)
#> 
#> $protOnlySharedPep
#>   [1] "1000" "1016" "1017" "1076" "1077" "1082" "1103" "1111" "1121" "1122"
#>  [11] "1148" "115"  "116"  "1175" "1179" "1199" "1229" "1246" "1247" "125" 
#>  [21] "1253" "1254" "1280" "1303" "1304" "1315" "132"  "1351" "137"  "1382"
#>  [31] "1395" "1398" "1414" "145"  "1468" "1496" "151"  "152"  "155"  "1551"
#>  [41] "1556" "1569" "1577" "1578" "1585" "1596" "1597" "1609" "1612" "166" 
#>  [51] "167"  "168"  "1680" "1681" "1685" "169"  "170"  "1713" "174"  "176" 
#>  [61] "1772" "1782" "1808" "1835" "184"  "1841" "1847" "1882" "1935" "1971"
#>  [71] "1974" "2012" "2022" "204"  "205"  "2050" "2056" "210"  "213"  "2145"
#>  [81] "2177" "2231" "2233" "2248" "2260" "227"  "2271" "2310" "2316" "2323"
#>  [91] "2327" "2337" "2340" "2342" "2343" "2349" "2372" "245"  "247"  "251" 
#> [101] "252"  "257"  "262"  "281"  "289"  "290"  "293"  "295"  "3"    "301" 
#> [111] "302"  "306"  "362"  "390"  "399"  "40"   "411"  "418"  "427"  "430" 
#> [121] "49"   "490"  "495"  "505"  "512"  "522"  "523"  "532"  "54"   "603" 
#> [131] "608"  "609"  "610"  "643"  "644"  "645"  "651"  "652"  "667"  "691" 
#> [141] "7"    "717"  "722"  "758"  "759"  "762"  "776"  "779"  "798"  "799" 
#> [151] "8"    "829"  "830"  "84"   "845"  "850"  "869"  "881"  "909"  "910" 
#> [161] "915"  "916"  "944"  "95"   "96"   "97"   "972"  "976"  "977"  "98"  
#> [171] "986"  "987" 
#> 
#> $protMixPep
#> character(0)
#>