Bar plot of missing values per lines using highcharter.

This method plots a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins).

mvPerLinesHisto(qData)

mvPerLinesHistoPerCondition(qData, conds, pal.name)

mvHisto(qData, conds, showValues = FALSE, pal.name = NULL)

mod_mv_plots_ui(id)

mod_mv_plots_server(id, obj, conds, pal.name)

Arguments

qData

A data.frame or matrix that contains the quantitative data.

conds

A Character()

pal.name

xxx

id

xxx

obj

An instance of the class SummarizedExperiment

Value

A plot

Details

distribution of the missing values per line,

a bar plot which represents the distribution of the number of missing values (NA) per lines (ie proteins) and per conditions,

Histogram of missing values.

Author

Samuel Wieczorek, Enora Fremy

Examples

library(QFeatures)
library(DaparToolshed)
data(ft)

qData <- assay(ft, 1)
conds <- design(ft)$Condition
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("QFeatures", package = "QFeatures")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("design"), generic = structure("design", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "design"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function 'design' for signature '"QFeatures"'
pal <- 'Dark2'

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvPerLinesHisto(qData)

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvHisto(qData, conds, pal.name = pal)
#> Error in unique(conds): object 'conds' not found

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvPerLinesHistoPerCondition(qData, conds)
#> Error in unique(conds): object 'conds' not found
mvPerLinesHistoPerCondition(qData, conds, pal.name = pal)
#> Error in unique(conds): object 'conds' not found

#----------------------------------------
# Launch a shiny module
#----------------------------------------

if(interactive()){
 library(QFeatures)
 library(shiny)
 library(DaparToolshed)
 data(ft)
 ui <- mod_mv_plots_ui('plot')

 server <- function(input, output, session) {
  conds <- design(ft)$Condition
 
  mod_mv_plots_server('plot',
                      obj = reactive({ft[[1]]}),
                      conds = reactive({conds}),
                      pal.name = reactive({'Dark2'})
                      )
  }
 
 shinyApp(ui=ui, server=server)
}