Display a a jitter plot for CC

Display a jitter plot for connex component with the package highcharter

plotJitter_hc(df, clickFunction = NULL)

Arguments

df

A dataframe containing xxx

clickFunction

A JavaScript code to show an information when a point is clicked

Value

A plot

Author

Thomas Burger, Samuel Wieczorek

Examples

df <- data.frame(x=seq_len(10), y=seq_len(10))
plotJitter_hc(df)