Finalizes the aggregation process

Finalizes the aggregation process

# S4 method for QFeatures
qMetadata(object, i, slotName = "qMetadata")

qMetadata(object, i, slotName = "qMetadata") <- value

# S4 method for QFeatures
aggregateQmetadata(object, i, fcol, name = "newAssay", fun = aggQmeta, ...)

# S4 method for SummarizedExperiment
aggregateQmetadata(object, fun, conds)

aggQmeta(object, conds)

# S4 method for QFeatures
FinalizeAggregation(object, from, to, ...)

Arguments

object

An object of class 'SummarizedExperiment'

i

When adding an adjacency matrix to an assay of a `QFeatures` object, the index or name of the assay the adjacency matrix will be added to. Ignored when `x` is an `SummarizedExperiment`.

value

An adjacency matrix with row and column names. The matrix will be coerced to compressed, column-oriented sparse matrix (class `dgCMatrix`) as defined in the `Matrix` package, as generaled by the [sparseMatrix()] constructor.

conds

xxx

qMetaName

`character(1)` with the variable name containing the adjacency matrix. Default is `"qMetadata"`.

obj.pep

A peptide object of class MSnset

pepData

xxxx

X

An adjacency matrix in which lines and columns correspond respectively to peptides and proteins.

protData

xxxxx

protMetacell

xxx

Value

NA A protein object of class MSnset

Author

Samuel Wieczorek

Examples

Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
#> Warning: cannot open compressed file '', probable reason 'Invalid argument'
#> Error in gzfile(file, "rb"): cannot open the connection
obj <- Exp1_R25_pept[1:100]
#> Error in eval(expr, envir, enclos): object 'Exp1_R25_pept' not found
FinalizeAggregation(obj, 2)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'FinalizeAggregation': object 'obj' not found