Quantitative cell metadata scopes for filtering

This function looks for the lines that respect the request in either all conditions or at least one condition.

This function looks for the lines where each element respect the query.

This function looks for the lines that respect the request in either all conditions or at least one condition.

qMetadataFilteringScope()

SymFilteringOperators()

qMetadatWholeMatrix(object, cmd, pattern, percent, th, operator)

qMetadataWholeLine(object, cmd, pattern)

qMetadataOnConditions(object, cmd, mode, pattern, conds, percent, operator, th)

Arguments

object

xxx

cmd

A character(1) xxx

pattern

A character(1) xxx

percent

xxx

th

The theshold to apply

operator

String for operator to use. List of operators is available with SymFilteringOperators().

conds

xxx

mask

xxx

type

Available values are:

'AllCond' (the query is valid in all the conditions),

'AtLeatOneCond' (the query is valid in at leat one condition.

op

String for operator to use. List of operators is available with SymFilteringOperators().

Value

NA NA NA NA NA #' @examples data(ft) obj <- ft[1] mask <- match.qMetadata(metadata = qMetadata(obj), pattern = 'missing', level = typeOfData(obj) ) type <- 'AllCond' conds <- design(ft)$Condition op <- '>=' th <- 2 percent <- FALSE ind <- qMetadataOnConditions(mask, type, conds, percent, op, th)

Examples

data(ft)
#> Warning: data set 'ft' not found
obj <- ft[[1]]
#> Error in eval(expr, envir, enclos): object 'ft' not found
level <- typeOfData(ft, 1)
#> Error in typeOfData(ft, 1): could not find function "typeOfData"
pattern <- 'missing'
mask <- match.qMetadata(metadata = qMetadata(obj), 
                        pattern = pattern, 
                        level = level)
#> Error in match.qMetadata(metadata = qMetadata(obj), pattern = pattern,     level = level): unused argument (metadata = qMetadata(obj))
percent <- FALSE
th <- 3
op <- '>='
ind <- qMetadatWholeMatrix(mask, op, percent, th)
#> Error in qMetadatWholeMatrix(mask, op, percent, th): object 'mask' not found

data(ft)
#> Warning: data set 'ft' not found
obj <- ft[[1]]
#> Error in eval(expr, envir, enclos): object 'ft' not found
mask <- match.qMetadata(metadata = qMetadata(obj), 
                        pattern = 'missing POV', 
                        level = typeOfData(obj)
                        )
#> Error in match.qMetadata(metadata = qMetadata(obj), pattern = "missing POV",     level = typeOfData(obj)): unused argument (metadata = qMetadata(obj))
ind <- qMetadataWholeLine(mask)
#> Error in qMetadataWholeLine(mask): object 'mask' not found