Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 5.923700 6.587886 9.084109 9.314632 7.503160 9.985728 8.003701 8.476625
#>    [9] 8.815725 6.356415 9.823102 8.211886 9.390990 6.464962 5.810739 7.229930
#>   [17] 7.283777 8.881097 8.906803 9.957363 5.176343 5.368309 5.084592 6.850362
#>   [25] 9.013271 5.154174 7.550320 5.916799 8.308626 5.291257 5.506719 7.067262
#>   [33] 9.239059 6.626040 8.611558 7.816547 9.824658 6.433030 5.926429 8.503543
#>   [41] 6.051831 8.789166 9.975509 9.995857 7.794877 9.130096 8.932386 9.772736
#>   [49] 5.109249 8.979679 5.154138 6.166799 9.975096 7.940065 7.071421 9.475777
#>   [57] 5.400513 9.456689 9.049652 7.219238 5.835355 8.473454 7.607699 5.829023
#>   [65] 7.490508 6.506622 9.529444 9.803409 7.873595 9.902418 9.464807 5.425941
#>   [73] 7.380880 7.076306 9.063471 9.944132 8.754367 7.772478 9.655478 6.634402
#>   [81] 7.895802 8.279032 7.476086 7.961757 7.889303 8.778176 7.770426 8.456364
#>   [89] 6.835062 9.121823 9.592214 9.751665 7.922560 6.508612 7.220719 6.223319
#>   [97] 8.903770 7.756297 8.729663 7.303048 9.345833 7.261030 7.150519 9.596880
#>  [105] 8.602312 6.289265 6.735215 6.828622 9.578890 5.655455 5.872637 7.478722
#>  [113] 5.442771 9.364005 7.737153 7.359428 5.405186 9.706324 9.430676 7.698767
#>  [121] 7.037761 6.175742 6.264459 6.154907 9.067630 5.515447 7.582174 6.096010
#>  [129] 5.958204 8.512121 5.880240 7.121686 7.724717 9.966316 7.755810 5.985161
#>  [137] 6.341053 7.527333 7.099079 8.643422 7.895735 5.554100 9.303963 7.934571
#>  [145] 6.108205 6.346269 8.452872 5.717319 5.766249 6.962249 6.970106 5.409198
#>  [153] 8.751333 5.050077 8.094622 9.163323 5.562144 8.137697 8.798890 7.358365
#>  [161] 8.652095 8.747476 7.811226 5.661191 7.815617 9.549026 6.067748 8.023295
#>  [169] 7.115876 9.504835 5.984923 8.987851 8.717625 6.758626 7.893818 8.900834
#>  [177] 7.800861 8.124086 9.828491 7.106897 5.395882 7.920476 8.194714 7.860194
#>  [185] 7.557827 9.219965 6.535052 8.347486 7.756521 6.088327 7.661915 5.956027
#>  [193] 7.562797 8.311783 5.776495 8.158715 9.903461 5.596414 6.642451 8.853992
#>  [201] 5.057509 9.852723 8.453788 5.353976 7.630507 7.995211 7.297079 5.864041
#>  [209] 7.712269 5.492660 7.186083 6.036926 7.595831 7.669832 9.258337 5.911247
#>  [217] 7.565972 9.706531 6.802291 9.627693 7.779967 9.515389 7.824615 9.503286
#>  [225] 7.287693 7.277322 7.902682 8.891840 8.184116 7.398516 8.359299 9.161855
#>  [233] 7.403054 6.452794 9.853261 7.815023 7.081255 7.093165 7.040613 6.102820
#>  [241] 7.147468 9.283645 5.069816 8.017098 5.186867 8.401297 5.735877 6.823298
#>  [249] 9.305583 9.625671 8.074574 8.812569 6.794433 6.504406 7.496401 8.264041
#>  [257] 6.065608 8.203131 9.806796 9.753846 6.438774 7.006324 5.157219 9.596185
#>  [265] 6.023573 9.884453 9.833565 7.553336 8.673715 9.682799 6.303490 7.788231
#>  [273] 5.499564 6.765747 5.132144 5.770680 9.138548 7.076444 8.966153 5.501459
#>  [281] 7.327808 9.669450 5.395178 7.120730 9.170088 6.957259 9.523135 7.997942
#>  [289] 6.018439 6.564147 6.779649 9.979886 6.703137 9.810253 7.101725 9.570750
#>  [297] 6.854803 5.382721 6.620999 7.067603 9.802031 7.258773 8.201659 8.894013
#>  [305] 6.508583 6.930523 5.656547 7.157276 9.794054 9.801051 9.444618 8.245547
#>  [313] 5.302516 6.306299 6.530701 6.455566 6.116093 7.223499 5.941416 7.018475
#>  [321] 9.019123 6.616336 5.080320 6.155714 8.015042 6.270929 9.361753 6.044069
#>  [329] 7.954988 5.884054 8.293164 5.703630 7.526996 9.854145 7.793458 6.411176
#>  [337] 5.899349 5.765412 7.694629 6.175377 6.233589 6.615467 7.553793 7.121333
#>  [345] 8.382234 6.676060 5.420353 6.944190 9.236135 9.128990 6.916343 5.001160
#>  [353] 7.664097 6.045123 9.564720 6.112408 6.006525 9.694376 9.632479 5.558909
#>  [361] 8.765846 6.079655 8.694417 6.387384 6.935433 7.086236 6.224711 9.281127
#>  [369] 6.522672 7.328291 5.255717 5.860057 7.619468 9.686448 7.633720 8.921539
#>  [377] 6.501038 7.147684 9.521276 6.569282 9.374638 5.547351 9.209396 7.778222
#>  [385] 7.115573 9.501954 8.310925 7.516861 7.142179 7.952175 7.823102 5.555044
#>  [393] 8.741573 6.164310 5.713216 9.667859 7.083925 5.064430 5.793403 6.087955
#>  [401] 9.604000 7.330304 6.752505 6.955495 8.799042 8.877519 6.492135 8.710206
#>  [409] 8.577588 5.616180 9.584959 7.836944 7.880972 9.651940 7.569825 9.261274
#>  [417] 7.736253 6.125314 9.947437 5.768484 8.352136 7.692601 9.002636 6.850838
#>  [425] 9.074001 5.946702 6.551541 6.260495 5.291855 5.058669 5.334467 5.844440
#>  [433] 9.393119 6.161072 6.217601 5.508238 5.452985 7.093324 9.196519 6.591279
#>  [441] 8.839624 8.768171 7.889783 8.523665 9.893726 5.904296 5.279287 8.108627
#>  [449] 5.632466 9.070871 6.923708 9.132613 5.509771 7.450813 9.670031 5.877863
#>  [457] 9.987847 5.543394 7.541096 7.614143 9.912040 5.256408 7.860131 9.714425
#>  [465] 7.221211 6.637509 8.739000 8.760133 6.640157 7.832227 5.628638 6.932541
#>  [473] 7.915447 9.328243 9.941660 9.617285 6.461229 7.027755 9.167192 5.051472
#>  [481] 8.721321 6.862585 9.433955 8.112540 5.762262 6.662228 7.809754 5.934601
#>  [489] 8.556951 5.389364 5.851437 6.173298 5.331909 6.135697 5.176797 9.120364
#>  [497] 7.502110 7.877428 6.170410 8.991675 6.032558 5.138074 9.603014 5.674253
#>  [505] 6.702759 5.162584 7.371243 9.016612 5.873456 6.986083 6.038835 9.303680
#>  [513] 7.807152 5.456866 5.448662 7.201626 7.499064 9.171919 8.535003 6.445870
#>  [521] 7.248283 8.152650 9.664673 7.680655 9.310882 7.363583 8.268585 7.677209
#>  [529] 9.293615 8.302295 6.424786 6.324007 6.686845 8.719596 9.245507 8.124571
#>  [537] 7.896859 8.655313 8.485962 5.469224 7.872338 8.044949 8.248562 5.587674
#>  [545] 7.660837 7.486413 8.946759 5.017002 7.008325 6.797591 8.892865 8.823556
#>  [553] 8.263519 6.489723 7.365881 7.938905 9.254760 9.326837 8.241412 8.508805
#>  [561] 7.130994 6.143812 7.228816 5.007446 8.916086 6.807302 8.233185 7.602435
#>  [569] 7.769282 8.312314 7.781465 7.961021 5.508950 6.806186 9.923226 6.705370
#>  [577] 6.100503 9.927048 6.251675 5.997513 8.298610 7.251647 8.404661 7.963836
#>  [585] 7.261289 9.785431 7.002751 6.386778 6.896424 5.412070 6.870006 9.940222
#>  [593] 9.627752 6.833964 8.164922 8.216338 6.577392 9.320014 7.525628 5.020045
#>  [601] 5.854914 8.323729 8.630029 9.457570 7.914930 6.243652 8.372755 8.047921
#>  [609] 7.979972 5.594847 7.569195 8.352945 6.154937 7.423048 6.831106 9.542124
#>  [617] 7.356219 5.628554 8.825505 7.833318 6.966199 6.158243 8.386411 6.452046
#>  [625] 7.213208 7.461944 5.397952 5.889020 8.503683 7.573044 8.303372 7.013479
#>  [633] 8.730602 9.029254 7.629486 6.995798 5.612143 5.795334 6.490609 5.270578
#>  [641] 8.959411 5.247859 9.914664 8.294307 9.868500 8.412865 8.715641 8.068358
#>  [649] 8.645742 8.824139 6.082903 6.281281 5.790540 5.923546 9.317902 8.573145
#>  [657] 7.112973 6.230699 6.799984 6.550107 7.359509 7.660575 9.127745 5.755907
#>  [665] 9.153353 6.629886 7.572539 8.569911 6.672136 9.042204 8.089848 9.756514
#>  [673] 6.234733 8.412276 6.162737 9.030433 7.542732 7.656957 6.865673 6.738966
#>  [681] 7.320104 9.130845 6.810905 9.651193 5.996795 5.839795 9.324714 8.324969
#>  [689] 7.142738 5.434331 8.125082 9.981794 9.993309 8.473829 7.452370 7.018364
#>  [697] 7.667985 9.169454 6.745851 6.424048 5.606535 5.193330 6.117415 7.424644
#>  [705] 5.259163 9.421286 7.185081 9.615786 7.673590 9.728926 7.212562 8.823047
#>  [713] 9.289137 5.332514 7.135466 9.399868 5.615383 8.853033 5.904980 9.919376
#>  [721] 7.844775 9.373367 8.839200 7.827847 5.753686 7.323389 7.834857 9.376292
#>  [729] 9.613089 8.062899 5.455912 5.022507 6.192832 6.379016 9.208278 6.991038
#>  [737] 6.176806 5.110405 7.930780 7.368669 5.258116 9.176489 8.607269 6.102818
#>  [745] 8.268717 6.024000 9.179279 5.777193 7.695929 9.316932 6.229886 5.611713
#>  [753] 6.278959 7.351335 9.287003 9.238654 9.582073 7.440102 8.562425 7.772281
#>  [761] 7.215630 5.526752 9.613754 6.729851 9.994977 6.313540 7.464837 8.084409
#>  [769] 5.504605 5.007309 9.928938 9.746505 9.699448 9.488351 9.478131 5.846396
#>  [777] 7.265729 9.322650 8.813141 5.863572 9.378919 7.835673 9.178914 9.887424
#>  [785] 6.794781 8.437783 6.417436 8.524658 6.982650 7.794119 9.713260 6.135712
#>  [793] 7.880270 6.764916 7.726090 5.164430 8.871531 8.644954 8.724768 5.599925
#>  [801] 6.868392 9.225136 6.712813 9.554042 9.850513 5.545517 7.094723 6.823341
#>  [809] 6.383483 8.926657 8.054370 9.329370 8.968878 8.204646 9.187882 9.072573
#>  [817] 9.355833 9.550166 5.447624 5.211788 5.142408 7.441855 7.590224 6.797839
#>  [825] 8.436096 9.329916 8.340174 5.915458 8.421825 7.396626 5.450978 5.447640
#>  [833] 9.476243 8.611544 6.524478 8.370048 5.595985 8.364785 9.125338 8.068416
#>  [841] 9.873889 5.881000 5.241995 9.622924 6.297089 7.772763 5.419240 5.282533
#>  [849] 7.150932 8.701109 6.657637 7.543243 6.409674 7.219360 6.297465 7.115817
#>  [857] 9.650067 9.341004 5.885158 5.777758 5.709234 5.466694 6.639103 5.617428
#>  [865] 8.411805 9.632396 9.573395 6.817017 7.487602 8.277783 8.450250 7.116445
#>  [873] 5.895764 8.261227 7.538883 8.705802 7.683908 7.481310 6.093037 9.877593
#>  [881] 8.539798 7.683034 8.674391 9.687702 8.694668 7.942079 9.684106 8.923341
#>  [889] 9.965793 7.867943 6.463793 9.898696 9.720818 6.829915 5.343778 9.948196
#>  [897] 5.143712 8.370317 9.033106 7.504658 7.329500 8.271497 6.569504 6.401293
#>  [905] 9.189178 7.964210 9.270618 5.402312 7.134038 6.782203 7.484394 5.206828
#>  [913] 9.140503 5.125063 7.522747 5.584526 6.163278 6.285225 8.470056 5.378802
#>  [921] 7.605885 6.563485 5.631785 8.305472 9.318934 9.235944 8.193737 5.907081
#>  [929] 6.411543 6.985513 6.247603 5.562247 9.701844 6.743786 6.801073 7.629810
#>  [937] 6.350905 7.953239 5.654601 7.478024 8.369079 7.510179 6.402250 5.748106
#>  [945] 8.089142 7.169046 9.013066 6.295983 6.072102 9.970551 9.887442 8.073237
#>  [953] 7.405854 7.700057 8.217482 5.694665 9.953610 7.480244 5.456471 6.599824
#>  [961] 8.653074 7.820137 8.032762 9.926936 8.536778 5.852871 6.538888 5.558825
#>  [969] 8.221030 8.918950 5.745763 8.634521 5.528326 9.735880 9.701181 5.204062
#>  [977] 7.486542 7.959510 5.184387 7.324168 9.314688 5.730204 6.161979 6.357158
#>  [985] 5.699720 7.677728 8.195326 6.201689 7.275444 6.749791 9.577769 9.307075
#>  [993] 5.543224 6.462901 7.821098 9.937104 6.950563 7.786626 7.887637 6.629107