Visualize the clusters according to pvalue thresholds

Visualize the clusters according to pvalue thresholds

visualizeClusters(
  dat,
  clust_model,
  adjusted_pValues,
  FDR_th = NULL,
  ttl = "",
  subttl = ""
)

Arguments

dat

the standardize data returned by the function checkClusterability()

clust_model

the clustering model obtained with dat.

adjusted_pValues

vector of the adjusted pvalues obtained for each protein with a 1-way ANOVA (for example obtained with the function wrapperClassic1wayAnova()).

FDR_th

the thresholds of FDR pvalues for the coloring of the profiles. The default (NULL) creates 4 thresholds: 0.001, 0.005, 0.01, 0.05 For the sake of readability, a maximum of 4 values can be specified.

ttl

title for the plot.

subttl

subtitle for the plot.

Value

a ggplot object NA

Author

Hélène Borges

Examples

library(QFeatures)
utils::data(Exp1_R25_pept, package='DaparToolshedData')
#> Error in find.package(package, lib.loc, verbose = verbose): there is no package called 'DaparToolshedData'
test_anova <- wrapperClassic1wayAnova(Exp1_R25_pept)
#> Error in wrapperClassic1wayAnova(Exp1_R25_pept): object 'Exp1_R25_pept' not found
vizu <- visualizeClusters(dat = standardized_means,
                          clust_model = km_model,
                          adjusted_pValues = test_anova$P_Value$anova1way,
                          FDR_th = c(0.001,0.005,0.01,0.05),
                          ttl = "Clustering of protein profiles")
#> Error in cut(adjusted_pValues, breaks = c(-Inf, FDR_th, Inf), labels = c(str_try,     str_max)): object 'test_anova' not found