This function exports a QFeatures object to a Excel file.

This function exports a QFeatures data object to a Excel file. Each of the three data.frames in the QFeatures object (ie expression data, colData and rowData are respectively integrated into separate sheets in the Excel file). The colored cells in the experimental data correspond to the original missing values which have been imputed.

writeQFeaturesToExcel(obj, filename)

Arguments

obj
An object of class QFeatures.

filename
A character string for the name of the Excel file.

Value

A Excel file (.xlsx)

Author

Samuel Wieczorek

Examples

# \donttest{
Sys.setenv("R_ZIPCMD"= Sys.which("zip"))
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[seq_len(1000)]
writeQFeaturesToExcel(obj, "foo")

#> Error in openxlsx::addStyle(wb, sheet = i, cols = listMEC[, "col"] + 1,     rows = listMEC[, "row"] + 1, style = MEC_Style): objet 'listMEC' introuvable

# }