Aggregate an assay of peptides into proteins.

This function is a wrapper for the function QFeatures::aggregateFeatures(). It just add some meta informations to the aggregated assay, specific to DaparToolshed and implement methods to aggregate the quantitative metadata.

This function aggregates the quantitative features of an assay, applying a summarization function (fun) to sets of features. The fcol variable name points to a rowData column that defines how to group the features during aggregate. This variable can either be a vector (we then refer to an aggregation by vector) or an adjacency matrix (aggregation by matrix).

The quantitative metadata are aggregated with a function (fun.qmeta).

The list of agregation methods can be obtained with aggregateMethods(). This function compiles both methods from the packages DaparToolshed and QFeatures.

xxx

# S4 method for QFeatures
aggregateFeatures4Prostar(
  object,
  i,
  fcol,
  name = "newAssay",
  fun = MsCoreUtils::robustSummary,
  ...
)

# S4 method for SummarizedExperiment
aggregateFeatures4Prostar(
  object,
  fcol,
  fun = MsCoreUtils::robustSummary,
  conds,
  ...
)

aggQmetadata(qMeta, X, level, conds)

aggregateMethods()

Arguments

object

An object of class 'SummarizedExperiment'

i

The index or name of the assay which features will be aggregated the create the new assay.

fcol

A character(1) naming a rowdata variable (of assay i in case of a QFeatures) defining how to aggregate the features of the assay. This variable is either a character or a (possibly sparse) matrix. See below for details.

name

A character(1) naming the new assay. Default is newAssay. Note that the function will fail if there's already an assay with name.

fun

A function used for quantitative feature aggregation. See Details for examples.

...

Additional parameters passed the fun and fun.qmeta.

conds

A character() vector which is the names of conditions for each sample in the dataset.

X

xxxx

level

A character(1) which is the type of dataset

fun.qmeta

A function used for quantitative metadata aggregation. See Details for examples.

Value

A QFeatures object with an additional assay or a SummarizedExperiment object (or subclass thereof). xxxxx

Details

Iterative aggregation function

xxxxxx xxxxx

Quantitative metadata aggregation

xxxxxx xxxx

The function to aggregate the quantitative metadata is

aggQmetadat() xxxxx

See also

Examples

## ---------------------------------------
## An example QFeatures with PSM-level data
## ---------------------------------------
data(ft)
#> Warning: data set 'ft' not found
ft
#> Error in eval(expr, envir, enclos): object 'ft' not found

## Aggregate peptides into proteins
## using the adjacency matrix
ft <- ft <- aggregateFeatures4Prostar(object = ft,
                                      i = 1,
                                      name = 'aggregated', 
                                      fcol = 'adjacencyMatrix', 
                                      fun = colSumsMat,
                                      fun.qmeta = aggQmeta)
#> Error in aggregateFeatures4Prostar(object = ft, i = 1, name = "aggregated",     fcol = "adjacencyMatrix", fun = colSumsMat, fun.qmeta = aggQmeta): could not find function "aggregateFeatures4Prostar"
feat1
#> Error in eval(expr, envir, enclos): object 'feat1' not found

assay(feat1[[1]])
#> Error in assay(feat1[[1]]): could not find function "assay"
assay(feat1[[2]])
#> Error in assay(feat1[[2]]): could not find function "assay"
aggcounts(feat1[[2]])
#> Error in aggcounts(feat1[[2]]): could not find function "aggcounts"
assay(feat1[[3]])
#> Error in assay(feat1[[3]]): could not find function "assay"
aggcounts(feat1[[3]])
#> Error in aggcounts(feat1[[3]]): could not find function "aggcounts"
rowData(ft[[2]])
#> Error in rowData(ft[[2]]): could not find function "rowData"
data(ft)
#> Warning: data set 'ft' not found
qMeta <- qMetadata(ft, 1)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'qMetadata': object 'ft' not found
X <- adjacencyMatrix(ft, 1)
#> Error in adjacencyMatrix(ft, 1): could not find function "adjacencyMatrix"
level <- typeDataset(ft, 1)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'typeDataset': object 'ft' not found
conds <- colData(ft)$Condition
#> Error in colData(ft): could not find function "colData"
aggQmeta <- aggQmetadata(qMeta, X, level, conds)
#> Error in aggQmetadata(qMeta, X, level, conds): could not find function "aggQmetadata"