Additional accessors for instances of class Qfeatures.

These names are common to all assays contained in the object. This is why they are stored in the global metadata. This function is used whenever it is necessary to (re)detect MEC and POV (new dataset or when post processing protein qMetadata after aggregation)

# S4 method for QFeatures
qMetadata(object, i, slotName = "qMetadata")

# S4 method for SummarizedExperiment
qMetadata(object, slotName = "qMetadata")

qMetadata(object, i, slotName = "qMetadata") <- value

# S4 method for QFeatures
typeDataset(object, i, slotName = "typeDataset")

# S4 method for SummarizedExperiment
typeDataset(object, slotName = "typeDataset")

typeDataset(object, i, slotName = "typeDataset") <- value

# S4 method for QFeatures
idcol(object, i, slotName = "idcol")

# S4 method for SummarizedExperiment
idcol(object, slotName = "idcol")

idcol(object, i, slotName = "idcol") <- value

# S4 method for QFeatures
parentProtId(object, i, slotName = "parentProtId")

parentProtId(object, i, slotName = "parentProtId") <- value

# S4 method for QFeatures
analysis(object, i, slotName = "analysis")

# S4 method for SummarizedExperiment
analysis(object, slotName = "analysis")

analysis(object, i, slotName = "analysis") <- value

# S4 method for QFeatures
version(object, slotName = "version")

version(object, slotName = "version") <- value

# S4 method for QFeatures
design(object, slotName = "design")

design(object, slotName = "design") <- value

Arguments

object

An instance of class SummarizedExperiment or QFeatures.

i

The index or name of the assays to extract the quantitative metadata from. All must have a rowdata variable named as slotName

slotName

xxx

value

xxx

Details

Additional slots for rowdata of a SummarizedExperiment object:

qMetadata: xxx

Additional slots for Metadata for a QFeatures object:

xxx: xxxx

Additional slots for Metadata for a SummarizedExperiment object:

qMetadata: xxxx

parentProtId: xxx

idcol: xxxx

typeDataset: xxx

Quantitative metadata

Default slotName is "qMetadata". The value is an adjacency matrix with row and column names. The matrix will be coerced to compressed, column-oriented sparse matrix (class dgCMatrix) as defined in the Matrix package, as generaled by the sparseMatrix() constructor.

Examples

data(ft)
#> Warning: data set 'ft' not found
design(ft)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'design': object 'ft' not found