Exports a QFeatures object to a Excel file.

This function exports an instance of the class QFeatures to a Excel file. The resulting file is composed of four sheets:

quantitative data which contains the content of assay() object whith a color code for each cell w.r.t. to cell quantitative metadata.

metadata which is the content of rowData() with only one-dimensionnal data (i.e. the adjacencyMatrix and the qMetadata slots are not part of the sheet),

exp. design which is the content of colData(). Each condition in the table is colored with a different color,

quantitative metadata which is the content of qMetadata(). There is a color code for the different tags.

# S4 method for QFeatures
write2excel(object, i = NULL, filename = "newFile", ...)

# S4 method for SummarizedExperiment
write2excel(object, filename, exp.design, ...)

Arguments

filename

A character string for the name of the Excel file.

obj

An object of class QFeatures.

Value

A Excel file.

Author

Samuel Wieczorek

Examples

# \donttest{
data(ft)
#> Warning: data set 'ft' not found

#---------------------------------------
# Export the whole dataset
#---------------------------------------

write2excel(ft, filename = 'foo')
#> Error in write2excel(ft, filename = "foo"): could not find function "write2excel"
write2excel(ft, 1, 'foo')
#> Error in write2excel(ft, 1, "foo"): could not find function "write2excel"
# }