Filter a peptide assay on the basis of its adjacency matrix.

These functions filters (delete) peptides of an assay, applying a function on peptides and proteins. They can be used alone but the usual usage is to create an instance of a class FunctionFilter and to pass it to the function filterFeaturesOneSE in order to create a new assay, embedded into the QFeatures object.

AdjMatFilters

AdjMatFilters()

allPeptides(object, ...)

specPeptides(object, ...)

subAdjMat_specificPeptides(X)

sharedPeptides(object, ...)

subAdjMat_sharedPeptides(X)

topnFunctions()

topnPeptides(object, fun, top)

subAdjMat_topnPeptides(X, qData, fun, top)

Format

An object of class function of length 1.

Arguments

object

An object of class SummarizedExperiment

fun

A list() of additional parameters

top

A integer(1) which is the number of xxx

Details

This function builds an intermediate matrix with scores for each peptide based on 'fun' parameter. Once this matrix is built, one select the 'n' peptides which have the higher score

The list of filter functions is given by adjMatFilters():

specPeptides(): returns a new assay of class SummazizedExperiment with only specific peptides;

sharedpeptides(): returns a new assay of class SummazizedExperiment with only shared peptides;

topnPeptides(): returns a new assay of class SummazizedExperiment with only the 'n' peptides which best satisfies the condition. The condition is represented by functions which calculates a score for each peptide among all samples. The list of these functions is given by topnFunctions():

See also

Author

Samuel Wieczorek

Examples

#------------------------------------------------
# This function will keep only specific peptides
#------------------------------------------------

f1 <- FunctionFilter('specPeptides', list())

#------------------------------------------------
# This function will keep only shared peptides
#------------------------------------------------

f2 <- FunctionFilter('sharedPeptides', list())

#------------------------------------------------
# This function will keep only the 'n' best peptides
# w.r.t the quantitative sum of each peptides among 
# all samples
#------------------------------------------------

f3 <- FunctionFilter('topnPeptides', fun = 'rowSums', top = 2)

#------------------------------------------------------
# To run the filter(s) on the dataset, use [xxx()]
# IF several filters must be used, store them in a list
#------------------------------------------------------

data(ft)
#> Warning: data set 'ft' not found
lst.filters <- list()
lst.filters <- append(lst.filters, f1)
lst.filters <- append(lst.filters, f3)
 
ft <- filterFeaturesOneSE(object = ft,
                          i = 1,
                          name = 'filtered', 
                          filters = lst.filters
                          )
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'filterFeaturesOneSE': object 'ft' not found