Compare two assays

This plot compares the quantitative proteomics data between two assays. It can be used for example to compare the effect of the normalization process.

The comparison is made with the division operator.

Usage

plotCompareAssays(
  obj,
  i,
  j,
  info = NULL,
  pal.name = "Set1",
  subset.view = NULL,
  n = 100,
  type = "scatter",
  FUN = NULL
)

Arguments

obj

An instance of the class omXplore.

i

A numeric matrix containing quantitative data after normalization.

j

A numeric matrix containing quantitative data after normalization

info

xxx

pal.name

xxx

subset.view

xxx

n

xxx

type

The type of plot. Available values are 'scatter' (default) or 'line'

FUN

xxx

Value

A plot

Author

Samuel Wieczorek, Enora Fremy

Examples

data(Exp1_R25_prot, package="DaparToolshedData")
plotCompareAssays(Exp1_R25_prot, 1, 2, n = 5)