Density plots of logFC values

This function show the density plots of Fold Change (the same as calculated by limma) for a list of the comparisons of conditions in a differential analysis.

Usage

hc_logFC_DensityPlot(df_logFC, th_logFC = 0, pal = NULL)

Arguments

df_logFC

A dataframe that contains the logFC values

th_logFC

The threshold on log(Fold Change) to distinguish between differential and non-differential data

pal

xxx

Value

A highcharts density plot

Author

Samuel Wieczorek

Examples

library(SummarizedExperiment)
data(Exp1_R25_pept, package="DaparToolshedData")
obj <- Exp1_R25_pept
# Simulate missing value imputation
assay(obj[[1]])[which(is.na(assay(obj[[1]])))] <- 0
assay(obj[[2]])[which(is.na(assay(obj[[2]])))] <- 0

qData <- as.matrix(assay(obj[[2]]))
sTab <- SummarizedExperiment::colData(obj)
limma <- limmaCompleteTest(qData, sTab)
#> Warning: Zero sample variances detected, have been offset away from zero
pal <- ExtendPalette(2, "Dark2")
hc_logFC_DensityPlot(limma$logFC, th_logFC = 1, pal = pal)
#> Warning: The color palette has not the same dimension as the 
#>                 number of samples