Similar to the function is.na but focused on the equality with the paramter 'type'.

Similar to the function is.na but focused on the equality with the paramter 'type'.

is.OfType(data, type)

Arguments

data

A data.frame

type

The value to search in the dataframe

Value

A boolean dataframe

Author

Samuel Wieczorek

Examples

library(QFeatures)
data(ft)
#> Warning: data set 'ft' not found
obj <- ft[[1]]
#> Error in eval(expr, envir, enclos): object 'ft' not found
data <- qMetadata(obj)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'qMetadata': object 'obj' not found
is.OfType(as.data.frame(data), "MEC")
#> Warning: The parameter param is not a data.frame.
#> NULL