Search lines which respects request on one or more conditions.

This function looks for the lines that respect the request in either all conditions or at least one condition.

Usage

SymFilteringOperators()

qMetacellFilteringScope()

qMetacellWholeMatrix(
  object,
  cmd,
  pattern,
  percent = "Percentage",
  th,
  operator
)

qMetacellWholeLine(object, cmd, pattern)

qMetacellOnConditions(
  object,
  cmd,
  mode,
  pattern,
  conds,
  percent = "Percentage",
  operator,
  th
)

Arguments

object

xxx

cmd

Available values are:

pattern

ss

percent

xxx

th

The threshold to apply

operator

String for operator to use. List of operators is available with 'SymFilteringOperators()'.

mode

xxx

conds

xxx

Value

A vector of operators

NA

A vector of filtering scopes

NA

NA

NA

Examples

SymFilteringOperators()
#>    lessthan        less greaterthan     greater       equal   different 
#>        "<="         "<"        ">="         ">"        "=="        "!=" 

data(ft, package='DaparToolshed')
obj <- ft[[1]]
level <- typeDataset(obj)
pattern <- "Missing"
mask <- match.metacell(
    metadata = qMetacell(obj),
    pattern = pattern,
    level = level
)
percent <- FALSE
th <- 3
op <- ">="
cmd <- 'delete' 
ind <- qMetacellWholeMatrix(obj, cmd, pattern, percent, th, op)

data(ft, package='DaparToolshed')
ind <- qMetacellWholeLine(obj, cmd, pattern)

conds <- design.qf(ft)$Condition
op <- ">="
th <- 0.5
percent <- "Percentage"
mode <- "AllCond"
ind <- qMetacellOnConditions(obj, cmd, mode, pattern, conds, percent, op, th)

qMetacellFilteringScope()
#> [1] "None"           "WholeLine"      "WholeMatrix"    "AllCond"       
#> [5] "AtLeastOneCond"

data(ft_na)
obj <- ft_na[[1]]