Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 9.774674 8.571805 8.251604 7.778575 5.141571 6.504491 5.441515 7.990560
#>    [9] 7.415997 7.691578 6.594492 9.849188 7.678640 6.990911 9.308754 6.348039
#>   [17] 5.614800 5.492842 5.436608 8.511694 5.385240 7.785954 9.625497 6.989869
#>   [25] 8.274854 5.814787 7.783212 7.980475 8.033260 9.301921 6.365214 8.615755
#>   [33] 7.274503 6.739380 9.501626 8.622259 8.534725 6.192334 9.534294 8.966338
#>   [41] 5.184206 7.564789 6.257595 8.701252 5.814729 8.552210 6.824994 5.697412
#>   [49] 6.791313 8.382538 8.456266 6.298807 9.859739 7.218248 5.549465 8.711499
#>   [57] 5.872593 8.584746 6.953312 5.624634 5.952076 8.245908 6.556177 8.540521
#>   [65] 5.029451 9.641455 8.457817 7.101513 7.782034 7.422543 6.233799 7.103983
#>   [73] 9.646150 9.832260 9.231399 6.048673 9.379214 8.637448 5.513582 6.225828
#>   [81] 9.694870 5.282886 5.310257 8.949152 6.762894 5.587841 7.299380 5.307911
#>   [89] 8.663450 8.453848 5.108523 8.682022 6.419471 5.323054 6.072859 8.807149
#>   [97] 5.943145 5.934531 7.303359 7.724950 5.789886 8.899084 6.831916 6.494653
#>  [105] 9.959147 9.675571 5.533120 6.023864 7.052947 7.278504 8.802244 8.849872
#>  [113] 5.369036 7.745157 5.827391 5.870200 5.413580 9.354375 6.121857 6.127649
#>  [121] 5.295380 6.564765 6.051611 8.164995 6.527849 5.208751 9.858800 5.353588
#>  [129] 6.845558 9.513297 6.226105 9.647240 6.980143 7.817397 9.787426 8.406292
#>  [137] 8.918312 9.619978 8.927235 8.723620 7.864747 7.372613 7.847337 8.871762
#>  [145] 8.717896 7.516734 7.466708 7.460512 9.519979 5.487852 9.084835 5.925489
#>  [153] 6.005773 5.802018 5.836118 8.553613 9.419784 8.223700 8.632294 7.649387
#>  [161] 9.046489 8.240672 5.874674 8.641659 6.080373 5.847957 9.218190 5.330311
#>  [169] 8.458126 5.050083 6.488764 7.504720 9.625217 5.486913 5.236534 6.642945
#>  [177] 6.173236 9.455225 9.913146 6.157078 8.510438 8.168095 5.182014 9.958343
#>  [185] 8.824640 7.426558 7.281275 6.891961 7.361359 6.927841 8.322411 6.217591
#>  [193] 9.184176 9.884422 5.801580 7.523021 6.924199 9.585466 6.341495 7.284367
#>  [201] 8.897420 9.834843 6.656748 7.970572 6.638986 7.362131 7.985641 8.431074
#>  [209] 9.977243 7.946572 7.487511 6.473206 9.033433 5.409516 6.337070 9.924798
#>  [217] 7.601229 7.231869 5.574388 8.901114 7.022743 8.311689 5.756022 6.350274
#>  [225] 7.294737 6.259096 9.074011 7.781393 7.163584 8.383512 8.465911 8.437492
#>  [233] 6.450472 9.251993 6.951385 8.420493 7.954214 7.223774 7.490794 8.355341
#>  [241] 7.007101 5.022972 9.963559 8.762342 7.750602 6.259333 6.035861 9.936345
#>  [249] 8.399054 5.268835 8.129670 5.972392 8.521020 6.161506 6.273733 5.022182
#>  [257] 8.874999 8.376136 5.419636 8.385079 9.498990 6.001308 9.049338 6.429731
#>  [265] 6.982414 6.933667 6.835830 6.078425 9.822462 9.412439 5.801850 7.692100
#>  [273] 5.426254 5.541158 5.422381 8.099404 7.603058 6.179929 5.559093 9.424790
#>  [281] 7.986650 5.945560 5.785624 8.183245 5.569057 8.615022 5.500954 6.412121
#>  [289] 8.726096 5.049936 7.013154 5.231181 5.207507 5.446559 5.793324 8.125350
#>  [297] 8.984590 5.560520 6.157142 8.606718 6.945668 7.933001 9.314865 7.945381
#>  [305] 6.368050 5.029059 5.840628 5.717569 7.893054 9.487938 7.036311 5.746388
#>  [313] 8.886033 9.106081 7.742145 8.132076 7.255241 7.555108 8.723972 7.663616
#>  [321] 8.325142 7.855623 9.387136 6.430917 5.198219 9.522901 6.923574 5.268189
#>  [329] 6.327924 5.381517 6.556827 5.601731 7.729823 7.800106 7.046307 6.179850
#>  [337] 5.861996 7.274223 9.415515 7.860960 8.257526 7.705611 6.379658 7.085286
#>  [345] 7.564041 7.391262 8.597577 5.748030 7.547479 7.241549 7.793744 8.753194
#>  [353] 6.137929 8.071408 8.916202 9.859713 8.754706 6.783406 6.751440 9.606109
#>  [361] 8.140981 8.041613 7.231319 7.315077 5.250476 5.057656 5.487856 8.021389
#>  [369] 5.765218 6.820488 6.161823 7.671671 9.553120 9.122611 6.076699 7.157010
#>  [377] 9.706703 5.622946 6.422735 7.946758 8.325121 8.564923 7.266094 5.237768
#>  [385] 8.616230 5.021597 5.493728 6.724569 7.939457 6.078237 6.937655 5.258244
#>  [393] 9.722795 9.344919 9.007621 9.453118 7.143217 7.090915 8.221403 7.154010
#>  [401] 5.823476 6.172696 5.827455 7.569146 7.446284 8.787413 7.895220 9.670035
#>  [409] 7.942011 7.290613 7.374776 9.994492 8.754461 9.029229 5.021208 9.015461
#>  [417] 6.227725 9.970116 7.744249 5.558141 6.048951 8.241698 8.034341 6.228703
#>  [425] 8.888223 7.272792 5.136723 8.233415 8.075609 8.246327 7.260010 9.026060
#>  [433] 5.881301 8.724432 5.657700 9.772678 5.852766 5.663385 9.475171 7.427874
#>  [441] 9.684474 8.238710 8.641492 6.693863 7.973188 8.685567 7.343027 7.541990
#>  [449] 5.400285 7.191637 9.013701 7.464461 5.686409 5.479318 9.026614 6.292909
#>  [457] 6.986734 8.959504 7.922582 5.609356 9.482324 5.278346 9.120176 7.547649
#>  [465] 8.278472 8.839784 7.940141 5.332740 8.029438 5.857488 8.209486 8.091129
#>  [473] 6.919939 7.014022 8.502146 5.482607 7.392146 8.644988 6.229351 9.555556
#>  [481] 8.205809 7.837257 6.440285 6.603368 6.049086 7.829643 5.351984 6.397756
#>  [489] 8.138841 6.235449 7.687763 7.872811 6.237139 5.914072 5.442966 8.084456
#>  [497] 6.365111 5.858676 5.307049 7.109067 9.167077 5.976386 8.888262 8.628959
#>  [505] 9.808910 9.278731 5.651204 8.784703 8.733348 6.428280 8.474590 9.291183
#>  [513] 8.334322 5.614424 5.161909 7.088790 5.283504 7.304269 7.173133 5.498878
#>  [521] 5.417366 5.303070 9.365417 5.739450 5.426224 9.516466 8.225239 9.534394
#>  [529] 5.892053 8.767296 9.958449 8.139163 9.958725 7.203285 5.199750 8.790449
#>  [537] 9.269752 6.846090 6.462155 8.936703 6.470375 9.702711 9.718495 7.586624
#>  [545] 5.347453 7.553540 6.886704 8.915369 5.315247 7.407159 7.815013 9.372070
#>  [553] 7.227617 9.647436 8.665723 5.405116 8.683529 6.785261 6.623958 8.498941
#>  [561] 5.321578 9.617321 5.406799 6.164483 8.352087 5.077725 7.459451 6.761062
#>  [569] 6.114089 8.709071 6.710444 7.819088 5.622133 7.840996 5.052869 9.641921
#>  [577] 7.847261 5.610590 5.760366 8.308805 8.651095 5.980415 6.610677 6.332602
#>  [585] 7.452934 7.538593 5.353140 9.690138 8.671065 8.695420 5.443864 7.899967
#>  [593] 5.676619 9.524825 6.741030 7.743294 8.769561 8.383360 7.392871 6.264764
#>  [601] 6.305507 7.495197 8.705564 8.142725 7.002464 5.799400 7.966915 9.045908
#>  [609] 6.930686 6.767598 7.706092 8.605719 6.090044 7.689392 7.698971 5.996814
#>  [617] 9.985999 6.938796 6.603242 9.000927 7.683594 7.232727 8.066884 8.708223
#>  [625] 7.136511 6.764607 8.787658 9.594615 6.961271 6.900673 7.860852 6.688059
#>  [633] 7.659838 5.248813 9.163336 5.760773 7.077420 6.893978 9.188626 6.695207
#>  [641] 6.675038 8.562815 7.500181 8.399148 9.990825 7.606890 6.821189 5.748591
#>  [649] 5.064728 5.316308 8.800043 7.785429 8.841132 6.708043 5.516555 8.620139
#>  [657] 5.070842 8.369198 8.963534 9.306174 6.483580 6.680851 6.875750 9.815204
#>  [665] 6.975495 9.858753 8.563896 7.772894 8.473108 7.887105 6.821299 6.388682
#>  [673] 7.748941 9.016719 9.460070 6.836034 5.163679 5.676639 9.767835 7.401507
#>  [681] 8.195450 5.615787 6.629298 9.251388 8.983519 8.797796 6.387087 9.514509
#>  [689] 6.015088 7.001567 7.776346 8.698471 6.812247 5.335101 6.439545 8.572470
#>  [697] 8.781534 8.183847 6.989370 9.911520 5.011306 8.450948 8.634293 8.145499
#>  [705] 5.399415 9.085767 8.083705 9.767093 5.368527 5.189870 7.204773 8.996960
#>  [713] 5.433341 6.763851 5.440954 5.106102 6.408214 7.089806 9.345163 8.877268
#>  [721] 6.612729 7.007798 7.663527 7.947986 7.756541 7.099856 5.264642 7.158529
#>  [729] 8.881779 9.252336 7.321116 8.861609 5.408313 9.180129 5.455766 9.449053
#>  [737] 6.396825 9.783780 9.550258 9.425727 7.430354 6.775724 6.050170 7.049938
#>  [745] 5.932127 6.279040 9.309246 8.208449 6.498662 7.481629 7.684417 9.342868
#>  [753] 5.882265 6.848003 6.099828 9.865723 7.240554 5.162475 6.276629 9.659553
#>  [761] 5.042563 7.695710 8.468306 7.710284 6.516876 9.543581 6.749905 9.251528
#>  [769] 7.153850 6.706045 6.592483 9.315750 8.159190 8.715580 7.849603 5.177655
#>  [777] 7.007751 6.466704 5.236388 6.917576 7.401379 5.252081 6.457064 9.878000
#>  [785] 5.075605 8.853358 5.700870 6.162144 7.055793 8.786795 9.411588 8.854237
#>  [793] 9.141426 8.131057 8.285706 5.831079 5.110788 9.613990 9.720151 6.353336
#>  [801] 6.235643 6.492500 5.739467 7.409651 8.556649 5.628895 8.568534 8.337270
#>  [809] 8.950999 8.948855 7.882575 9.102777 8.539144 6.266655 7.104386 9.501758
#>  [817] 6.195533 7.746876 6.129994 7.871334 7.184715 7.947071 6.486158 5.045657
#>  [825] 7.208213 5.344938 9.709313 5.467134 5.506216 9.606968 7.118988 5.255182
#>  [833] 8.017386 8.572815 9.707325 6.501335 8.012479 5.590534 8.942719 5.096485
#>  [841] 6.657877 5.190258 8.360291 5.036198 8.089913 5.055457 6.381275 8.919614
#>  [849] 8.846378 8.293019 8.134485 8.355640 8.144556 9.056963 8.124721 8.596610
#>  [857] 8.831300 8.608569 6.302096 8.457503 6.442283 6.767003 9.534614 9.930587
#>  [865] 5.816784 9.290584 6.746230 5.644113 5.539784 6.749672 7.484051 9.103898
#>  [873] 9.899395 7.500514 7.378619 9.900632 5.373483 8.616169 7.648809 7.798890
#>  [881] 8.243606 5.620987 9.911052 7.918289 9.360578 8.811755 5.951853 8.739048
#>  [889] 5.853905 6.467743 5.294897 5.422814 7.163781 6.364585 8.305708 9.557988
#>  [897] 7.943073 6.609635 5.086547 9.462240 6.364600 6.539826 5.598967 7.997012
#>  [905] 9.732428 7.242599 6.764181 8.843432 7.503727 6.964124 8.038640 8.100363
#>  [913] 6.919010 5.701180 5.343857 9.977113 9.575073 7.020597 8.515989 8.137331
#>  [921] 7.901822 8.250435 8.484516 8.277965 5.016543 5.449303 5.077338 5.374554
#>  [929] 7.645176 6.244041 5.967539 5.061836 5.985311 8.735660 5.751455 9.589084
#>  [937] 8.453685 7.357177 9.650911 6.636174 8.978958 6.760991 6.467894 5.686177
#>  [945] 7.099772 9.817757 7.582049 6.065475 7.027135 7.066827 6.325044 5.323341
#>  [953] 5.184356 6.084476 5.181994 8.112216 6.320634 8.667475 8.032905 6.590325
#>  [961] 6.145000 7.795591 7.579781 8.782919 7.192030 9.232046 5.434167 9.814469
#>  [969] 7.490396 5.393215 9.662269 7.368793 7.006781 6.917513 8.263573 7.524782
#>  [977] 5.517137 9.196531 9.946146 5.423224 6.635648 9.805543 7.306184 8.183586
#>  [985] 7.319237 5.360946 7.815522 7.981234 9.131886 8.985840 8.934948 8.106240
#>  [993] 5.447977 7.426562 9.850720 6.674479 8.402710 6.563411 9.133971 5.105461