Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 8.144773 9.829983 8.095988 7.915229 8.846555 8.216512 5.832411 6.127046
#>    [9] 5.753320 7.102589 5.752735 5.370166 5.037093 9.005501 6.616109 6.001379
#>   [17] 9.081832 9.098204 6.261899 6.514728 8.496235 8.378456 9.872102 5.771501
#>   [25] 9.795666 7.942054 5.575360 6.312106 7.239428 5.458835 7.658362 7.418285
#>   [33] 8.829006 6.914419 9.735965 8.234654 9.146400 9.973688 8.918504 8.374968
#>   [41] 5.170487 9.252885 5.730984 9.758896 6.611337 9.553484 6.928860 7.083815
#>   [49] 5.467096 6.087330 9.300362 9.246820 9.490844 7.537145 7.508233 9.129319
#>   [57] 8.992549 5.042578 5.104618 5.368335 6.330919 8.972183 8.288680 7.003240
#>   [65] 5.639944 7.592842 8.239217 7.947146 8.914546 5.636347 9.293427 9.697191
#>   [73] 5.656929 6.262117 6.819790 8.179145 8.197862 7.121081 8.712145 5.925431
#>   [81] 9.277773 7.952344 9.684109 7.316264 9.468067 5.194272 7.638208 6.912049
#>   [89] 7.570766 6.521422 8.999854 5.660239 6.830135 7.632548 9.244194 8.751773
#>   [97] 6.222273 6.523376 6.952749 6.518738 9.538192 6.927230 9.704293 5.759103
#>  [105] 7.444564 6.487001 5.439296 5.609263 8.263442 6.285478 9.073126 7.928387
#>  [113] 6.707834 8.650883 5.435998 8.145925 8.365482 5.497669 6.314875 8.200236
#>  [121] 7.110981 6.105232 6.160371 5.476809 8.913298 8.000684 8.548820 6.577162
#>  [129] 5.700438 7.502915 9.494706 8.865423 6.553318 7.207907 7.504130 7.065998
#>  [137] 8.316734 6.551680 7.106205 9.533211 5.404114 7.507712 5.945690 7.700404
#>  [145] 9.555750 9.449308 5.642536 5.381984 8.649771 6.710585 7.879338 9.077545
#>  [153] 9.085210 9.669660 8.041382 9.690320 9.382057 5.258838 7.022920 5.631403
#>  [161] 7.477879 9.152341 7.212118 8.723610 8.082362 6.682721 9.764539 6.656167
#>  [169] 7.199863 5.115845 5.215262 6.063913 8.257495 7.087326 7.828505 8.949372
#>  [177] 7.029532 8.881916 5.547334 9.271658 5.029711 6.817644 6.113342 9.125062
#>  [185] 7.943875 9.240949 7.891629 9.940340 8.813787 8.011541 9.621138 6.177287
#>  [193] 6.328823 5.740841 8.970357 6.645082 8.774970 5.372052 6.434269 6.388869
#>  [201] 7.856348 8.718781 6.527381 9.023310 9.633089 7.934287 6.989759 7.527590
#>  [209] 8.442939 5.228142 9.244827 7.697248 8.778556 6.164611 8.764389 7.904507
#>  [217] 5.312759 7.141851 9.629864 7.889628 5.472649 5.224931 8.736998 7.384896
#>  [225] 9.317927 8.367340 5.359276 5.979417 5.539566 8.361483 7.276734 7.021317
#>  [233] 9.545963 5.129226 5.606869 8.181201 8.929545 6.737483 9.893141 8.142420
#>  [241] 6.640219 9.982429 6.024482 7.657159 5.312444 5.297725 8.718421 8.040587
#>  [249] 5.364014 9.439433 5.925370 8.498318 8.082352 8.643044 5.700808 6.291726
#>  [257] 6.119486 8.155102 5.474416 5.470989 5.551618 9.101898 9.862716 8.261130
#>  [265] 8.732586 8.603930 5.880309 8.290554 5.593681 7.115615 6.515881 5.520093
#>  [273] 5.302178 8.863287 7.023891 8.397917 9.866821 6.452158 9.517661 9.623053
#>  [281] 5.364728 6.295495 8.289429 9.258814 8.799952 8.945429 8.724279 9.229796
#>  [289] 9.523499 6.057695 7.780315 7.918071 8.058839 6.204684 9.762301 6.169310
#>  [297] 8.222575 9.004866 6.914987 8.527710 5.140032 6.065497 9.988797 6.638787
#>  [305] 9.759159 6.250252 6.983647 5.339302 9.815780 5.016970 7.784671 7.386910
#>  [313] 8.757087 6.743790 7.889122 7.076221 8.182914 6.086604 7.175908 5.452840
#>  [321] 8.404320 6.704883 7.306070 9.947534 8.740156 8.855007 6.787263 6.391808
#>  [329] 5.019110 8.090331 6.995884 8.269343 8.475716 6.293977 6.249129 9.539207
#>  [337] 8.830892 8.344462 5.125330 7.106777 7.437235 7.338343 7.385081 9.515129
#>  [345] 5.849471 6.222424 7.391858 8.344796 7.139244 8.565173 6.575928 9.531110
#>  [353] 5.530466 5.119966 8.028082 5.093682 8.719183 9.806855 5.995718 6.076919
#>  [361] 5.669398 8.422343 9.444604 5.610310 9.659960 5.180570 8.324361 7.263797
#>  [369] 6.235452 7.519412 6.573736 6.887242 5.394222 7.452245 6.750447 8.991134
#>  [377] 7.892244 5.365071 9.087026 8.799796 8.644095 7.170070 8.615132 5.507486
#>  [385] 7.585034 8.074834 7.473405 5.859977 6.441615 8.271486 6.757636 8.654345
#>  [393] 6.990493 6.590856 8.951026 6.718211 6.011538 9.003460 8.079744 6.100477
#>  [401] 9.706084 7.711962 5.338128 9.355584 9.572305 5.564546 9.827582 7.025078
#>  [409] 5.645598 8.546384 8.811289 8.578189 7.659940 6.824275 8.408810 7.096084
#>  [417] 5.519179 7.488534 5.329478 6.676314 8.228170 7.173655 7.188340 9.833742
#>  [425] 5.762552 6.405165 8.601883 5.619217 8.300535 7.500054 9.974604 6.649971
#>  [433] 8.119769 6.920916 7.657906 5.379619 9.033861 8.327405 8.577134 5.645305
#>  [441] 5.201333 6.635670 8.102940 9.908155 7.516661 9.451709 7.191863 9.894343
#>  [449] 7.171749 9.484478 9.930140 6.140332 8.070554 7.825611 8.520595 6.083572
#>  [457] 5.426471 7.243064 6.510155 5.988219 6.554583 6.896233 5.708280 5.614552
#>  [465] 9.457841 6.911693 5.123917 8.076218 6.853384 5.758566 5.283631 5.486060
#>  [473] 9.734460 6.725932 7.510893 6.861897 8.915019 8.721359 6.533854 6.809729
#>  [481] 9.929414 5.846743 8.284388 6.704573 6.729849 7.915460 8.830077 7.546431
#>  [489] 6.190918 6.503935 9.420972 7.166666 7.995606 9.421706 5.486524 9.120465
#>  [497] 7.773294 5.655410 9.388971 9.195686 8.706824 5.065287 7.556025 9.694802
#>  [505] 9.663691 7.600386 9.609538 6.803705 5.869285 9.596091 8.400388 8.900440
#>  [513] 9.515989 8.397973 6.772978 5.324319 8.026987 7.754238 8.841955 6.733477
#>  [521] 5.202492 9.970128 6.537086 7.755020 5.102698 7.431819 8.798058 8.502452
#>  [529] 7.157475 7.830330 8.235128 7.590619 7.057277 7.226345 5.735872 8.641575
#>  [537] 8.864262 9.096658 7.570301 9.222084 8.023219 6.543023 7.931134 6.824177
#>  [545] 6.450629 8.080885 8.317830 8.191786 9.497562 5.596541 6.061016 8.280685
#>  [553] 8.950515 6.633543 7.212503 6.900008 5.194606 5.862264 6.619143 9.319309
#>  [561] 5.106783 8.846471 9.683482 8.247346 6.495600 6.326919 5.115894 7.357340
#>  [569] 9.905030 8.730933 5.164799 9.935041 8.595567 6.592454 9.097297 6.531125
#>  [577] 6.671721 8.844638 6.323777 8.887060 6.965554 8.406503 6.739835 8.205810
#>  [585] 5.787569 8.987854 5.793163 9.505769 5.331445 9.647829 8.575773 7.819960
#>  [593] 5.812885 6.100871 5.456891 8.413469 9.849521 9.390260 6.588295 6.374422
#>  [601] 5.764740 7.327206 5.321005 6.414682 5.254544 5.029438 7.637122 8.689406
#>  [609] 8.447737 7.279788 8.411147 7.757018 6.338110 5.901680 7.108713 6.980800
#>  [617] 6.945719 6.367691 9.331349 5.298931 7.643315 8.207101 5.621087 9.037371
#>  [625] 9.788279 8.764669 5.671370 8.641672 5.524265 8.946561 5.806475 7.379318
#>  [633] 6.224306 5.014840 7.183138 9.089165 6.517509 8.172800 6.363991 5.055398
#>  [641] 5.366472 5.844636 6.402762 5.753172 6.135845 9.295385 7.153795 9.453318
#>  [649] 9.998876 9.218147 5.969760 7.257014 7.365679 8.946935 9.722142 5.023419
#>  [657] 7.735630 5.546151 5.089821 8.258472 6.125862 6.344551 7.639883 9.119392
#>  [665] 9.343764 6.262562 6.086722 6.313651 6.441037 6.904791 8.780719 9.565421
#>  [673] 7.634294 5.253576 8.033329 9.894147 5.627733 6.058174 9.926598 8.347637
#>  [681] 7.370647 8.379904 5.277814 9.399603 5.024890 8.855053 6.248705 7.728771
#>  [689] 9.697719 9.097520 8.786333 9.449349 5.919544 9.615981 5.026650 6.220730
#>  [697] 8.085106 5.848761 7.321007 5.811942 5.568239 8.607298 5.364713 9.733503
#>  [705] 5.688505 9.558631 7.453929 9.450498 6.636970 8.425043 8.031298 6.303138
#>  [713] 6.131396 7.047896 9.636081 5.139070 5.339596 6.798868 8.910059 5.188916
#>  [721] 8.914246 9.484019 5.624796 8.719892 5.937977 9.437687 5.407354 6.158295
#>  [729] 7.389745 7.489564 5.994728 5.210726 9.983191 9.017908 5.569785 9.579018
#>  [737] 5.250709 6.522040 7.948524 8.464146 5.666022 6.871749 9.748515 6.241478
#>  [745] 6.129006 5.608603 9.143883 7.373254 8.736318 9.259964 6.970865 8.513983
#>  [753] 9.099200 5.296611 7.840011 7.405566 5.364202 8.151667 8.811390 7.983314
#>  [761] 8.470299 9.533632 6.028148 5.814821 7.782843 6.416276 9.289264 8.533997
#>  [769] 7.090530 9.521481 8.183667 5.481951 5.564173 6.380957 8.591304 8.253247
#>  [777] 5.382123 5.588158 9.168989 5.978983 7.166380 6.958122 7.489271 9.860417
#>  [785] 8.811759 7.878807 7.204558 8.489577 7.106194 7.409892 6.865588 6.510493
#>  [793] 9.605804 7.875642 8.833844 9.064749 6.104691 6.678723 9.945181 5.949823
#>  [801] 9.696181 6.696443 8.056153 9.419919 6.995103 7.401696 6.822412 5.772953
#>  [809] 8.085707 5.719464 8.649885 5.141197 9.502777 7.957282 6.349411 6.286829
#>  [817] 5.158451 5.506803 6.049600 5.288149 5.928217 9.583633 9.702469 8.501902
#>  [825] 6.992906 5.394214 9.158198 7.207265 7.023267 8.484488 9.387752 5.956000
#>  [833] 6.685881 5.281048 9.216610 8.849166 7.524360 5.438170 5.051610 9.861497
#>  [841] 7.272800 9.150446 9.166593 8.368701 5.530578 5.059965 5.380391 6.555062
#>  [849] 9.011665 9.032781 7.690035 7.027774 6.799027 7.306296 6.577778 8.498711
#>  [857] 9.135289 8.452443 5.592896 7.258710 7.530572 7.095843 9.616816 6.906324
#>  [865] 6.815128 9.355873 8.402480 6.470491 9.198711 7.478366 6.179541 5.556219
#>  [873] 6.484477 8.001097 5.875362 9.045937 9.248058 8.346175 8.745348 7.414514
#>  [881] 8.373476 7.565742 8.358703 7.914474 5.356113 7.793685 6.436633 7.805637
#>  [889] 9.704290 6.234623 7.905496 8.283596 6.666080 7.526842 6.094532 6.776744
#>  [897] 9.910760 8.404174 5.454455 8.617842 5.411736 8.450940 7.334622 6.303296
#>  [905] 9.340934 7.697167 5.771557 6.412605 6.849495 7.235000 8.785450 5.260513
#>  [913] 8.419368 8.486354 6.286079 8.916486 7.411527 8.987048 8.356600 7.072752
#>  [921] 5.486633 9.333537 6.388048 5.648996 5.910091 6.632523 7.344916 8.400103
#>  [929] 7.725576 5.245699 7.699777 6.190461 5.389210 9.106743 9.142403 6.546822
#>  [937] 7.895359 7.440172 9.452813 6.578683 7.915304 7.059768 8.767670 6.734647
#>  [945] 6.677066 6.441892 6.548788 9.735265 6.285338 5.926493 8.534513 6.739455
#>  [953] 6.634452 6.833585 9.631758 9.108973 8.795297 5.174702 6.763261 9.013241
#>  [961] 8.777780 9.347472 5.794692 9.530014 8.388894 5.407623 8.403671 9.888794
#>  [969] 5.395978 5.417349 7.718556 7.936355 6.819854 5.664019 8.522930 6.888223
#>  [977] 7.676164 6.438487 5.334231 8.569559 5.456014 6.031176 6.664033 7.724191
#>  [985] 8.034776 5.462145 9.843053 9.249367 7.318406 5.199281 5.209347 6.798676
#>  [993] 5.420919 7.335235 7.934275 6.382946 6.551219 5.641683 7.566678 7.942728