Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 5.513421 6.124006 5.007568 6.632324 5.591847 8.592462 5.055938 9.145320
#>    [9] 5.437575 8.722721 5.605743 9.473978 6.314823 7.092120 6.455601 6.855496
#>   [17] 7.712717 8.400727 5.592299 5.831300 6.534844 6.672607 6.407471 8.978034
#>   [25] 8.405932 5.143626 9.959496 6.464933 6.413798 8.066770 7.578803 8.289645
#>   [33] 8.785184 8.673495 5.793585 9.665898 9.332946 5.789116 7.761723 5.435956
#>   [41] 7.298215 8.916847 5.775607 7.903380 5.234346 9.715873 9.938305 7.436391
#>   [49] 7.023709 9.770063 7.769389 5.675589 5.984975 7.459981 7.039582 8.586848
#>   [57] 6.766355 6.228439 5.111695 6.030828 5.722546 8.331374 7.380072 8.883130
#>   [65] 9.931863 8.967635 8.357696 6.292873 5.035775 6.581990 8.002302 7.437494
#>   [73] 9.838698 5.887952 6.198358 8.257530 9.605923 6.089384 6.682737 7.305815
#>   [81] 6.995338 8.870853 9.119040 8.322169 6.654522 6.186067 6.592746 5.871307
#>   [89] 6.656887 6.308805 7.578411 7.255058 6.165029 9.837555 9.348450 5.659190
#>   [97] 6.518139 8.838918 9.171973 8.691806 6.083766 6.981325 7.832235 9.198190
#>  [105] 7.126079 6.178451 7.712637 8.785354 7.693155 5.029516 8.408768 6.181830
#>  [113] 6.305079 5.612035 9.121939 8.685243 5.496096 5.179723 6.804091 9.147641
#>  [121] 5.711414 5.960879 5.963955 6.204692 7.236520 6.267699 7.303708 5.773487
#>  [129] 8.836505 9.082810 9.068282 5.178383 9.985078 8.622901 7.429168 7.420383
#>  [137] 7.475458 6.842512 8.292205 6.319517 9.882800 6.619899 9.938239 6.247104
#>  [145] 9.851316 6.668828 5.097638 9.908804 7.520655 7.263517 6.496230 9.272981
#>  [153] 9.564598 6.886127 9.774329 5.336602 5.177856 5.746714 6.077464 6.479041
#>  [161] 9.513969 8.824659 6.352910 7.111036 5.704686 9.541184 7.643925 9.819229
#>  [169] 8.743013 9.422891 9.346325 7.865594 7.873649 8.253484 5.290937 8.313499
#>  [177] 9.129786 9.701296 8.843159 6.807589 8.997858 9.322238 8.156446 6.972833
#>  [185] 6.037856 6.182326 7.957272 8.918512 6.136721 9.770871 5.425437 8.557042
#>  [193] 5.436276 7.226191 6.729762 5.094644 8.634542 6.963557 9.498964 9.063497
#>  [201] 7.241019 8.573504 9.192250 7.239371 6.992136 5.278186 6.158115 9.382383
#>  [209] 8.881653 7.384508 5.977022 9.641998 9.755635 7.212362 8.457437 9.702987
#>  [217] 8.868689 6.228452 9.834173 8.196439 9.678577 9.558182 6.851366 8.204932
#>  [225] 5.377518 6.031967 7.974572 6.377663 6.575947 9.628722 8.773666 5.518965
#>  [233] 6.796234 7.519326 8.508608 8.351246 7.861117 7.965990 7.948824 7.564325
#>  [241] 8.058248 8.416692 7.149224 9.945968 6.398357 7.821524 5.079434 8.779663
#>  [249] 9.416527 8.968069 5.434514 8.560517 8.961112 8.719151 6.411545 9.570748
#>  [257] 9.234657 6.370713 6.007234 5.990421 5.261544 6.743488 8.756313 6.425136
#>  [265] 5.952550 7.560004 7.806486 9.827655 7.415760 6.079589 7.456884 9.304522
#>  [273] 9.478901 6.411467 7.138388 7.341998 8.957227 5.083907 8.760498 6.925558
#>  [281] 8.883523 5.501148 6.379496 7.490978 6.850882 8.736816 7.949168 6.649338
#>  [289] 8.284582 8.418856 6.708676 7.334569 9.269084 7.858261 7.267953 7.595722
#>  [297] 9.494951 7.716196 9.879419 8.416370 7.501817 7.080118 6.663022 5.378807
#>  [305] 5.602989 5.075034 7.803662 9.780982 7.597848 8.915711 8.453151 8.005340
#>  [313] 7.648328 6.266439 9.350802 5.965393 7.063873 6.926445 6.888232 6.336868
#>  [321] 6.290186 5.812900 6.158538 9.465375 6.511712 7.702195 8.340760 5.579634
#>  [329] 7.352429 8.159782 9.357436 7.576910 9.110613 8.748574 5.629700 7.324333
#>  [337] 5.466302 7.997092 7.105065 6.596884 9.397244 5.125631 7.754005 9.761336
#>  [345] 5.621922 9.688034 9.011532 5.451547 6.634563 6.392824 9.513092 9.523748
#>  [353] 5.306567 9.920372 8.416628 9.998763 8.244211 9.166163 8.880760 7.663772
#>  [361] 6.674496 9.552813 6.972291 9.855076 6.564058 9.330831 9.045109 8.757005
#>  [369] 7.219328 7.545151 6.974374 8.768798 9.426861 7.964074 6.010601 5.322773
#>  [377] 6.677418 9.644558 6.194547 7.367338 5.603487 5.075979 5.479712 8.341093
#>  [385] 7.214337 5.076826 7.781954 5.415341 6.091323 5.307316 5.982242 5.220144
#>  [393] 9.026568 6.189071 5.113660 6.584216 9.745797 5.862340 8.928577 8.064910
#>  [401] 7.810348 6.275923 9.663201 6.129249 5.646497 6.659944 9.715392 8.087264
#>  [409] 9.484004 5.398040 7.668867 5.630292 7.245778 5.515664 6.151726 6.803200
#>  [417] 8.658278 9.694868 9.761099 7.918735 6.328781 5.658338 8.400296 5.907942
#>  [425] 9.894952 7.946795 6.097112 5.671917 9.220653 9.775436 9.727797 5.183864
#>  [433] 9.664841 8.458229 9.435146 6.411861 9.705393 6.151831 5.671136 8.688580
#>  [441] 6.969716 8.212698 5.250827 7.260120 5.295193 7.562924 8.358220 8.091487
#>  [449] 8.017891 7.636255 6.310726 6.335505 7.533463 8.646499 5.209889 8.276198
#>  [457] 6.165385 9.458695 5.570950 9.470418 8.714036 9.183724 5.233809 6.060966
#>  [465] 7.303820 7.674623 9.074186 7.453819 6.381615 9.413992 5.784584 9.805041
#>  [473] 6.552161 6.195402 9.651527 5.885000 8.810562 7.230504 8.034206 7.957374
#>  [481] 5.409934 7.301571 7.139073 7.983007 8.663643 9.625293 8.755242 9.929731
#>  [489] 9.521788 7.104097 8.540945 7.974634 6.181636 5.293839 6.299731 9.993578
#>  [497] 6.038499 8.599888 6.059897 5.782865 7.312441 6.000668 6.024270 5.907511
#>  [505] 5.001006 7.053973 9.981172 9.708872 6.877239 9.671282 5.488216 5.112566
#>  [513] 6.552199 8.649603 7.450203 5.624306 6.403205 7.802293 5.407899 9.858913
#>  [521] 8.181810 7.893768 9.703175 6.112133 9.657017 7.857138 9.331199 8.301375
#>  [529] 6.580087 7.845093 5.553694 8.153274 7.093717 8.989319 6.621818 5.253521
#>  [537] 7.705827 5.881517 9.675307 7.211533 6.899979 6.263259 5.199265 5.500964
#>  [545] 5.982511 5.481397 9.766825 8.185211 9.248604 7.929540 7.813757 6.822291
#>  [553] 5.307025 6.777398 5.228288 5.889312 6.732885 9.932675 5.274235 8.175458
#>  [561] 9.357680 7.973213 9.550334 8.946091 5.048703 7.727610 5.508963 8.215107
#>  [569] 9.342841 8.458356 7.132412 5.486752 5.468855 5.325178 8.063410 8.433162
#>  [577] 7.154389 7.305597 5.360262 8.667939 9.764812 9.136571 8.848388 5.673566
#>  [585] 7.650851 9.668244 6.627145 9.257626 5.034575 6.776767 7.046827 7.746196
#>  [593] 7.205763 8.316553 7.653843 9.740785 9.144847 8.069688 6.211695 9.194479
#>  [601] 7.364311 5.999859 9.797153 6.148042 6.352546 9.663535 9.345017 7.856313
#>  [609] 8.168790 7.759294 8.760283 7.776058 7.433573 7.635333 7.524562 8.821240
#>  [617] 7.653664 8.822173 7.511176 7.761348 9.320731 5.013735 6.104903 6.149555
#>  [625] 6.561075 6.830530 9.960313 5.843832 6.988903 5.499390 6.895967 8.562241
#>  [633] 9.767321 5.778498 8.172156 9.620622 8.465292 6.774174 5.092345 7.481700
#>  [641] 9.283099 7.383387 7.177490 6.203170 9.021122 5.980811 6.510359 6.882432
#>  [649] 9.510066 6.495839 5.832823 5.384533 7.831989 7.513273 8.102337 5.749433
#>  [657] 9.022970 5.426891 6.507529 7.988386 7.162738 8.145691 7.194952 6.713873
#>  [665] 6.026708 6.125592 8.015870 8.186425 6.691723 9.945348 9.922222 7.949958
#>  [673] 7.364047 8.470841 9.338917 5.181623 5.065910 9.004151 7.014597 8.211085
#>  [681] 9.162870 5.943978 7.055684 8.898129 8.917336 6.730725 8.227401 6.526540
#>  [689] 7.442690 5.785370 5.044201 6.519788 9.997665 9.392601 7.740269 7.247588
#>  [697] 6.502107 9.164725 7.515819 9.802996 5.236772 6.869165 6.160091 5.111204
#>  [705] 5.492482 8.046980 6.048543 5.556270 5.840115 6.667912 7.591281 9.307701
#>  [713] 7.827102 9.895432 5.627365 5.381044 7.677701 9.365293 9.634573 6.271372
#>  [721] 5.073785 8.820278 5.906812 9.453740 7.379598 5.052706 7.708710 9.522394
#>  [729] 8.718716 5.100911 5.663022 5.484100 9.657284 5.107282 7.735299 8.437213
#>  [737] 5.624381 8.052158 5.839732 5.694868 5.402637 9.735883 9.518250 9.282509
#>  [745] 9.700815 7.780308 5.523579 7.687545 7.417800 7.065067 8.850699 6.620697
#>  [753] 7.722427 5.820184 6.477041 6.786588 5.369815 9.443558 6.177471 6.239924
#>  [761] 9.700851 6.727401 5.649290 8.388910 5.344262 9.881143 8.193204 8.892341
#>  [769] 9.525814 6.654289 7.928529 5.562621 5.300805 6.448359 6.995184 8.392044
#>  [777] 6.155261 9.661168 9.504975 7.922128 6.925166 6.662801 5.225951 8.489916
#>  [785] 9.826933 7.388967 6.923100 8.243295 8.127874 9.843986 7.805591 6.676575
#>  [793] 7.276972 8.752963 9.118981 8.216633 7.291910 8.140583 7.269664 7.780297
#>  [801] 5.154296 7.303943 6.660475 6.591860 9.110467 7.551118 6.817493 9.287791
#>  [809] 9.933232 8.299110 7.064555 6.098910 7.110263 9.476474 8.395963 8.122765
#>  [817] 5.101078 9.085037 7.011460 7.239495 9.212964 6.789444 9.638565 9.184238
#>  [825] 5.594365 6.082686 5.442716 6.514196 7.922925 5.726359 7.037318 6.318175
#>  [833] 5.210829 6.995268 5.252048 8.427908 5.073140 9.264598 5.564402 9.659271
#>  [841] 5.881036 7.200428 9.529591 5.155776 7.165755 7.970663 6.096533 5.294105
#>  [849] 5.884224 5.639194 7.008172 6.369857 8.319196 8.655353 5.479538 6.440216
#>  [857] 9.248947 8.330540 7.879348 9.938546 9.960522 5.112167 8.905661 6.909034
#>  [865] 6.936691 9.018851 8.191442 6.087517 7.035633 5.426267 6.758400 5.971140
#>  [873] 6.796554 5.188079 7.377343 6.994133 8.046589 7.830714 5.363570 5.563424
#>  [881] 7.092339 5.529629 6.569477 6.106101 7.155469 7.391692 6.919070 9.434731
#>  [889] 8.217382 5.357768 8.646285 8.702063 9.250665 5.943887 7.468705 9.706666
#>  [897] 8.844530 7.933379 6.687791 8.432028 9.089509 7.942945 9.619246 6.178916
#>  [905] 9.405416 8.221988 5.514308 8.776118 9.469295 5.289761 7.954117 8.494366
#>  [913] 9.163745 9.955818 9.418230 9.205485 7.252599 7.822908 8.430449 8.421505
#>  [921] 5.743703 9.966037 9.068643 8.523891 7.968673 7.211912 6.012880 8.301968
#>  [929] 9.338913 5.851491 9.866333 7.095129 8.267725 6.747598 5.826331 7.026747
#>  [937] 8.229286 9.673643 7.790905 8.207011 5.237820 6.657013 8.571608 6.771036
#>  [945] 6.585710 9.474848 6.610160 8.251477 5.551410 5.961780 9.152197 6.014299
#>  [953] 8.207352 8.821985 5.467605 7.962532 6.756859 5.360777 5.751630 8.817627
#>  [961] 9.368929 7.351797 7.291776 9.163266 9.490202 6.322914 6.274359 8.041879
#>  [969] 6.612549 8.824167 7.789927 8.133662 6.406539 9.516360 8.161944 5.324783
#>  [977] 8.570243 6.757539 8.183473 5.823334 5.350365 7.628700 7.466844 7.623663
#>  [985] 6.700355 9.169166 8.152008 5.889505 8.701296 8.807227 8.787373 7.187072
#>  [993] 9.874504 8.869321 8.741399 6.903324 7.436220 5.184194 5.308928 9.141286