Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 9.117680 7.269730 8.001004 9.807388 6.734531 7.858939 5.340092 7.407357
#>    [9] 6.426267 7.680870 9.417957 6.592090 8.526577 9.324752 5.125308 8.509468
#>   [17] 8.864006 6.302380 7.092525 8.498706 8.469163 9.208647 9.042760 7.837233
#>   [25] 6.845552 9.839111 5.713047 5.668259 5.855580 8.076175 8.957468 7.982310
#>   [33] 8.878978 7.529350 9.541088 8.682594 8.549294 5.673453 7.464515 7.755657
#>   [41] 8.603643 8.385798 9.717646 9.165309 8.141336 7.627748 6.343838 6.089268
#>   [49] 9.489632 7.833697 6.568375 6.012407 7.310524 9.312734 6.114670 5.681168
#>   [57] 5.500164 8.254583 5.849491 7.714603 7.135877 5.982945 7.001797 6.314891
#>   [65] 5.728599 6.878382 7.887040 9.169984 5.305413 5.074513 6.874625 8.743896
#>   [73] 8.015536 7.286650 6.874950 6.469436 8.698811 5.936645 6.845404 7.922582
#>   [81] 8.165590 7.757711 6.622312 5.976637 9.898518 6.490677 8.302532 7.544020
#>   [89] 6.211106 8.209634 6.343171 7.882040 5.163507 8.933622 9.337856 5.630688
#>   [97] 9.378074 9.122687 9.450437 5.677777 9.604026 7.866901 8.423473 7.253898
#>  [105] 8.460251 7.059159 8.629684 6.464458 8.100811 9.887606 6.693176 8.204641
#>  [113] 5.278268 7.563184 5.899104 7.223987 9.338530 5.393566 5.413110 5.494902
#>  [121] 9.015328 5.151110 8.341308 6.033060 7.057365 9.255138 9.965403 5.925315
#>  [129] 7.604480 9.749530 5.180734 7.145113 7.284033 9.413631 6.624296 9.587363
#>  [137] 8.433173 9.887541 9.774107 6.834353 5.501235 5.597601 6.340389 8.280382
#>  [145] 5.656804 5.369109 8.738329 5.161155 7.219309 7.532435 7.657579 6.791484
#>  [153] 9.901809 5.666818 7.027000 5.024922 5.724368 9.391958 7.602075 8.522599
#>  [161] 8.617527 8.459836 6.800114 6.793278 7.127845 9.949313 7.164171 9.701155
#>  [169] 7.205320 6.166981 9.572376 5.020785 5.249404 7.523247 6.228387 5.051043
#>  [177] 7.644949 7.587113 7.833739 5.759940 6.851402 9.351354 7.589896 8.219223
#>  [185] 9.265919 6.765699 8.538013 7.366900 8.765028 8.478749 6.915845 8.548411
#>  [193] 6.701405 5.514130 5.702924 8.964387 7.216381 9.912283 6.884862 7.249617
#>  [201] 8.265226 5.377906 5.478520 9.877717 6.682756 6.025124 8.452495 5.866127
#>  [209] 7.668779 8.322511 7.783531 9.884927 9.208881 7.318696 8.617573 8.514964
#>  [217] 8.993034 8.590692 6.776587 9.375918 6.321220 8.525067 5.975698 8.859425
#>  [225] 6.382255 7.543462 6.342789 6.040297 8.949665 6.282876 7.040316 6.634482
#>  [233] 6.686306 8.162233 9.942477 8.616562 5.570146 7.418255 7.898000 9.850814
#>  [241] 8.686223 6.245921 8.574643 5.252306 7.341943 9.829450 7.801203 9.417967
#>  [249] 8.720979 8.961221 9.017719 7.764623 9.478443 6.415287 6.374071 5.376500
#>  [257] 7.460255 8.118988 7.303051 7.649817 5.488208 5.320497 6.848254 7.860071
#>  [265] 8.781896 7.244682 8.007450 6.609258 6.979929 6.987274 8.580017 9.263330
#>  [273] 9.033229 8.705135 5.263213 7.825437 7.359841 7.422886 9.317160 6.139003
#>  [281] 9.349070 5.570963 8.934409 8.586387 7.531050 8.976526 7.852249 9.166910
#>  [289] 5.525829 7.050947 6.401312 5.876621 5.989784 7.514628 6.018977 9.638086
#>  [297] 8.576376 9.166869 7.046864 6.783622 9.268671 8.743928 9.801611 6.154982
#>  [305] 8.823820 7.390196 7.344955 6.564433 5.346001 9.445988 9.790771 8.697571
#>  [313] 7.150594 6.000062 6.954850 9.095164 6.337511 6.387621 7.006401 5.808253
#>  [321] 6.488890 5.433445 7.752315 6.693081 6.073812 7.974834 5.092547 6.506525
#>  [329] 7.587652 5.711288 5.781568 6.984949 9.834607 9.827485 7.406334 7.643621
#>  [337] 9.314059 6.254447 9.538693 7.554289 7.947969 8.183834 6.605206 5.928268
#>  [345] 6.770171 7.269335 5.596017 5.058893 8.822472 6.827021 8.157754 9.603347
#>  [353] 6.196439 9.719732 9.770533 8.523312 9.419979 5.634861 8.146135 9.465706
#>  [361] 9.956708 9.612006 9.025035 9.015742 5.584938 6.565914 5.274968 7.789419
#>  [369] 6.870156 9.823252 9.184159 7.208981 9.037702 7.341620 7.903050 6.834495
#>  [377] 5.404796 5.584450 7.923348 8.681158 8.575607 7.218139 8.305040 6.515225
#>  [385] 6.246261 7.193510 7.596829 9.533814 8.084242 6.271794 6.820958 5.956987
#>  [393] 5.603360 9.121146 9.116505 7.951702 6.842983 8.363507 9.370016 9.033430
#>  [401] 5.819911 6.553935 9.639255 6.903590 5.029058 8.602410 5.264779 7.814403
#>  [409] 9.719505 9.088461 7.508852 7.631715 9.915252 5.242152 9.947236 8.835144
#>  [417] 6.846543 7.864316 9.723820 5.093361 6.219999 7.312013 9.676073 9.400656
#>  [425] 5.092312 9.442361 5.444564 9.568390 8.636117 5.785429 7.887011 5.653571
#>  [433] 6.923982 5.967659 8.218089 9.684087 7.245877 7.052027 7.469754 7.193078
#>  [441] 5.147582 9.326344 5.245835 9.463014 5.426379 8.164471 5.941373 5.234556
#>  [449] 7.101232 9.668411 7.256027 5.846441 6.047563 5.449464 7.257447 6.401584
#>  [457] 8.297298 6.061389 8.455570 8.350049 5.644691 8.170353 6.931651 9.576750
#>  [465] 6.446215 8.368789 9.995421 8.268548 5.600176 9.808094 9.460110 9.633528
#>  [473] 9.635463 6.137704 8.300607 7.167277 8.926389 6.721853 7.440322 8.541780
#>  [481] 6.085499 6.136417 5.043851 9.101613 8.087323 8.294674 5.451339 7.241162
#>  [489] 8.952420 8.757678 5.625595 7.118105 9.270056 7.590108 9.856332 5.997965
#>  [497] 9.656583 5.430447 5.264859 8.261475 7.620584 5.306625 6.550677 8.746894
#>  [505] 6.557847 9.615529 8.166741 6.961476 6.236188 6.584552 9.192632 7.428145
#>  [513] 6.110352 8.379081 6.721037 7.838840 6.097913 7.067951 6.183105 6.155930
#>  [521] 7.620378 8.237752 9.753355 9.561275 5.359153 9.778722 9.872592 8.086370
#>  [529] 9.085215 9.722569 9.847739 7.950995 5.985649 9.658507 8.834763 8.347577
#>  [537] 5.466198 9.932222 5.066664 9.713306 8.194887 9.053541 9.591467 5.821729
#>  [545] 8.423545 9.782432 8.496896 7.216137 5.457188 7.584534 6.033950 7.133657
#>  [553] 6.036722 9.765217 9.540258 8.101492 7.440983 9.963802 7.528429 7.985682
#>  [561] 8.744676 9.239902 5.688430 7.560106 7.318679 8.208246 6.792075 7.190868
#>  [569] 7.016293 9.279989 5.508087 9.162597 8.862833 8.725759 7.743209 7.193646
#>  [577] 6.218020 7.756207 7.497673 8.435206 7.091776 6.044504 7.936906 7.074277
#>  [585] 6.641580 8.861918 5.057400 6.736051 5.548530 6.933632 9.804398 8.593254
#>  [593] 6.835606 8.271947 7.472298 6.002838 9.663470 8.665411 6.805566 7.888044
#>  [601] 8.838059 7.474471 5.464783 5.037373 8.095819 6.399664 6.780575 6.997185
#>  [609] 6.735000 6.740567 7.553845 9.507414 6.537978 5.372460 8.041920 7.960162
#>  [617] 8.649239 6.028051 5.357380 6.454861 6.956085 6.959170 6.676716 6.138398
#>  [625] 5.647221 5.244565 9.842742 7.989141 5.192810 6.365458 9.843282 7.770465
#>  [633] 8.471592 5.687748 9.089620 7.646975 8.237838 5.728361 8.030939 9.279366
#>  [641] 9.502491 6.615447 8.360592 8.325983 6.615052 6.352657 7.129517 9.011887
#>  [649] 9.407891 6.718742 5.455598 7.834274 5.389128 5.835141 7.836942 7.564509
#>  [657] 9.113569 8.131442 8.352672 6.537835 8.438440 9.805773 9.452828 9.115736
#>  [665] 7.949948 9.357224 7.429207 8.877810 7.012559 5.961052 8.688388 6.642302
#>  [673] 6.002826 6.567427 5.293962 6.355054 6.249733 8.341558 6.938534 6.717898
#>  [681] 8.705078 7.313861 9.346001 5.456810 8.879050 5.914570 7.283896 9.286210
#>  [689] 9.231072 9.324859 5.121752 6.274977 7.926864 5.191687 7.095749 8.890594
#>  [697] 5.815341 5.403597 7.772174 9.025518 7.597332 5.761540 8.374280 8.606092
#>  [705] 7.293030 7.214172 7.359175 7.281915 8.521614 5.493531 6.576966 7.394064
#>  [713] 6.224242 7.171665 9.281411 5.050845 6.891503 5.586756 7.893093 7.464446
#>  [721] 6.714020 8.335965 5.153144 6.726112 8.214470 9.941868 7.884998 5.836515
#>  [729] 6.424802 6.548927 7.895185 5.926898 5.618900 9.683113 5.071753 8.717228
#>  [737] 9.980626 5.577555 8.466829 9.081399 6.793514 9.762784 9.636095 5.660707
#>  [745] 6.616449 5.345025 5.025926 8.312791 6.460606 8.311227 7.705148 8.911164
#>  [753] 8.620169 9.577722 9.782090 5.167674 5.399030 7.569948 8.420696 8.529347
#>  [761] 9.813290 5.429066 7.485055 5.769527 8.494520 9.174434 8.025698 6.320684
#>  [769] 7.709431 8.850417 6.400215 8.160443 5.289980 8.158487 6.033417 6.311275
#>  [777] 5.638819 7.683726 8.406554 9.021036 8.575520 8.465846 8.906902 8.929714
#>  [785] 7.080605 8.090121 7.405918 5.026912 6.362637 8.740554 7.304284 7.415214
#>  [793] 7.772688 7.887612 9.219507 6.392777 7.388975 7.107265 9.827528 8.774062
#>  [801] 6.805570 9.026824 7.858553 9.588082 5.862364 9.441739 5.331145 8.160653
#>  [809] 7.807219 9.203660 5.604976 6.453577 9.541235 8.495511 6.654743 8.887986
#>  [817] 6.950207 5.697860 8.390233 7.144883 9.982433 5.270795 9.764801 5.781059
#>  [825] 5.019858 9.504217 5.660475 8.037819 7.704208 6.817721 7.453132 5.222099
#>  [833] 8.649664 6.124624 5.750725 7.966061 7.667296 7.529677 9.817252 6.106407
#>  [841] 7.909414 6.455978 5.934097 6.686632 5.650331 8.981820 8.280101 5.283715
#>  [849] 9.360773 6.650636 6.082938 9.947439 9.374542 6.518931 9.291969 6.753935
#>  [857] 5.034198 6.115956 5.376973 6.641067 5.479048 7.805779 8.836664 8.506626
#>  [865] 7.234994 8.411105 5.090112 8.477718 9.329172 5.738278 7.103012 6.066115
#>  [873] 7.048901 8.054086 7.251017 5.089070 5.105100 7.574126 5.454411 8.610072
#>  [881] 8.090554 6.584681 8.268703 8.424556 5.209871 7.261653 9.770188 9.587512
#>  [889] 9.212005 8.196869 8.647177 8.388632 9.749360 6.605931 9.678065 5.918747
#>  [897] 6.986895 9.944815 6.095058 8.328690 5.253214 8.978492 7.187465 7.618536
#>  [905] 7.073924 5.832933 8.537395 7.965518 5.246223 9.936410 5.756578 9.136013
#>  [913] 8.450140 6.163163 9.152918 7.742515 5.268489 5.734318 7.522020 8.863606
#>  [921] 6.834175 8.095916 7.809707 7.503955 7.876894 7.978549 9.466097 9.918763
#>  [929] 6.015618 9.551984 7.731910 7.220787 9.428602 7.732355 7.502170 8.879888
#>  [937] 9.019014 6.719018 5.074676 6.394769 6.027292 7.563834 6.411812 5.288059
#>  [945] 5.297608 5.463527 9.710295 7.574350 9.057001 9.621354 7.200096 6.360536
#>  [953] 5.837687 5.717004 5.804683 7.873649 6.187632 6.261489 9.673658 8.879345
#>  [961] 8.872423 9.155584 8.142025 7.360456 5.292389 7.137230 6.799971 6.693816
#>  [969] 6.620616 8.963510 7.630323 5.773855 5.653337 7.600049 8.034604 5.129031
#>  [977] 8.093990 5.820325 5.069856 8.479809 7.282279 8.012901 8.663757 7.830640
#>  [985] 5.215138 5.909738 5.622372 9.560273 9.672103 7.626729 6.662512 7.700944
#>  [993] 8.987887 5.642593 8.873825 8.231264 7.728330 6.370271 7.280675 7.075558