Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 8.040927 5.679377 9.229242 7.783929 6.894933 9.154014 7.496967 7.748764
#>    [9] 9.890385 5.219427 9.781203 5.795850 5.796914 6.062762 9.396652 5.566992
#>   [17] 5.597811 5.050990 6.091496 5.280839 7.553525 8.771139 7.195371 7.460835
#>   [25] 8.460987 9.423246 9.875617 5.628911 8.806796 5.295426 9.630672 5.631096
#>   [33] 8.522091 7.318446 6.013262 8.982207 5.517036 8.025014 6.947060 7.023595
#>   [41] 6.073801 6.642832 8.862378 8.514892 5.809226 7.862980 5.893611 9.562527
#>   [49] 6.097116 8.496048 5.189132 6.105863 9.757689 6.632212 5.772055 5.341216
#>   [57] 6.672665 8.818262 6.677665 6.718283 5.947877 9.710936 7.970701 5.691643
#>   [65] 9.419444 7.394414 9.471653 7.995831 8.052443 8.001758 6.137689 9.338252
#>   [73] 8.244219 8.334146 9.770415 5.187126 8.192790 9.770695 7.350436 5.905718
#>   [81] 7.816200 8.035300 5.751873 6.101630 6.314002 9.275612 6.810847 6.615668
#>   [89] 7.506906 5.545745 6.660107 8.813307 9.847374 9.969549 9.165697 8.794910
#>   [97] 6.038818 9.040860 5.104368 8.426559 7.034913 7.552245 8.169407 9.252100
#>  [105] 7.128945 9.848245 6.429030 8.212576 6.515534 6.058882 5.370125 7.631313
#>  [113] 5.241337 6.001784 7.685481 7.528080 7.695435 6.781919 8.723954 7.040807
#>  [121] 7.207325 6.144653 5.763165 7.704349 9.195243 5.456628 7.628040 9.343998
#>  [129] 9.498855 8.349216 7.561113 6.114051 5.036647 7.844218 6.436738 5.962487
#>  [137] 9.139222 8.319960 8.139287 6.073009 9.881816 9.296383 9.746689 6.937095
#>  [145] 6.344911 9.655390 7.849982 5.828134 5.938430 8.488881 8.571336 8.722865
#>  [153] 7.579874 7.737596 5.847318 5.276546 6.673266 8.313833 8.322514 5.595466
#>  [161] 9.442369 9.674199 7.980243 8.662851 5.103396 5.814308 9.421850 8.377567
#>  [169] 5.952071 8.632344 7.693008 5.877922 9.819579 8.135251 5.076753 8.375097
#>  [177] 9.449886 6.545824 8.116493 6.466617 9.082793 9.814922 6.746524 7.436151
#>  [185] 5.845363 8.554380 9.660341 5.317510 5.342811 9.412430 6.882882 6.311736
#>  [193] 9.255315 6.659913 9.193071 8.638078 5.577967 5.726012 9.847596 7.231132
#>  [201] 5.094217 8.881836 6.970552 7.358676 9.199724 5.897204 6.181221 9.454632
#>  [209] 5.776217 7.257843 5.359202 8.376915 8.731331 6.701589 8.690302 8.708520
#>  [217] 9.132675 6.916609 7.842540 5.138902 7.724008 7.913340 9.626089 9.293592
#>  [225] 7.669562 8.236964 8.255955 9.885000 8.282143 8.625368 5.872635 6.910334
#>  [233] 9.757418 8.052354 7.436708 5.391579 9.744543 7.704586 6.828273 7.995699
#>  [241] 9.057647 5.935303 9.417665 5.330489 6.065019 8.131792 6.575454 5.894359
#>  [249] 9.610826 5.339548 6.280576 6.506582 5.108784 7.013076 6.000923 9.376913
#>  [257] 5.007640 5.838518 9.944553 5.974180 8.645315 5.932418 6.635344 7.670799
#>  [265] 8.392878 7.409184 9.048964 8.889724 9.057981 8.720794 7.516363 7.982990
#>  [273] 8.965998 6.456538 8.155392 8.661122 5.786629 8.812352 8.710460 8.351011
#>  [281] 6.938834 9.002882 7.584009 9.145735 7.529605 9.270590 7.042227 9.783067
#>  [289] 7.632410 8.634842 9.680430 7.867894 5.784437 9.412833 5.735010 8.299012
#>  [297] 7.879930 5.649100 5.090839 8.345543 6.518140 6.398662 7.085877 6.939498
#>  [305] 7.923707 7.906902 5.411838 9.793366 5.646781 9.131567 6.072626 5.046093
#>  [313] 6.397103 6.997559 9.771475 7.808452 9.226739 7.958443 7.580646 7.233498
#>  [321] 9.354988 5.805744 7.492651 7.066050 5.661151 6.193992 5.375285 7.468586
#>  [329] 9.960276 9.096975 8.786047 8.585303 6.367623 5.055391 6.496615 8.868199
#>  [337] 6.106718 6.299051 7.247535 9.575805 8.037622 5.231482 7.016794 8.145873
#>  [345] 9.917654 8.455261 7.686383 8.735832 6.098361 6.700216 7.999894 5.817877
#>  [353] 6.286671 6.207083 6.175647 8.860175 8.048767 6.833938 7.089053 6.770390
#>  [361] 8.636532 9.853107 7.204682 8.570895 9.322575 7.465787 7.225085 6.247201
#>  [369] 8.270964 8.909739 8.474311 9.236316 5.022644 8.129157 6.557243 5.118570
#>  [377] 8.715067 9.764556 7.447046 5.561574 6.016789 8.953036 7.118630 5.972597
#>  [385] 5.403171 7.050589 8.782225 5.578288 5.454881 9.808070 6.838104 8.628705
#>  [393] 7.283752 9.869023 6.099771 7.983755 7.057095 5.160208 7.171199 9.329778
#>  [401] 9.458419 7.974105 8.676513 6.488942 9.251932 6.955778 8.677527 6.293298
#>  [409] 8.709643 8.187471 7.222486 8.621218 7.943527 9.942639 8.977092 9.219968
#>  [417] 6.149790 6.310533 7.522693 7.077597 6.332561 9.145075 7.226906 5.997516
#>  [425] 6.377177 9.350852 5.227950 9.300467 7.462892 5.708996 5.320343 7.425606
#>  [433] 9.759283 6.882016 6.742730 8.720260 6.500520 5.130403 6.632831 5.021777
#>  [441] 7.641252 8.155362 7.488564 9.947261 5.914644 5.621768 7.631403 6.422037
#>  [449] 7.862094 5.826635 5.939391 7.186103 7.043652 5.697795 7.917924 9.398978
#>  [457] 8.213778 9.273100 5.156485 9.459765 6.077198 5.437753 5.731371 6.927784
#>  [465] 7.586594 8.278581 5.012104 5.021199 6.270623 5.112273 9.221026 8.409809
#>  [473] 5.092280 9.126503 8.302539 7.340417 5.316864 6.920133 8.441172 8.101179
#>  [481] 8.367053 8.512534 9.307248 7.456263 7.063133 7.647297 5.167316 5.075200
#>  [489] 9.981644 5.691628 7.154511 6.771667 8.415885 7.843865 5.208139 6.729810
#>  [497] 9.493719 9.453347 6.410793 6.074646 9.723199 9.783299 6.753867 7.353133
#>  [505] 8.996573 9.498677 8.572678 8.087181 9.803059 5.335249 9.972134 6.778334
#>  [513] 7.925546 9.201864 6.458641 8.418103 7.861934 5.140650 5.580614 6.297315
#>  [521] 7.141009 9.299557 8.856903 9.954008 9.915701 9.440030 5.937643 8.929818
#>  [529] 8.598346 5.031952 8.608152 8.228422 5.728638 6.856533 7.160871 8.118071
#>  [537] 5.944716 9.521375 9.646827 9.292009 8.360024 5.736339 7.184378 6.331856
#>  [545] 7.597711 8.786363 5.742300 6.076140 9.005275 9.703113 7.540345 8.459088
#>  [553] 5.840905 6.272068 9.366305 8.699181 5.797003 7.750620 8.210932 8.053161
#>  [561] 7.691379 9.733825 9.887260 8.244601 8.175375 7.289623 9.714496 6.073060
#>  [569] 8.085659 6.332014 8.331721 6.300098 9.749102 9.056889 7.152300 5.139162
#>  [577] 6.030935 6.530432 9.629745 6.363439 7.032592 8.362441 5.690126 5.862247
#>  [585] 8.050254 6.835840 6.704918 7.126326 7.894554 7.509887 6.838594 7.806952
#>  [593] 9.823270 8.442940 8.335732 7.431761 6.252283 5.420646 7.361846 8.873688
#>  [601] 9.526969 5.073064 8.817264 8.742524 5.233172 7.243089 6.746066 9.615374
#>  [609] 6.617531 8.576860 8.143186 7.528089 9.481749 9.915035 9.896494 9.279672
#>  [617] 5.983024 8.457557 7.777746 7.335248 5.327504 6.950422 8.905361 5.296550
#>  [625] 7.345915 5.103610 5.172634 9.293272 8.754942 9.476165 7.565546 8.364244
#>  [633] 8.151112 6.721975 7.191695 9.091583 9.885408 5.915285 5.950252 8.053207
#>  [641] 6.620724 9.752899 5.650515 7.075406 9.239537 5.795115 6.418169 6.595889
#>  [649] 6.223813 7.923130 8.030089 8.311747 7.588260 8.493427 5.515270 7.719567
#>  [657] 5.619878 6.356103 9.413342 5.098942 5.984805 7.096315 5.071109 7.131217
#>  [665] 7.535655 7.418903 6.340014 7.175369 8.482733 6.739413 5.143210 9.214587
#>  [673] 7.147224 7.977886 8.730547 5.696462 6.819927 8.679991 8.122343 6.867525
#>  [681] 7.283756 8.070011 9.100046 5.936909 7.351780 6.756745 9.358311 7.122902
#>  [689] 9.058949 9.902963 6.706583 6.202060 5.767226 5.490985 6.607453 5.315937
#>  [697] 7.771652 6.321193 8.163605 9.772181 6.001701 9.096838 5.782469 7.187654
#>  [705] 5.737107 7.354281 5.598779 7.096716 5.211425 6.023709 8.882837 6.767323
#>  [713] 9.108683 5.806632 8.791542 8.745777 7.278486 9.992153 7.702710 5.522578
#>  [721] 8.189498 6.326758 7.249543 5.104338 6.445906 7.751370 6.780907 7.321359
#>  [729] 7.435748 7.147733 8.006868 7.104867 6.393081 8.150648 6.448613 8.527261
#>  [737] 9.089551 8.638183 7.748312 8.370317 7.456654 9.125414 5.634559 5.122532
#>  [745] 6.148145 6.067065 9.897612 8.884221 5.361535 9.509820 6.018907 5.956888
#>  [753] 7.264898 9.078914 8.574360 6.715197 8.062356 5.817921 6.547677 8.701687
#>  [761] 5.838112 8.066420 6.989551 6.780283 9.453782 8.122588 7.204271 5.745320
#>  [769] 7.120266 6.479966 5.603846 5.213863 7.805805 6.796629 5.322000 9.045957
#>  [777] 8.879154 9.403408 6.382106 9.852568 7.755500 5.486197 6.939228 7.313789
#>  [785] 7.673338 8.672869 9.106844 6.924990 9.721071 5.086640 6.622903 8.469073
#>  [793] 8.928885 8.007987 8.871880 8.264975 8.958659 8.456109 6.152767 8.048512
#>  [801] 9.769750 7.779121 8.657516 8.091606 6.104990 9.712742 7.162049 7.243968
#>  [809] 6.214241 5.356059 7.270561 5.743010 8.957281 6.961197 8.096096 5.112586
#>  [817] 5.837458 8.741460 8.050833 7.767454 6.842646 9.951991 5.813867 5.760471
#>  [825] 8.521837 8.447241 6.511598 5.938989 8.338456 8.613816 7.394423 9.550488
#>  [833] 5.188309 8.072873 5.273629 5.930344 9.915422 7.385017 5.574489 5.151616
#>  [841] 5.353378 9.113640 5.971599 8.150341 8.716097 5.967137 6.495461 6.603834
#>  [849] 5.997759 7.030411 7.308380 6.510751 6.590358 5.912320 7.443678 8.259555
#>  [857] 5.156975 6.137807 7.906784 5.215602 5.240808 9.858287 7.615797 8.777157
#>  [865] 8.124108 7.539381 8.059463 6.485179 6.489925 5.491573 5.983283 5.897555
#>  [873] 9.423719 8.146177 5.769218 5.229264 5.937500 7.835969 6.203819 7.501847
#>  [881] 6.298171 5.608051 9.451541 9.436388 6.122406 6.294241 5.616960 7.728405
#>  [889] 7.922516 7.462147 5.339256 9.249669 7.608136 8.353675 7.372056 7.999670
#>  [897] 9.596637 8.281882 5.539295 8.690571 8.569278 7.173454 9.874452 6.582022
#>  [905] 9.528913 5.495922 9.833235 8.106282 6.451602 7.179047 9.623833 9.148833
#>  [913] 5.756935 8.030219 8.491097 8.034166 7.104771 5.744787 7.362016 9.784265
#>  [921] 7.359424 5.395254 7.016716 8.346333 7.978849 6.860851 8.967941 9.383874
#>  [929] 8.706525 9.797376 9.140640 7.582674 5.936977 6.620424 6.616054 6.487143
#>  [937] 6.695085 5.696598 9.360970 8.792100 8.781861 8.716556 6.153058 7.786249
#>  [945] 7.721890 6.297825 5.380202 7.790451 5.607157 5.346004 7.834118 9.174222
#>  [953] 6.948082 5.178185 9.295278 7.166089 9.063972 8.278370 8.149691 5.845443
#>  [961] 5.999057 5.308690 5.108870 6.115927 8.257624 6.628533 5.350346 9.624493
#>  [969] 9.998769 8.397875 5.772189 6.029364 6.706512 7.972782 7.309040 8.607030
#>  [977] 7.040811 7.239337 6.222778 9.778257 9.889407 9.141279 6.153689 9.587932
#>  [985] 6.910811 6.422654 6.259692 8.228600 6.976835 5.677734 6.982097 6.269611
#>  [993] 9.557337 9.430475 5.990767 6.456412 6.041454 7.187567 5.232928 7.996047