Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 7.094004 5.914165 8.073710 6.629363 8.185619 6.722836 8.915157 7.247248
#>    [9] 8.361965 8.090563 7.749767 7.058595 5.279866 6.341398 6.123850 7.497176
#>   [17] 5.398893 9.089157 6.546342 9.834906 9.449923 8.969233 7.456081 7.676531
#>   [25] 8.770104 9.849145 5.909522 5.963530 7.289054 6.005381 8.053256 9.404646
#>   [33] 7.830450 7.441883 9.662909 5.001199 6.974855 7.871669 6.072949 7.615970
#>   [41] 8.361017 8.390982 9.852734 8.903255 8.638540 6.083721 6.319197 9.266548
#>   [49] 5.229374 8.642191 5.802452 7.691598 8.712185 7.376510 9.852988 6.645685
#>   [57] 7.755675 6.603926 7.808753 6.936383 5.970108 8.118296 8.705534 8.133366
#>   [65] 6.100350 8.494603 5.745028 6.387042 8.461393 6.453597 5.805416 7.247801
#>   [73] 6.766539 7.608534 6.948922 6.560168 8.786994 6.875032 6.122337 8.789764
#>   [81] 6.760486 5.474813 6.068912 5.787353 5.291257 5.767843 5.515630 5.066643
#>   [89] 6.252368 6.922356 9.700941 7.298996 5.438984 7.442898 5.691467 5.729762
#>   [97] 8.070349 8.592357 6.812984 5.392679 6.407492 9.676063 6.873150 5.089079
#>  [105] 8.810498 6.373660 7.156866 5.882828 6.417913 8.329264 5.464614 5.745469
#>  [113] 6.961396 9.047643 6.202079 5.946243 9.636264 5.422398 7.110703 9.602095
#>  [121] 7.818084 6.824479 8.960323 8.639727 6.784310 5.065419 9.468673 5.444120
#>  [129] 7.221414 7.324177 7.941628 6.433200 5.788093 8.145744 5.095960 6.063483
#>  [137] 9.463835 7.147221 8.706042 9.420068 5.838620 5.302563 7.438843 7.362533
#>  [145] 7.270273 7.198642 9.031736 6.337990 6.980707 9.959457 8.972896 9.038302
#>  [153] 5.570355 5.743607 9.303451 5.015587 7.792552 9.538704 7.143224 9.804372
#>  [161] 5.256596 9.791002 6.607045 7.409409 7.519577 8.084576 7.235518 7.094221
#>  [169] 8.699913 7.488132 8.596863 8.492853 7.764927 7.278922 5.334735 7.477333
#>  [177] 6.363639 6.678264 6.687521 6.853229 9.731375 7.012367 8.061224 9.262465
#>  [185] 8.538057 6.287328 6.909397 6.778546 8.191963 7.764442 7.896466 9.205312
#>  [193] 7.825250 9.121235 6.259983 6.120114 9.838042 6.168039 9.868449 5.359136
#>  [201] 8.255634 8.000496 6.538376 7.868867 9.180544 6.901043 9.943460 9.789492
#>  [209] 5.679395 9.282026 8.840062 5.194140 7.125697 9.218191 9.221576 9.202018
#>  [217] 9.882646 5.175758 9.858909 5.277021 5.992653 8.937195 8.045317 7.498521
#>  [225] 6.755363 8.863853 8.615037 7.838550 7.247923 6.509421 6.207933 7.326080
#>  [233] 6.944084 7.711813 6.677650 9.327171 7.159985 9.842000 9.471991 8.575835
#>  [241] 5.958133 9.397303 8.316527 5.779833 6.356209 5.740508 7.757863 9.909232
#>  [249] 6.280350 5.095666 5.816494 9.633240 5.485847 7.511604 9.024480 8.039314
#>  [257] 8.363248 5.440492 8.352470 8.133568 6.100904 9.956179 7.549097 9.481562
#>  [265] 9.199638 9.824261 7.172675 5.937486 5.925773 8.282549 9.642482 8.646959
#>  [273] 7.054382 9.718760 9.180335 7.035565 8.623350 9.498766 8.788129 6.063017
#>  [281] 5.748947 6.429146 9.122885 8.702870 8.499555 8.924657 5.099654 5.577779
#>  [289] 5.586177 5.748020 6.918966 7.911965 8.581093 6.643954 8.512185 8.306505
#>  [297] 7.127755 9.940435 9.267888 8.783291 5.420905 8.043836 5.581737 5.807902
#>  [305] 9.050075 6.666805 9.898001 8.747478 5.362565 6.626813 5.685782 7.731669
#>  [313] 6.587894 6.507938 8.709246 7.430520 9.759513 8.548887 7.724822 8.455944
#>  [321] 6.311610 9.134412 8.867694 6.709686 8.795887 8.938219 6.021944 5.761631
#>  [329] 5.306697 8.713009 9.463200 9.742244 8.925241 5.692502 7.116831 9.951903
#>  [337] 5.334708 8.256750 7.093323 8.471949 6.892345 5.943566 9.320253 7.856803
#>  [345] 8.961458 9.730409 9.603112 8.577885 6.722633 6.353343 5.847082 9.477032
#>  [353] 9.263807 8.183294 8.240231 9.691378 7.109241 8.541480 8.076750 8.896361
#>  [361] 8.842021 8.892914 6.057081 7.458677 8.347530 9.536531 7.768256 8.776490
#>  [369] 7.582284 6.411471 7.037991 5.979180 7.989086 9.110175 8.418485 9.382538
#>  [377] 8.229087 6.381508 9.640403 7.499778 8.300670 9.919604 8.151336 6.702908
#>  [385] 7.853810 9.274712 9.355085 6.903138 8.944451 6.053169 8.336501 6.741260
#>  [393] 8.694849 7.580552 5.399396 5.611093 9.500387 5.359613 8.300283 5.276492
#>  [401] 8.318412 7.899790 5.947331 9.320456 8.355357 5.037726 7.243043 7.716887
#>  [409] 7.203288 9.343029 8.151074 9.069279 7.330527 9.202404 9.569021 5.768195
#>  [417] 9.250977 5.145411 6.796145 5.837884 6.376578 5.471041 7.187722 7.583313
#>  [425] 9.912303 8.921120 5.612993 6.628716 7.862677 7.592097 5.948134 5.572721
#>  [433] 7.095767 6.630224 6.528710 5.153808 9.421148 8.698389 6.568145 7.526511
#>  [441] 5.974772 6.929201 9.842163 7.534752 7.033828 9.123738 5.927667 6.464033
#>  [449] 6.166894 7.132596 6.805274 8.215722 9.672391 7.323344 5.688464 9.774303
#>  [457] 5.466482 6.486701 6.394259 9.800690 5.191709 5.085361 8.405763 6.718852
#>  [465] 8.919349 6.325385 6.622022 9.585550 7.768112 8.914031 5.714075 7.724584
#>  [473] 7.967464 6.110148 7.939560 6.371838 8.978546 8.014513 8.198762 8.574680
#>  [481] 8.453945 6.409191 5.033909 6.833981 7.935657 9.630959 6.231812 9.675761
#>  [489] 5.764354 7.133869 7.293565 8.917568 5.800908 6.301768 7.606834 5.249812
#>  [497] 8.285977 6.871155 6.783079 8.346679 7.700575 6.271850 6.413123 5.463399
#>  [505] 5.350214 6.852183 5.356520 5.933191 7.739568 5.853937 8.516321 6.685998
#>  [513] 5.860352 7.133390 7.083792 7.359287 6.848885 8.702155 7.141979 9.956866
#>  [521] 6.886883 9.393397 7.282246 8.414929 6.258926 7.897251 9.094786 8.411148
#>  [529] 8.549964 8.264454 5.592289 5.856061 8.152000 6.126378 8.809685 7.297501
#>  [537] 9.218201 9.989474 6.501956 6.066132 7.659766 7.427687 6.051682 6.090121
#>  [545] 5.380368 8.600281 6.487288 7.312774 6.765025 6.636562 5.381846 8.788691
#>  [553] 7.134480 8.481204 5.898353 7.800880 6.798211 9.603515 5.639068 7.231198
#>  [561] 8.216907 6.460873 7.403864 5.142621 6.579985 7.236627 9.465121 7.565212
#>  [569] 5.529544 5.247470 8.954272 8.192161 8.384365 5.688269 7.852964 7.856221
#>  [577] 6.350139 6.583332 5.412869 9.080409 9.417850 6.332623 9.884173 6.379659
#>  [585] 8.430672 6.778133 9.829515 8.900627 9.988350 7.377833 8.506969 5.927875
#>  [593] 5.192946 9.440525 9.533237 5.546408 6.062213 7.077436 7.404606 8.760929
#>  [601] 5.718798 5.364197 8.468650 5.812855 9.825024 6.852015 7.671090 5.897048
#>  [609] 5.938241 6.774799 8.967155 9.280144 9.066672 7.537183 7.965786 5.348224
#>  [617] 8.522664 6.997667 5.485981 6.287765 7.754891 5.014480 7.797286 8.039163
#>  [625] 9.359616 6.915918 5.381861 9.343360 8.511766 9.984109 5.972734 5.946498
#>  [633] 9.750408 5.457326 5.969908 6.288479 8.137071 9.691222 8.529379 7.719321
#>  [641] 5.384618 7.841891 6.170566 6.850882 8.311886 6.294061 9.552740 6.161980
#>  [649] 6.355026 7.186451 9.434690 9.321912 8.906608 6.192954 8.354206 7.254536
#>  [657] 7.691122 6.536953 7.609633 6.576725 9.085295 6.654736 5.730710 6.548630
#>  [665] 6.793671 5.435022 5.926743 8.969713 5.454555 8.047974 5.645768 5.239097
#>  [673] 7.443072 5.418364 5.014256 5.374625 8.077535 5.982779 5.722302 8.733809
#>  [681] 5.762502 9.372303 9.088736 9.820198 5.961744 8.216215 8.908426 6.985079
#>  [689] 9.731089 9.947215 9.240135 5.928030 9.901905 5.952483 6.049662 7.215668
#>  [697] 9.994944 9.687478 5.351020 9.842540 8.124752 5.342853 7.696930 8.367686
#>  [705] 6.790188 7.477915 8.553316 7.394997 5.096562 6.900086 7.944923 6.974504
#>  [713] 6.739584 7.277843 7.678264 8.261527 5.450674 6.457217 9.001192 9.161264
#>  [721] 5.772440 9.711967 6.483581 5.439237 6.528641 5.292737 5.483694 9.382926
#>  [729] 8.201532 9.259949 9.996211 6.214354 7.288674 9.479393 8.839665 8.466621
#>  [737] 8.423251 9.337795 7.771384 9.920223 8.216687 5.482698 9.922698 8.132491
#>  [745] 6.355558 9.574942 5.116448 7.217074 7.494427 6.711204 5.301219 5.426395
#>  [753] 7.206785 5.847650 6.648130 6.195844 9.179982 7.666690 7.954146 5.924891
#>  [761] 5.293150 9.576917 7.660762 8.687695 6.635353 6.012350 7.599203 7.659840
#>  [769] 7.616971 6.449419 6.619840 6.868842 7.571971 5.563443 5.529955 5.407353
#>  [777] 8.417958 7.333421 5.240703 7.278600 9.878297 5.621679 5.588821 7.284315
#>  [785] 6.653045 6.334088 8.173076 5.873688 6.662930 9.655183 9.090238 6.584807
#>  [793] 8.270243 7.625900 5.548059 5.928039 6.421265 6.502604 5.503712 6.255821
#>  [801] 7.352866 5.987468 6.021136 9.605758 5.404294 8.023481 7.333585 6.568882
#>  [809] 9.591405 5.228801 5.712331 5.510328 7.758259 7.188786 9.178245 7.609938
#>  [817] 8.898628 7.728224 5.901530 8.325492 7.051735 7.997611 8.551846 7.651418
#>  [825] 8.115316 8.306393 9.615134 5.871839 9.876927 7.550721 8.937962 9.542814
#>  [833] 5.482010 8.658788 8.998521 7.517591 7.916554 7.630857 5.551016 8.487562
#>  [841] 5.356287 9.094200 9.106714 6.452995 5.377793 7.123457 7.906212 8.579559
#>  [849] 6.105739 5.767221 7.005805 5.617520 8.532773 8.945106 7.587236 6.227045
#>  [857] 5.034653 5.082274 8.850529 5.650657 5.047567 8.772019 7.798539 5.091797
#>  [865] 6.860399 9.033675 5.499377 5.386211 9.566777 5.175120 7.493142 8.539795
#>  [873] 5.549721 8.744558 7.045086 8.138429 7.604563 6.383872 9.095455 7.351363
#>  [881] 8.168902 9.561800 6.946355 7.148647 8.831258 7.315508 5.820781 7.218212
#>  [889] 5.443204 6.589782 5.343488 8.501420 6.852972 7.847720 7.317125 9.320487
#>  [897] 8.751600 8.835618 8.356043 9.068598 6.313739 9.224980 8.634225 6.312217
#>  [905] 9.419244 9.265069 7.194509 6.208152 9.523374 7.815439 9.106438 6.248932
#>  [913] 5.696148 7.035852 6.819173 9.744986 7.827904 7.976003 9.083538 5.357574
#>  [921] 9.131442 5.755369 9.715266 9.946892 5.462912 5.270873 9.045835 8.192712
#>  [929] 7.033778 6.756831 8.091977 8.818900 8.950548 8.543383 9.189403 8.912952
#>  [937] 9.498690 5.366976 9.518271 6.961258 8.700639 7.066796 8.740367 8.270029
#>  [945] 7.139019 8.801169 8.348367 8.067192 7.824688 9.261867 9.762288 9.727292
#>  [953] 7.451004 6.765259 8.739604 7.700467 8.887143 7.948151 5.687943 7.361444
#>  [961] 7.907051 9.249230 8.573798 8.281163 7.665159 5.292577 8.179158 9.131071
#>  [969] 7.610457 5.038433 9.649793 6.455713 5.498358 8.716347 9.271518 8.807681
#>  [977] 6.948124 7.138851 6.949559 5.040411 5.119556 6.103846 8.636030 8.983107
#>  [985] 5.787419 6.384700 8.848581 6.940336 9.286574 5.830668 6.141513 9.128576
#>  [993] 7.947483 8.650505 6.906421 5.328262 6.080483 7.864906 7.796531 8.499603