Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 5.216892 7.420817 5.677995 5.341730 8.089506 6.581815 8.254556 8.189183
#>    [9] 5.876881 9.595561 7.531984 5.358841 6.196730 8.930154 8.696368 5.108712
#>   [17] 7.360989 5.012155 8.594783 9.315195 5.654150 7.070792 6.180219 8.990987
#>   [25] 9.601381 6.904951 9.293764 6.254592 8.785769 5.564708 8.282059 5.042966
#>   [33] 8.126024 6.183811 8.117104 8.137324 7.487876 5.495407 9.253438 6.749114
#>   [41] 5.134770 8.516289 8.425864 9.925366 8.820972 7.828552 5.157199 6.209575
#>   [49] 9.179588 7.203454 9.423443 6.703440 7.390840 8.138439 5.092618 6.878164
#>   [57] 9.678669 9.156586 6.638538 5.602032 5.615083 9.033704 9.531310 9.165726
#>   [65] 5.625470 5.121859 8.932002 5.261424 9.987123 6.288976 7.205468 9.142121
#>   [73] 9.890566 9.793527 5.080103 8.414171 8.786414 6.656269 5.961745 6.413013
#>   [81] 6.297782 6.055550 5.997037 8.385717 7.204733 8.499351 7.363053 9.433621
#>   [89] 8.114307 5.457319 5.220177 7.031530 9.714284 6.730306 9.763593 8.540574
#>   [97] 8.349864 9.203677 5.411734 7.231104 8.812992 8.218579 8.084614 9.566432
#>  [105] 6.991233 5.902082 6.577252 7.952480 6.245982 7.016399 5.652164 6.671089
#>  [113] 5.464564 9.939786 7.144953 8.545391 6.332484 8.379646 7.278288 6.491050
#>  [121] 7.931596 5.551798 8.294970 8.496471 6.335690 8.037303 5.845321 5.851339
#>  [129] 8.637319 5.305707 9.411435 5.248637 7.469147 9.877887 8.842475 7.925543
#>  [137] 5.009005 6.679810 9.691252 5.464102 6.318708 7.412615 7.021096 9.739603
#>  [145] 5.023283 6.862247 6.050583 8.912442 6.525528 9.317472 5.506396 6.004534
#>  [153] 7.586854 5.836010 8.311382 8.420445 9.378068 8.764085 6.237922 5.955595
#>  [161] 9.627877 8.367485 5.871139 6.078277 5.215138 8.280801 6.637509 9.669582
#>  [169] 5.918064 5.408995 7.743525 7.886984 6.552262 7.019612 9.494590 9.063821
#>  [177] 8.626560 8.739146 6.576847 7.150544 5.097274 5.985039 9.380564 7.956270
#>  [185] 5.342045 5.228163 8.031718 5.391962 6.422075 8.083029 9.478148 6.397611
#>  [193] 8.404128 5.734387 5.455520 6.146328 9.212229 5.132167 6.779810 5.426021
#>  [201] 7.714425 5.619846 7.200865 5.579920 8.351231 9.659322 7.598785 9.538566
#>  [209] 8.677666 6.987929 9.702062 7.654550 6.796388 9.060455 5.149149 6.508151
#>  [217] 5.785359 8.986041 8.910362 5.209179 7.762268 6.249171 8.365947 6.818019
#>  [225] 6.820365 9.360321 8.308627 9.190981 7.996435 8.857815 6.665103 6.770421
#>  [233] 7.332008 9.407325 6.409585 5.362409 6.350998 7.452450 6.781063 6.348122
#>  [241] 5.181345 8.209776 5.200635 6.419276 6.736907 7.548798 8.902153 6.120621
#>  [249] 9.107519 8.377488 5.513785 6.418411 5.514314 8.100639 8.561528 8.626287
#>  [257] 8.148191 9.743080 5.935017 7.856845 6.872599 7.847134 7.472924 9.405311
#>  [265] 6.120765 9.558921 5.888998 8.335422 9.338388 5.861599 8.797674 9.017336
#>  [273] 8.018725 9.497190 5.501336 5.351653 5.301792 7.838548 5.253210 5.603860
#>  [281] 5.235001 9.394151 9.890833 8.854830 9.540288 6.086063 6.743492 9.552597
#>  [289] 9.658327 9.391674 7.825695 6.965981 5.341209 5.303458 6.437017 7.793245
#>  [297] 8.718064 9.554718 8.638949 6.681973 9.946962 8.368441 6.504055 7.161212
#>  [305] 9.006126 6.700172 9.217042 7.315939 9.870211 6.035180 7.676340 8.488028
#>  [313] 9.112109 7.706109 8.588559 6.271075 8.908810 6.782178 9.560285 8.648888
#>  [321] 5.910246 9.439728 5.903204 7.508790 5.562743 8.392180 7.820309 7.906572
#>  [329] 9.386690 9.127611 9.502510 5.991968 9.894809 5.886281 6.234428 5.480358
#>  [337] 5.461807 5.656587 7.992553 9.976850 5.200831 8.162142 8.287225 9.205486
#>  [345] 7.216813 7.648547 9.665545 8.753014 7.060251 7.459748 5.370612 6.694579
#>  [353] 5.603142 9.222040 9.442903 8.461479 9.367465 9.166693 6.773239 7.460978
#>  [361] 8.514016 9.706517 7.274345 8.351697 8.047581 7.550626 8.492317 8.093116
#>  [369] 8.625190 5.925240 5.594497 8.929572 8.118852 5.106603 5.742165 9.472583
#>  [377] 9.541165 5.694810 7.920806 9.050971 5.363996 6.722801 9.870675 5.162324
#>  [385] 7.880970 6.822898 8.143090 5.451823 7.113841 7.722731 7.453656 8.444337
#>  [393] 9.595760 6.627445 8.454208 7.937304 6.417865 9.824194 7.145058 6.294357
#>  [401] 7.113648 9.188445 7.158161 5.798927 7.136649 5.347852 7.136057 9.264566
#>  [409] 8.238080 5.683203 8.516406 9.781748 7.010153 6.181838 9.646468 9.657044
#>  [417] 5.354588 8.340962 7.981341 9.339182 5.601624 5.744950 5.808676 5.282112
#>  [425] 9.266477 9.205063 7.308759 7.628500 8.761955 7.148051 7.624830 5.340633
#>  [433] 5.839698 7.933154 7.472378 9.184390 5.795788 8.456928 6.480820 9.067080
#>  [441] 5.604067 7.241641 5.619431 8.912034 8.876944 8.399240 6.715484 9.286803
#>  [449] 7.569133 8.104214 5.722162 6.695645 7.754222 6.374183 8.654006 7.994232
#>  [457] 8.856465 7.077800 7.992148 8.545766 5.304124 5.578527 7.707077 8.043952
#>  [465] 7.877613 9.242579 7.712471 5.861232 5.833829 6.963769 5.260473 7.089037
#>  [473] 8.183803 5.279864 5.919908 8.010662 6.863003 5.960841 7.471240 9.722238
#>  [481] 8.441922 7.639850 6.999557 5.489282 8.350799 6.356967 5.327759 7.262751
#>  [489] 6.453513 6.085913 9.457204 8.466175 6.832169 6.796530 8.535090 8.570179
#>  [497] 9.529570 5.944592 7.159981 5.659935 9.727212 7.629746 7.487862 9.845472
#>  [505] 5.855027 5.680858 8.962790 7.285482 6.399538 7.997499 5.605365 8.285880
#>  [513] 9.443798 6.702812 5.026882 7.145437 6.890468 8.021146 6.715216 7.683898
#>  [521] 6.971551 8.274511 9.199156 5.408181 8.027431 8.583387 6.013056 5.784781
#>  [529] 6.760104 5.381485 9.981789 5.085518 6.557182 5.170554 7.942622 8.608475
#>  [537] 7.005738 9.426854 5.557583 5.768039 5.885005 7.027300 5.364422 9.810766
#>  [545] 8.580163 8.996023 7.790415 8.024789 6.840155 9.479827 6.012651 5.018047
#>  [553] 6.005824 8.605987 5.286659 8.615334 9.702008 6.676412 8.921990 8.519801
#>  [561] 6.733632 8.696121 5.203890 8.934353 5.449910 8.281705 6.498477 9.977241
#>  [569] 5.655727 9.208276 8.570377 8.950443 7.754706 7.827880 6.246889 9.832385
#>  [577] 5.592033 9.364952 8.274275 5.166234 6.526212 5.276767 9.404067 6.927747
#>  [585] 6.999197 8.820497 5.237126 8.764837 8.943284 8.709073 9.355847 8.082763
#>  [593] 5.486874 7.993047 6.156999 5.915637 7.081047 9.096970 7.720203 7.905515
#>  [601] 5.038771 7.714758 8.433444 9.192022 6.252650 5.246471 8.388928 9.046502
#>  [609] 6.283068 5.292076 9.800365 7.353876 9.117292 6.430397 9.308442 9.106127
#>  [617] 9.007688 8.500934 5.250520 7.033379 6.227472 5.056281 5.495307 6.043999
#>  [625] 7.800643 9.637819 6.470918 6.153459 6.186034 5.628571 6.322446 6.090147
#>  [633] 5.362227 7.326573 9.971422 5.454103 8.431946 5.865972 9.155644 7.975755
#>  [641] 5.294506 8.992799 9.534567 8.225552 6.269064 8.836951 8.157636 9.425590
#>  [649] 6.699477 9.519876 6.066517 6.579569 5.871802 8.193350 5.438067 6.519141
#>  [657] 9.461319 7.924528 6.943261 9.533221 9.671345 7.708382 9.262497 9.386621
#>  [665] 5.106833 8.023978 8.379066 5.389247 6.801916 9.735833 6.511792 8.061458
#>  [673] 9.556655 9.840340 5.062857 7.763223 5.113679 8.753463 6.209437 6.795494
#>  [681] 7.382663 9.248074 9.687223 5.097806 5.090656 8.843470 5.695783 9.824904
#>  [689] 8.185279 9.016717 9.525371 9.445467 5.059852 8.839680 6.567011 6.488240
#>  [697] 6.882576 9.377266 8.434402 6.894466 6.319811 5.827293 5.809909 6.894887
#>  [705] 6.366085 8.439391 6.098448 6.197262 8.390103 8.398115 7.461139 7.847565
#>  [713] 9.143436 7.343622 5.641214 8.527042 7.365678 6.468172 7.265501 8.288652
#>  [721] 7.454039 6.707630 6.176463 8.156292 5.530716 7.791457 8.914978 9.822514
#>  [729] 6.510842 9.588805 7.129338 9.873927 8.728177 8.837581 5.981349 9.948733
#>  [737] 7.830699 8.150208 9.456374 9.538919 7.684686 6.437384 7.966209 9.878518
#>  [745] 8.087656 6.034600 8.864149 8.533252 5.489244 7.595149 8.719609 7.656807
#>  [753] 5.686728 6.138575 8.547420 5.323070 8.657187 5.816296 6.228044 6.246009
#>  [761] 7.618589 7.181991 9.458102 5.669735 5.507739 5.357708 9.627432 6.745310
#>  [769] 5.946482 7.581338 5.756433 9.663625 9.503472 5.401788 7.442831 5.143329
#>  [777] 9.333641 9.516562 5.538347 8.559177 7.226622 6.661844 8.717450 7.159044
#>  [785] 8.603789 8.183091 5.906791 5.600117 5.765121 5.168728 6.864129 5.053825
#>  [793] 7.300135 5.849801 6.976425 9.148486 6.085369 7.763067 9.037446 9.770341
#>  [801] 9.510216 5.141339 5.120563 7.458454 6.092339 7.939912 9.382365 5.880885
#>  [809] 7.809640 5.429812 8.378079 5.710860 8.589294 7.109914 7.011001 7.275261
#>  [817] 8.370247 6.079623 9.768761 6.985497 8.396383 6.671535 8.917697 7.113876
#>  [825] 5.512869 6.981910 5.264868 5.963595 9.019783 8.797445 5.940293 9.943352
#>  [833] 5.671668 6.739456 5.597395 7.054479 8.780884 7.831960 9.131546 7.360330
#>  [841] 7.013385 9.128041 7.553758 8.093664 7.314070 6.982703 6.091293 8.864313
#>  [849] 9.045965 8.676464 9.036941 7.866999 8.406631 9.918418 5.698810 7.345175
#>  [857] 5.238711 6.219644 9.982040 7.272307 6.423684 7.869847 5.173778 8.306114
#>  [865] 8.480177 8.382516 8.415910 7.575207 5.628651 5.942985 9.705779 7.225269
#>  [873] 5.982834 5.566758 8.257156 9.243388 9.120094 6.208981 7.036295 5.261933
#>  [881] 9.078778 9.244871 6.095544 7.923823 8.484209 5.995432 8.959599 5.188319
#>  [889] 8.205267 9.428892 7.149491 8.467225 9.173364 9.209256 6.120917 6.697155
#>  [897] 6.160214 8.139917 7.912027 5.986832 9.601688 5.750097 9.868319 6.308991
#>  [905] 9.744822 5.114228 5.095965 6.239125 6.901559 9.166353 8.714932 8.714876
#>  [913] 9.834534 7.281189 6.857250 5.834549 6.737973 9.262360 5.357109 6.860452
#>  [921] 6.644443 6.993591 7.483250 5.880594 8.261585 5.114214 7.671070 5.771598
#>  [929] 6.237798 9.900216 5.984024 9.420578 6.555178 5.004284 8.159035 9.365747
#>  [937] 7.701270 9.036798 6.417199 7.282598 6.292037 6.616026 7.003770 7.952349
#>  [945] 9.413466 6.406682 9.011039 5.472782 6.955856 6.455328 9.328333 8.259047
#>  [953] 5.020673 5.204024 7.525456 9.035581 8.708205 5.220480 9.677306 7.818100
#>  [961] 8.433578 9.714573 8.492508 8.876226 5.231308 8.910438 5.189705 9.161156
#>  [969] 7.366556 6.524077 5.829606 9.441159 8.975137 5.393970 9.597429 5.689355
#>  [977] 8.787100 7.851984 7.582848 6.145107 9.259788 6.037059 5.650568 5.822476
#>  [985] 9.936527 6.757955 5.732218 6.191256 9.188724 7.421278 5.567219 7.743097
#>  [993] 7.761385 5.597780 7.466167 5.698671 6.724034 5.223080 5.301007 5.326851