Generator of simulated values

Generator of simulated values

Usage

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n

An integer which is the number of simulation (same as in rbeta)

min

An integer that corresponds to the lower bound of the interval

max

An integer that corresponds to the upper bound of the interval

param1

An integer that is the first parameter of rbeta function.

param2

An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)
#>    [1] 5.906591 5.072919 9.795406 7.932444 7.534859 6.906614 6.870959 8.116349
#>    [9] 7.109800 8.228411 6.023027 9.318713 9.904831 5.396424 6.379007 7.838022
#>   [17] 9.786759 7.568882 6.158478 6.945618 7.954093 6.499443 5.321358 6.276561
#>   [25] 9.899073 7.108462 9.638392 9.099112 7.138512 9.062461 7.692694 6.038181
#>   [33] 8.513944 7.554102 8.417912 8.955091 8.943910 8.972265 8.715074 5.270903
#>   [41] 6.714713 8.942148 7.311505 9.147804 5.654141 5.144115 5.519725 8.652471
#>   [49] 5.278411 7.610358 9.489716 5.008790 8.709324 5.081796 5.105113 8.061686
#>   [57] 6.892687 8.104342 7.617498 7.345522 7.612860 5.082394 5.252813 8.376773
#>   [65] 8.215364 7.736505 8.700717 9.959748 6.796214 8.149784 9.120035 9.162044
#>   [73] 5.094190 7.719581 9.432776 7.058166 6.305577 9.947517 7.483495 9.766208
#>   [81] 5.767768 6.750531 8.348709 8.227869 8.057508 9.783646 5.841979 5.544732
#>   [89] 9.682450 7.436649 6.902881 6.712172 6.945288 6.441635 5.973920 6.741992
#>   [97] 9.915868 6.054047 8.448954 5.997512 5.578012 6.748063 6.613344 9.951952
#>  [105] 6.816282 9.167227 7.155918 8.091505 6.353436 9.846404 7.336122 5.589504
#>  [113] 6.242828 8.874457 7.529183 9.955707 5.397745 8.997037 7.943312 8.185433
#>  [121] 6.704990 5.359852 7.050877 8.213088 6.509413 6.333867 7.979259 9.813534
#>  [129] 8.725432 6.986062 7.267540 6.752144 6.033143 7.596424 8.226731 9.318112
#>  [137] 5.253595 6.293200 7.739922 9.752348 7.769575 9.584724 6.633994 9.555762
#>  [145] 8.201503 8.362855 9.367990 9.394536 8.927539 5.767369 7.341842 9.141804
#>  [153] 7.355040 6.451828 9.718925 9.878477 8.677601 8.368806 8.615267 6.417377
#>  [161] 9.194124 9.049653 6.081211 5.631466 5.262070 6.410378 5.863162 6.206615
#>  [169] 8.303709 9.983991 9.175092 9.431207 8.355904 8.564775 5.153775 7.044489
#>  [177] 6.628234 6.532274 5.843979 7.796311 9.719586 7.432104 8.804482 5.500714
#>  [185] 5.289174 7.484562 9.888483 7.064640 9.002732 7.972799 5.189094 8.931589
#>  [193] 8.291727 9.153182 8.204514 5.664530 9.173628 9.738809 5.680025 7.719397
#>  [201] 5.338512 7.739336 5.454515 5.385318 6.337332 6.356178 9.990430 8.613050
#>  [209] 6.379614 7.913715 7.290560 5.930335 7.158290 8.398368 5.177059 9.535018
#>  [217] 9.025497 6.501879 6.481198 9.536188 9.083070 9.589096 6.695829 7.763790
#>  [225] 8.585222 8.902512 7.631154 8.619886 9.742693 8.383390 9.590888 6.436476
#>  [233] 8.860475 5.781441 5.587221 9.850751 8.196205 6.649470 8.515979 9.502240
#>  [241] 6.690740 7.564265 5.676031 8.423766 6.902327 5.857960 7.652214 7.183777
#>  [249] 6.178087 7.402840 6.570055 5.115554 6.590212 6.482481 8.433241 7.501849
#>  [257] 7.285955 6.398414 8.488657 5.097121 5.199671 6.860022 5.391381 5.690878
#>  [265] 5.903440 8.498460 6.637223 5.428440 6.176392 8.266590 9.621907 6.332811
#>  [273] 6.318064 6.796928 9.968431 5.190232 6.085450 8.478658 9.642452 9.927081
#>  [281] 5.027874 6.071151 5.568408 5.189296 9.660783 5.170692 5.734472 6.311411
#>  [289] 9.956440 9.111865 9.513526 7.605146 7.090257 7.863638 6.907435 7.306600
#>  [297] 9.911739 6.549182 8.022389 8.379187 8.656268 7.553755 9.107039 5.762673
#>  [305] 5.896580 8.744905 6.610990 6.300573 8.277302 7.653913 5.795929 7.643231
#>  [313] 9.794253 8.901764 9.672568 8.249618 5.524067 7.894181 9.903989 6.653300
#>  [321] 7.558615 6.806240 7.101864 6.708226 6.606079 6.076533 9.658287 6.972485
#>  [329] 9.822547 6.262076 5.769086 9.025904 8.158977 9.730865 5.323322 5.554290
#>  [337] 6.017471 8.678179 6.099226 9.039779 6.183671 9.653567 8.631509 5.356544
#>  [345] 7.285532 7.993205 7.985643 9.210890 6.775009 7.950492 6.619601 5.813043
#>  [353] 8.959147 5.511291 5.146669 8.711303 8.880918 7.132524 5.214942 6.348653
#>  [361] 5.716838 5.503632 7.815847 9.731147 7.852383 5.112980 7.830379 8.012247
#>  [369] 7.937504 9.942962 9.369600 8.756188 9.042062 8.742441 6.813790 8.564063
#>  [377] 6.664901 8.231764 6.575519 5.673690 9.391860 7.337484 6.814711 5.119042
#>  [385] 8.202449 6.790996 6.707461 5.615590 9.382103 6.017361 5.338705 6.071392
#>  [393] 6.772972 6.354315 9.121814 6.361463 7.267430 8.294233 7.405112 6.049806
#>  [401] 7.861752 8.716033 7.844738 5.537512 8.801601 9.951002 8.327519 7.437194
#>  [409] 8.616341 9.110819 6.401964 7.170075 9.687906 7.381143 5.119992 9.518795
#>  [417] 5.555171 8.152771 7.788861 5.330064 6.237223 6.375501 7.951167 9.370538
#>  [425] 6.139279 9.117030 9.947050 7.172957 5.803297 7.399953 6.931691 8.244042
#>  [433] 9.825608 8.520761 9.066600 5.572616 7.905368 7.054636 5.368164 6.486299
#>  [441] 8.833960 9.058610 6.282575 7.751883 6.865885 9.644188 7.767104 7.011162
#>  [449] 7.302462 5.900835 6.049994 7.120430 8.665492 8.505360 7.636070 7.303489
#>  [457] 9.846108 6.713847 9.803978 8.373094 9.568069 8.035643 9.449472 9.597978
#>  [465] 9.843146 9.550507 7.270697 9.789114 9.596203 8.249116 6.600693 5.432247
#>  [473] 7.837719 6.145575 8.519295 5.843680 7.946230 7.465471 5.029311 9.991239
#>  [481] 5.912178 8.178273 6.082070 5.921845 9.965204 5.726542 6.486658 6.019149
#>  [489] 5.205871 9.706656 8.997176 8.051569 5.993919 5.052165 8.932836 9.016416
#>  [497] 6.403083 8.991428 7.157070 9.702643 7.419638 5.577297 9.878204 9.215898
#>  [505] 8.094490 9.723778 5.738416 9.225293 5.762139 9.029307 9.825334 5.657891
#>  [513] 5.840234 5.399897 8.219477 6.853489 6.880959 8.074951 7.363476 7.780666
#>  [521] 7.289650 6.012321 8.203718 6.405753 9.620379 5.332295 8.212402 9.153655
#>  [529] 5.047071 6.696769 8.337168 8.026622 6.020531 9.881809 9.382061 8.733146
#>  [537] 5.316324 9.442255 8.417763 6.022484 7.921824 9.323667 6.841568 8.425020
#>  [545] 9.950494 7.789309 5.590039 6.269049 7.000095 5.156971 9.264020 9.304509
#>  [553] 8.684799 8.640209 7.405663 6.360727 7.592137 9.330373 6.834114 8.346495
#>  [561] 5.240143 9.466652 7.206409 7.438095 7.629590 6.205570 7.947231 9.174428
#>  [569] 9.599694 8.704956 9.994885 7.447810 9.388663 5.847202 7.312743 8.086395
#>  [577] 5.761689 9.234329 6.995326 5.748431 8.224765 7.175260 6.425278 5.025032
#>  [585] 6.807877 9.911244 5.795280 8.473965 9.313647 5.232277 8.318621 8.326464
#>  [593] 8.829451 5.590693 8.713029 5.558885 6.864952 8.738155 8.509598 6.800998
#>  [601] 5.071865 6.165755 9.404612 6.075549 5.458851 6.862124 7.238748 8.499372
#>  [609] 6.095745 8.603655 8.729577 6.342101 6.111548 8.328045 8.800377 9.216450
#>  [617] 6.483477 7.110396 7.483241 6.333033 9.728413 5.272240 5.459322 7.669645
#>  [625] 7.933208 7.388074 8.100836 8.682197 7.994070 8.878685 8.809506 7.649392
#>  [633] 9.547491 9.972252 8.799324 7.281684 9.913877 9.880632 5.556873 7.268658
#>  [641] 7.834695 9.366876 7.790785 8.697975 5.305652 6.885238 7.747305 8.853588
#>  [649] 8.806891 5.534752 9.749421 9.865489 9.653993 5.405944 6.983828 8.331152
#>  [657] 8.976877 9.345364 5.139486 9.200051 8.258322 5.729421 8.263806 5.288307
#>  [665] 7.192353 5.712390 7.277096 6.822556 8.110170 6.583967 7.524203 9.657850
#>  [673] 6.287232 9.571448 7.234925 7.983145 8.134477 7.068724 6.407283 7.759387
#>  [681] 5.832226 7.246124 9.048789 7.578039 7.748414 5.342547 7.874233 9.631796
#>  [689] 6.086786 7.904696 5.941227 9.717902 7.481843 8.711716 5.696390 6.986882
#>  [697] 8.095377 5.806637 5.226925 7.761531 9.067119 6.219093 9.499541 7.880072
#>  [705] 9.155201 8.090280 7.771687 7.621637 6.182564 9.741408 8.534804 5.936682
#>  [713] 6.005835 5.308877 6.593110 7.486501 8.162047 8.452007 8.972608 9.754217
#>  [721] 9.100065 6.936354 7.715142 8.392582 9.456966 8.399653 5.002841 8.432208
#>  [729] 9.187267 6.070887 8.775274 6.261551 5.314818 8.285127 9.796989 7.741279
#>  [737] 9.633163 5.712676 5.114951 9.758397 6.031481 8.678311 5.765544 9.389095
#>  [745] 8.605963 5.133855 9.417716 9.463992 6.100732 9.726358 9.496048 7.987689
#>  [753] 7.970356 5.302926 9.011540 9.409377 6.095591 9.741921 7.289245 9.854464
#>  [761] 7.099745 6.554238 5.943979 9.237599 9.848696 6.114144 8.680126 5.884251
#>  [769] 6.082671 7.060449 8.186784 5.285975 7.431457 6.200072 7.888867 5.752306
#>  [777] 8.939736 5.948306 7.710279 7.051708 8.524683 9.236338 7.853883 6.464290
#>  [785] 6.044593 8.981945 9.140794 5.427656 6.248902 7.802765 9.322994 8.443181
#>  [793] 7.499486 8.810457 6.339560 6.986605 6.777191 7.783216 8.600758 7.499064
#>  [801] 8.741615 9.657806 6.740712 7.850525 9.774408 8.150786 6.142202 8.109995
#>  [809] 7.264217 6.994938 6.665519 6.078662 8.448800 8.051983 9.545561 9.267379
#>  [817] 5.127399 5.443463 5.934020 9.221157 8.538681 8.498536 6.325026 5.397438
#>  [825] 6.816348 6.215722 7.648471 6.934946 6.379591 5.937313 8.931245 7.002031
#>  [833] 8.527326 8.546396 6.493854 5.807539 6.664490 5.535524 9.835219 5.340165
#>  [841] 5.674808 7.202842 5.652482 8.047131 5.106684 8.843415 6.436623 9.322254
#>  [849] 7.328273 5.215150 7.126709 8.589689 7.624176 9.249002 7.535185 9.325243
#>  [857] 5.820681 5.242163 6.602545 8.952126 5.894578 8.509373 6.295885 6.051579
#>  [865] 5.361820 8.549609 8.821996 9.287074 8.126043 7.334739 7.670305 9.363039
#>  [873] 7.665105 8.196797 7.836400 8.214627 9.931051 7.285350 5.338529 9.346804
#>  [881] 5.486460 6.361802 6.875833 6.256652 8.890633 7.701144 8.813541 5.239324
#>  [889] 8.132285 9.964672 5.233440 9.663489 8.817216 5.775347 5.188558 6.128240
#>  [897] 6.353440 5.880085 9.521444 6.513363 5.572272 7.874545 9.175041 7.205139
#>  [905] 6.887198 5.511662 7.287699 9.906164 6.465669 5.089641 9.166113 7.792656
#>  [913] 9.835855 6.565953 8.482405 8.744231 8.140458 6.507547 7.833044 8.550954
#>  [921] 9.313666 6.463156 8.153782 8.540422 8.506800 8.340663 5.229876 5.330960
#>  [929] 9.641319 9.704339 8.183840 5.449086 9.781784 9.886282 9.345206 5.566785
#>  [937] 5.726912 5.186148 8.785020 6.912813 5.179084 5.300307 8.390774 5.055321
#>  [945] 5.405986 8.372079 8.571610 8.696457 9.562695 5.692174 9.237131 9.273200
#>  [953] 9.709634 5.440665 9.595647 5.117822 6.336816 7.073947 7.355137 5.399900
#>  [961] 7.985291 6.979595 8.110532 7.329671 5.840149 6.040987 7.917399 7.082101
#>  [969] 8.667927 5.582409 8.215349 9.141725 7.130183 7.601841 7.513247 8.030171
#>  [977] 6.321958 5.885224 9.377880 9.162146 6.676443 6.587106 5.210822 7.524919
#>  [985] 6.366052 8.039546 6.131283 6.264904 8.035527 7.184971 9.735140 7.259015
#>  [993] 7.523424 6.699554 9.075841 9.314176 7.214558 6.315667 7.068485 8.020811