Generator of simulated values

Generator of simulated values

translatedRandomBeta(n, min, max, param1 = 3, param2 = 1)

Arguments

n
An integer which is the number of simulation (same as in rbeta)

min
An integer that corresponds to the lower bound of the interval

max
An integer that corresponds to the upper bound of the interval

param1
An integer that is the first parameter of rbeta function.

param2
An integer that is second parameter of rbeta function.

Value

A vector of n simulated values

Author

Thomas Burger

Examples

translatedRandomBeta(1000, 5, 10, 1, 1)

#>    [1] 9.474182 6.298200 7.950159 7.897270 6.735575 9.520500 7.553268 8.497736
#>    [9] 9.945684 9.194153 7.558131 7.732007 9.107178 9.518303 9.490795 9.954433
#>   [17] 7.648163 9.021069 8.758738 5.083675 6.611264 7.922857 8.261554 6.022382
#>   [25] 8.840796 9.634366 8.566836 8.068589 8.823822 8.235465 9.863774 9.347683
#>   [33] 8.664209 6.480800 6.428856 6.035253 6.201738 6.753009 5.635388 9.625272
#>   [41] 8.406025 8.740183 8.748737 8.798390 7.123232 5.579354 5.143421 6.773376
#>   [49] 6.037568 8.280197 8.307410 8.837902 8.829500 6.807342 9.671797 8.155645
#>   [57] 8.521201 8.024524 6.844332 5.730918 7.330897 5.819409 6.882159 6.724848
#>   [65] 9.011178 8.507931 9.496597 7.839070 9.229173 6.013589 5.891994 7.019067
#>   [73] 7.683855 8.475493 9.330110 9.888529 5.285260 6.528921 5.821720 8.188855
#>   [81] 7.413910 9.033830 8.613707 5.414298 9.129163 5.579361 8.001616 7.961228
#>   [89] 9.777969 5.209421 5.180829 7.239662 7.284989 8.685370 7.339811 8.023234
#>   [97] 5.996780 8.054797 6.382774 6.460268 8.468582 7.119541 9.600547 7.052962
#>  [105] 5.327559 5.902712 8.766156 5.049584 8.260520 7.187299 7.109660 8.243218
#>  [113] 9.739116 5.798933 9.190693 7.135609 6.684089 8.763735 9.517728 5.044983
#>  [121] 5.863208 7.888224 5.925302 7.077524 6.142346 7.338233 7.286016 6.617215
#>  [129] 6.562566 6.227210 9.630001 6.533915 6.197940 8.814347 8.494728 8.097007
#>  [137] 7.367357 9.258258 5.113903 9.663889 9.799110 6.637114 5.493596 7.098248
#>  [145] 6.023281 7.471061 8.396094 8.284881 8.433069 6.602585 5.287788 8.792170
#>  [153] 6.959219 5.603367 7.003495 7.849515 8.856762 9.585904 6.311511 6.576152
#>  [161] 7.866131 6.147021 7.037524 5.818296 9.291576 5.336350 9.348964 9.482859
#>  [169] 8.145275 7.321580 6.595767 7.240980 8.000389 5.277277 5.305391 7.216714
#>  [177] 6.665785 5.391604 5.862463 6.040937 8.402538 7.041304 7.405906 5.689014
#>  [185] 9.794817 6.880925 8.722179 6.225625 5.899887 8.445329 9.341080 8.015927
#>  [193] 5.424940 9.968655 6.233968 7.770697 9.164227 8.101741 6.119802 9.141694
#>  [201] 5.545463 6.507951 9.220083 6.830328 9.412372 8.057722 6.443216 6.461084
#>  [209] 5.306705 5.166358 6.081287 7.969219 9.087596 9.465080 5.978485 6.872293
#>  [217] 6.524492 5.031845 7.090725 8.383213 6.758840 9.686290 6.583605 6.319358
#>  [225] 6.687879 9.917577 7.344691 6.615623 8.495691 7.169228 9.019421 9.760377
#>  [233] 8.124542 8.658897 6.919795 5.125711 9.284866 6.699847 6.762106 7.368585
#>  [241] 9.919448 7.262520 8.776205 6.221876 8.753935 5.974911 8.529574 7.831206
#>  [249] 9.678100 7.017558 7.821426 6.708672 6.712120 8.534845 6.429665 6.531681
#>  [257] 9.786483 7.204688 8.079805 8.951269 5.408799 6.426163 9.170173 9.036786
#>  [265] 5.213957 6.760130 5.292869 8.615670 6.692911 9.571278 7.675431 5.619278
#>  [273] 5.024563 6.893699 7.161515 6.111388 8.994309 8.146849 7.003487 8.005388
#>  [281] 7.292358 5.696577 5.054011 7.275945 5.743616 6.981138 9.020032 5.096822
#>  [289] 5.462888 7.677029 6.512623 7.394747 6.723387 9.787666 9.652813 5.155464
#>  [297] 5.915534 9.449559 5.957930 5.998196 5.496740 7.102054 8.028693 6.331710
#>  [305] 9.878101 5.970284 9.129152 9.855281 7.406274 8.650932 5.269767 7.383072
#>  [313] 7.250563 8.936003 9.667591 8.635535 8.299976 9.855059 6.735244 9.544148
#>  [321] 5.332669 5.364617 8.960590 7.833365 5.320620 6.656182 7.420237 5.307304
#>  [329] 8.874019 6.938669 7.328821 6.117445 5.851273 8.780154 7.862707 6.978689
#>  [337] 6.531193 6.235269 9.795512 6.171026 6.955000 7.851818 7.406627 9.203511
#>  [345] 9.213004 5.340933 8.072157 6.263056 9.454719 8.736950 5.111313 8.344069
#>  [353] 6.358415 5.973145 8.382045 5.027978 6.778970 6.657992 9.787804 7.514371
#>  [361] 5.855445 9.514183 7.793317 7.764288 7.601817 7.070069 5.680351 9.955472
#>  [369] 7.548716 6.824240 8.453847 7.085893 8.667365 6.665576 6.326291 6.579416
#>  [377] 7.234085 5.462844 9.895870 9.951655 8.862252 6.746411 5.768405 5.793947
#>  [385] 5.456874 5.170689 6.761050 9.515045 9.138567 5.707022 6.343507 7.478108
#>  [393] 7.487303 7.394419 6.962219 6.953955 8.889601 5.527573 9.422325 9.116027
#>  [401] 7.429004 9.409790 6.726564 6.049328 7.240440 9.817693 6.013571 5.145342
#>  [409] 8.553449 9.088297 5.265028 9.168277 7.672020 7.566678 5.285070 7.366249
#>  [417] 8.337940 5.443838 8.112232 6.595913 6.325142 7.738724 7.598260 6.412121
#>  [425] 6.346083 8.732287 6.179582 6.068190 6.718285 7.663966 6.403328 5.170436
#>  [433] 9.344129 8.268660 7.375138 7.418542 7.057054 5.908004 8.882681 9.143220
#>  [441] 9.765026 5.952681 5.458221 7.837730 9.325972 6.950620 7.841611 7.809475
#>  [449] 7.400882 9.787299 5.651281 6.163115 7.501879 9.604003 7.398825 7.572569
#>  [457] 5.784660 7.852574 8.390142 7.164796 5.871844 9.283275 7.223552 7.722618
#>  [465] 7.387297 7.141409 8.486774 7.624744 7.395694 9.784829 6.033169 7.674403
#>  [473] 5.773764 5.948289 6.361575 6.515678 9.358538 5.159229 5.270883 5.735502
#>  [481] 7.687698 9.182161 8.436138 7.663151 7.922144 9.196512 7.965343 9.266748
#>  [489] 5.504952 9.832377 5.139434 8.704291 8.636413 9.551968 7.598414 5.286442
#>  [497] 8.028392 5.477798 7.764557 7.940004 9.865535 8.668001 8.160451 8.620945
#>  [505] 5.375485 7.502326 8.720770 8.328948 6.413194 5.157575 8.940929 7.343365
#>  [513] 7.294816 8.851433 5.580733 5.661389 7.655607 6.351000 7.925747 9.798591
#>  [521] 7.802490 8.696995 7.748813 5.325978 5.602672 8.285394 8.032841 5.159029
#>  [529] 6.320082 9.748220 5.658483 9.209729 8.067190 8.378657 6.946682 6.942160
#>  [537] 7.589032 9.463424 7.406074 8.645212 9.869716 9.594551 9.573210 7.681643
#>  [545] 9.173272 6.220585 8.323925 9.914120 8.999815 8.064077 9.445907 5.626507
#>  [553] 9.397387 7.567957 9.761786 7.383905 6.894023 7.326877 6.811512 8.576592
#>  [561] 5.323960 5.244637 5.294419 6.895229 7.647887 7.782356 7.008133 8.723058
#>  [569] 7.072473 5.590013 8.920262 7.882686 6.698211 9.457273 6.418988 9.236077
#>  [577] 7.925727 5.879166 8.686346 7.550996 8.303205 6.835351 6.916637 7.091819
#>  [585] 9.317486 5.544650 6.423497 6.756377 5.220647 6.561483 8.260337 5.496380
#>  [593] 6.288540 7.996756 5.391564 7.139572 7.236962 5.843471 8.642600 9.728693
#>  [601] 7.508580 7.020450 6.634774 6.398471 8.053550 9.409782 6.489630 5.530881
#>  [609] 5.964710 7.823886 9.674774 8.578177 5.464629 8.035733 8.764587 5.015224
#>  [617] 9.809204 5.620976 9.317852 5.512345 8.031238 9.522204 5.154558 6.244997
#>  [625] 6.255878 5.608281 8.262383 7.262073 5.152180 5.769310 5.324431 5.793943
#>  [633] 5.867158 5.547341 6.936976 6.245264 6.228109 6.727524 8.382602 8.070209
#>  [641] 8.788070 9.777725 7.706493 7.226702 7.396708 6.958720 6.890900 7.610720
#>  [649] 9.444130 9.212039 5.137861 8.505236 8.717884 7.493493 6.338037 5.729234
#>  [657] 8.845602 6.127343 6.483827 6.225880 9.715253 8.941668 6.281940 7.379231
#>  [665] 6.391777 5.413323 6.519511 8.392103 9.608043 8.753361 5.717553 7.032485
#>  [673] 5.476082 7.119480 7.351789 8.087409 6.389988 6.485125 9.813863 6.297701
#>  [681] 5.370773 7.983660 8.224910 5.094126 9.209247 5.202192 5.690249 7.384682
#>  [689] 5.565026 6.788010 6.711880 6.474954 7.521961 6.520175 9.555376 6.530591
#>  [697] 7.555539 9.250799 7.521756 5.647562 9.585725 9.615418 9.582341 8.933600
#>  [705] 6.929281 6.158721 5.569458 8.958720 8.931862 6.740252 6.102215 7.638587
#>  [713] 5.521289 5.556416 7.278509 9.006206 5.381073 7.497266 5.563466 8.288760
#>  [721] 7.406177 5.016454 8.959539 6.683769 5.299339 5.608411 5.674993 8.526875
#>  [729] 5.899366 9.109896 9.560217 5.019217 6.812719 8.832958 8.489929 7.925772
#>  [737] 7.161924 6.771917 7.550434 6.161600 7.496629 8.215839 6.522330 5.918486
#>  [745] 9.864188 7.564118 6.744092 8.974407 9.103986 5.097922 9.923812 7.323495
#>  [753] 8.620317 8.288682 7.317568 9.681007 6.692415 5.874090 7.882247 6.882159
#>  [761] 5.832717 7.838534 6.417165 8.732432 8.650847 6.342769 5.316434 5.192151
#>  [769] 5.560539 9.408021 7.718415 9.426415 9.607061 9.817656 7.628982 9.710707
#>  [777] 6.783974 5.003330 5.654369 6.779874 9.253008 7.398449 7.736274 5.234035
#>  [785] 7.389195 9.792846 6.411892 5.870373 9.979641 6.361275 5.398324 9.888585
#>  [793] 5.940773 6.018457 5.035076 5.825484 8.637724 9.355899 7.579801 9.059975
#>  [801] 5.081759 8.195048 9.713979 5.187561 7.730562 5.381612 5.617714 8.722676
#>  [809] 7.927599 6.222612 9.409395 8.546653 8.088888 8.494860 8.260479 6.778935
#>  [817] 5.708467 8.553440 6.439724 6.559700 9.276330 5.032226 5.344270 7.897709
#>  [825] 6.234796 9.114814 8.570409 6.438279 7.121067 6.650876 5.400735 6.271284
#>  [833] 8.823052 7.265069 6.272106 6.879823 5.143435 9.252034 5.149734 6.555415
#>  [841] 7.154990 6.703118 9.512597 6.837406 7.166074 5.702537 5.569726 5.421297
#>  [849] 5.316011 7.449967 8.425902 5.098782 5.601143 6.499519 5.218070 7.582469
#>  [857] 7.706387 8.473086 7.704925 6.806014 8.230041 9.920506 8.563694 6.630018
#>  [865] 8.898965 8.188346 6.670139 9.628211 9.226471 9.007668 6.940496 8.117115
#>  [873] 5.115383 5.223666 6.571051 9.462826 6.572316 9.814100 7.010176 7.263763
#>  [881] 5.542186 5.389517 6.192998 8.243999 9.181633 9.840377 7.709590 5.756384
#>  [889] 5.559190 6.726677 7.804417 9.393849 5.149790 6.878412 7.942206 7.760265
#>  [897] 7.721266 8.745883 8.373752 6.885831 5.207769 6.535814 9.976745 9.194459
#>  [905] 8.788247 9.807612 6.438354 7.328129 7.905031 8.695431 7.380278 7.805448
#>  [913] 8.430837 7.162368 9.316355 5.534128 5.598191 6.539650 9.572537 8.459449
#>  [921] 8.481458 7.994842 6.244629 5.127381 7.315929 8.127555 6.335423 6.630542
#>  [929] 7.006380 5.479362 5.048078 8.579865 9.714458 6.852524 5.016758 9.088376
#>  [937] 8.284083 8.123222 9.808609 8.167665 5.715421 6.755968 6.788920 9.080951
#>  [945] 5.804121 7.749287 9.232350 5.230414 7.138302 5.788104 7.728959 6.538276
#>  [953] 8.542242 5.114819 6.899267 8.962524 7.943402 6.514852 6.371505 5.502563
#>  [961] 7.166173 7.182336 9.077357 9.548032 8.566003 7.449494 6.690784 5.386985
#>  [969] 9.641428 6.683488 5.061813 8.243976 5.838785 8.457437 8.833380 9.540304
#>  [977] 9.611584 9.197065 7.564307 6.416315 8.050929 5.242964 9.686460 6.639322
#>  [985] 5.529828 8.208610 9.505313 8.624338 6.966980 6.502136 9.089955 7.315416
#>  [993] 9.950623 8.645042 5.747566 5.742900 8.610566 5.637770 7.342915 9.791770